% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-snap-op-methods.R, R/snap-graph.R,
%   R/snap-operator-help.R, R/snap-operator.R
\name{as_xml}
\alias{as_xml}
\title{create xml_document object from snap_operator and snap_operator_help objects}
\usage{
as_xml(x, ...)
}
\arguments{
\item{x}{snap_read_operator object}

\item{...}{not used}
}
\description{
create xml_document object from snap_operator and snap_operator_help objects

create xml_document object from snap_graph

create xml_document object from snap_operator_help

create xml_document object from snap_operator

create xml_document object from snap_read_operator
}
