% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-op-methods.R, R/snap_operator.R
\name{as_xml_document}
\alias{as_xml_document}
\title{create xml_document object from snap_operator and snap_operator_help objects}
\usage{
as_xml_document(x, ...)
}
\arguments{
\item{x}{snap_read_operator object}
}
\description{
create xml_document object from snap_operator and snap_operator_help objects

create xml_document object from snap_operator

create xml_document object from snap_read_operator
}
