% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_builders.R
\name{build_xml_engine}
\alias{build_xml_engine}
\title{A function to write R functions for given xml graphs.
This function creates a function that takes the arguments and xml from
a snap_operator_help_bject to create a function that can then be used to
build gpt graphs.}
\usage{
build_xml_engine(operator, node = TRUE, null_src = FALSE)
}
\description{
A function to write R functions for given xml graphs.
This function creates a function that takes the arguments and xml from
a snap_operator_help_bject to create a function that can then be used to
build gpt graphs.
}
\keyword{internal}
