% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ANPPest.R
\name{calc_ANPPest}
\alias{calc_ANPPest}
\title{Actual Aboveground Net Primary Production}
\usage{
calc_ANPPest(ANPPmax, LAI)
}
\arguments{
\item{ANPPmax}{Maximum potential aboveground net primary production (g/m2).}

\item{LAI}{Leaf area index per unit ground area.}
}
\description{
Calculation of aboveground net primary production under the specified grazing management system (g/m2), from Ritchie 2014.
}
