% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ANPPmax.R
\name{calc_ANPPmax}
\alias{calc_ANPPmax}
\title{Maximum Potential Above-ground Net Primary Production}
\usage{
calc_ANPPmax(RAIN, SAND)
}
\arguments{
\item{RAIN}{Mean annual precipitation (mm/year).}

\item{SAND}{Sand fraction in the top 40 cm of soil (0 - 1).}
}
\description{
Calculation of the parameter ANPPmax, the maximum potential grass production for a given site, from Ritchie 2014.
}
