% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ANPPt_est.R
\name{calc_ANPPt_est}
\alias{calc_ANPPt_est}
\title{Estimated Above-ground Net Primary Productivity in year t (ANPPt est)}
\usage{
calc_ANPPt_est(GI, ANPPt_max)
}
\arguments{
\item{GI}{Grazing intensity for year t (proportion)}

\item{ANPPt_max}{Output of calc_ANPPt_max()}
}
\description{
This function takes GI and ANPPt_max to calculate ANPPt_est. Not generally called directly but incorporated in wrapper function.
}
\seealso{
\link{SNAP} which wraps this function.
}
