% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ANPPt_max.R
\name{calc_ANPPt_max}
\alias{calc_ANPPt_max}
\title{Above-ground Net Primary Productivity in year t (ANPPt max)}
\usage{
calc_ANPPt_max(RAIN, SAND)
}
\arguments{
\item{RAIN}{MAP for year t (mm/year)}

\item{SAND}{Sand \% in top 30 cm soil}
}
\description{
This function takes RAIN and SAND to calculate ANPPt_max. Not generally called directly but incorporated in wrapper function.
}
\seealso{
\link{SNAP} which wraps this function.
}
