% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_BNPPest.R
\name{calc_BNPPest}
\alias{calc_BNPPest}
\title{Actual Belowground Net Primary Production}
\usage{
calc_BNPPest(RAIN)
}
\arguments{
\item{RAIN}{Mean annual precipitation (mm/year).}
}
\description{
Calculation of belowground net primary production for a given site (g/m2), from Ritchie 2014.
}
