% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_BNPPt_est.R
\name{calc_BNPPt_est}
\alias{calc_BNPPt_est}
\title{Below-ground Net Primary Productivity in year t (BNPPt est)}
\usage{
calc_BNPPt_est(RAIN)
}
\arguments{
\item{RAIN}{MAP for year t (mm/year)}
}
\description{
This function takes RAIN to calculate BNPPt_est. Not generally called directly but incorporated in wrapper function.
}
\seealso{
\link{SNAP} which wraps this function.
}
