% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_DDSOC.R
\name{calc_DDSOC}
\alias{calc_DDSOC}
\title{Dung-derived Soil Organic Carbon Inputs}
\usage{
calc_DDSOC(LIGCELL, ANPPest, GI)
}
\arguments{
\item{LIGCELL}{Lignin and cellulose fraction of plant biomass (0 - 1).}

\item{ANPPest}{Actual aboveground net primary production (g/m2).}

\item{GI}{Grazing intensity (proportion of aboveground biomass removed per year, 0 - 1).}
}
\description{
Calculation of dung-derived soil organic carbon inputs under the specified grazing management system (g/m2), from Ritchie 2014.
}
