% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_LAI.R
\name{calc_LAI}
\alias{calc_LAI}
\title{Leaf Area Index}
\usage{
calc_LAI(GI, LEAFFRAC)
}
\arguments{
\item{GI}{Grazing intensity (proportion of aboveground biomass removed per year, 0 - 1).}

\item{LEAFFRAC}{Leaf fraction (the proportion of aboveground biomass that is leaf tissue, 0 - 1)}
}
\description{
Calculation of the parameter LAI, the leaf area index per unit ground area, from Ritchie 2014.
}
