% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_LEAFFRAC.R
\name{calc_LEAFFRAC}
\alias{calc_LEAFFRAC}
\title{Proportion of Aboveground Biomass as Leaf}
\usage{
calc_LEAFFRAC(GI)
}
\arguments{
\item{GI}{Grazing intensity (proportion of aboveground biomass removed per year, 0 - 1).}
}
\description{
Calculation of the parameter LEAFFRAC, the fraction of aboveground plant biomass that is leaf tissue, from Ritchie 2014.
}
