% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_PDSOC.R
\name{calc_PDSOC}
\alias{calc_PDSOC}
\title{Plant-derived Soil Organic Carbon Inputs}
\usage{
calc_PDSOC(LIGCELL, ANPPest, GI, FIRE, BNPPest)
}
\arguments{
\item{LIGCELL}{Lignin and cellulose fraction of aboveground plant biomass (0 - 1).}

\item{ANPPest}{Actual aboveground net primary production (g/m2).}

\item{GI}{Grazing intensity (proportion of aboveground biomass removed per year, 0 - 1).}

\item{FIRE}{Proportion of aboveground biomass removed by fire per year (0 - 1).}

\item{BNPPest}{Belowground net primary production (g/m2).}
}
\description{
Calculation of plant-derived soil organic carbon inputs under the specified grazing management system (g/m2), from Ritchie 2014.
}
