% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_SOCeq.R
\name{calc_SOCeq}
\alias{calc_SOCeq}
\title{Equilibrium SOC Stock}
\usage{
calc_SOCeq(PDSOC, DDSOC, wetdays, SAND)
}
\arguments{
\item{PDSOC}{Plant-derived soil organic carbon inputs (g/m2).}

\item{DDSOC}{Dung-derived soil organic carbon inputs (g/m2).}

\item{wetdays}{Number of days wet enough to support soil microbial respiration (days).}

\item{SAND}{Sand fraction in the top 40 cm of soil (0 - 1).}
}
\description{
Calculation of the SOC stock (tSOC/ha) at equilibrium under the specified grazing management system, from Ritchie 2014. This is the point at which SOC inputs from plants and dung are offset by SOC losses due to microbial respiration (dsoc = 0).
}
