% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_wetdays.R
\name{calc_wetdays}
\alias{calc_wetdays}
\title{Wet Days}
\usage{
calc_wetdays(RAIN, Gdays)
}
\arguments{
\item{RAIN}{Mean annual precipitation (mm/year).}

\item{Gdays}{Number of days in growing season (days). Default is 240.}
}
\description{
Calculation of number of wet days per year able to support soil microbial respiration, from Ritchie 2014.
}
