% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esa-snappy-py-tools.R
\name{configure_snappy_python}
\alias{configure_snappy_python}
\title{Configure the snappy python environment
congfigures the python library "esa_snappy"}
\usage{
configure_snappy_python()
}
\description{
Configure the snappy python environment
congfigures the python library "esa_snappy"
}
\details{
This function creates a virtual python environment with the
required packages for the "esa_snappy" library. The function also sets the
"snappy_python" option to the python executable in the virtual environment.
}
\keyword{internal}
