% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esa-snappy-py-tools.R
\name{destroy_snappy_env}
\alias{destroy_snappy_env}
\title{Destroy the snappy python environment}
\usage{
destroy_snappy_env(ask = interactive())
}
\arguments{
\item{ask}{logical if TRUE, will ask for confirmation}
}
\description{
Destroy the snappy python environment
}
\keyword{internal}
