% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt-cli-tools.R
\name{get_operators}
\alias{get_operators}
\title{function to get the operators from the gpt executable}
\usage{
get_operators(gpt_path = "/home/hugh/esa-snap/bin/gpt")
}
\arguments{
\item{gpt_path}{character path to the gpt executable}
}
\value{
tibble of operators and descriptions
}
\description{
function to get the operators from the gpt executable
}
