% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esa-snappy-py-tools.R
\name{get_param_defaults}
\alias{get_param_defaults}
\title{Get parameter defaults for a SNAP operator
function calls the esa_snappy python library to get the default parameters
for a SNAP operator}
\usage{
get_param_defaults(operator, py_env = find_snappy_install())
}
\arguments{
\item{operator}{character name of the SNAP operator}

\item{py_env}{character path to the python environment}
}
\value{
named list of parameter defaults
}
\description{
Get parameter defaults for a SNAP operator
function calls the esa_snappy python library to get the default parameters
for a SNAP operator
}
\keyword{internal}
