% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esa-snap-setup.R
\name{install_snap}
\alias{install_snap}
\title{install snap
download and install ESA SNAP.}
\usage{
install_snap(
  installer = download_installer(url = get_os_installer(), dest =
    snapr_dir("snap-installers"), chmod = TRUE),
  install_dir = snapr_dir("esa-snap")
)
}
\arguments{
\item{installer}{character path to the snap installer}

\item{install_dir}{character path to the installation directory}
}
\description{
install snap
download and install ESA SNAP.
}
\details{
This function will download the installation file and install SNAP
in an appropraite location. If you already have SNAP installed and dont't
wish to install a second version (not sure if this is a real issue or not),
you can set the SNAPR_BIN environment variable to the path of the SNAP bin
directory.
}
