% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_aatsr_sst.R
\name{op_aatsr_sst}
\alias{op_aatsr_sst}
\title{Aatsr.SST: snap operator function}
\usage{
op_aatsr_sst(
  operator_id,
  sourceProduct,
  dual = TRUE,
  dualCoefficientsFile = "AVERAGE_POLAR_DUAL_VIEW",
  dualMaskExpression =
    "!cloud_flags_nadir.LAND and !cloud_flags_nadir.CLOUDY and !cloud_flags_nadir.SUN_GLINT and !cloud_flags_fward.LAND and !cloud_flags_fward.CLOUDY and !cloud_flags_fward.SUN_GLINT",
  nadir = TRUE,
  nadirCoefficientsFile = "AVERAGE_POLAR_SINGLE_VIEW",
  nadirMaskExpression =
    "!cloud_flags_nadir.LAND and !cloud_flags_nadir.CLOUDY and !cloud_flags_nadir.SUN_GLINT",
  invalidSstValue = "-999.0f"
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{dual}{Enables/disables generation of the dual-view SST Default value
is 'true'.}

\item{dualCoefficientsFile}{Coefficient file for the dual-view SST Value
must be one of 'AVERAGE_POLAR_DUAL_VIEW', 'AVERAGE_TEMPERATE_DUAL_VIEW',
'AVERAGE_TROPICAL_DUAL_VIEW', 'GRIDDED_POLAR_DUAL_VIEW',
'GRIDDED_TEMPERATE_DUAL_VIEW', 'GRIDDED_TROPICAL_DUAL_VIEW',
'GRIDDED_DUAL_VIEW_IPF'. Default value is 'AVERAGE_POLAR_DUAL_VIEW'.}

\item{dualMaskExpression}{ROI-mask used for the dual-view SST Default value
is '!cloud_flags_nadir.LAND and !cloud_flags_nadir.CLOUDY and
!cloud_flags_nadir.SUN_GLINT and !cloud_flags_fward.LAND and
!cloud_flags_fward.CLOUDY and !cloud_flags_fward.SUN_GLINT'.}

\item{nadir}{Enables/disables generation of the nadir-view SST Default value
is 'true'.}

\item{nadirCoefficientsFile}{Coefficient file for the nadir-view SST Value
must be one of 'AVERAGE_POLAR_SINGLE_VIEW', 'AVERAGE_TEMPERATE_SINGLE_VIEW',
'AVERAGE_TROPICAL_SINGLE_VIEW', 'GRIDDED_POLAR_SINGLE_VIEW',
'GRIDDED_TEMPERATE_SINGLE_VIEW', 'GRIDDED_TROPICAL_SINGLE_VIEW'. Default value
is 'AVERAGE_POLAR_SINGLE_VIEW'.}

\item{nadirMaskExpression}{ROI-mask used for the nadir-view SST Default
value is '!cloud_flags_nadir.LAND and !cloud_flags_nadir.CLOUDY and
!cloud_flags_nadir.SUN_GLINT'.}

\item{invalidSstValue}{Value used to fill invalid SST pixels Default value
is '-999.0f'.}
}
\value{
snap_op_aatsr_sst object
}
\description{
Aatsr.SST: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Computes sea surface temperature (SST) from (A)ATSR products."
}
