% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_aatsr_ungrid.R
\name{op_aatsr_ungrid}
\alias{op_aatsr_ungrid}
\title{AATSR.Ungrid: snap operator function}
\usage{
op_aatsr_ungrid(
  operator_id,
  sourceProduct,
  L1BCharacterisationFile = NULL,
  cornerReferenceFlag = TRUE,
  topographicFlag = FALSE,
  topographyHomogenity = 0.05
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{L1BCharacterisationFile}{L1B characterisation file is needed to
specify first forward pixel and first nadir pixel}

\item{cornerReferenceFlag}{Choose the pixel coordinate reference point for
use in the output file. Check for Corner (default), un-check for Centre.
Default value is 'true'.}

\item{topographicFlag}{Option to apply topographic corrections to tie points
Default value is 'false'.}

\item{topographyHomogenity}{Distance (image coordinates) pixel can be from
tie-point to have topo correction applied Default value is '0.05'.}
}
\value{
snap_op_aatsr_ungrid object
}
\description{
AATSR.Ungrid: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Ungrids (A)ATSR L1B products and extracts geolocation and pixel field of
view data."
}
