% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_adaptive_thresholding.R
\name{op_adaptive_thresholding}
\alias{op_adaptive_thresholding}
\title{AdaptiveThresholding: snap operator function}
\usage{
op_adaptive_thresholding(
  operator_id,
  sourceProduct,
  targetWindowSizeInMeter = 50,
  guardWindowSizeInMeter = 500,
  backgroundWindowSizeInMeter = 800,
  pfa = 6.5,
  estimateBackground = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{targetWindowSizeInMeter}{Target window size Default value is '50'.}

\item{guardWindowSizeInMeter}{Guard window size Default value is '500.0'.}

\item{backgroundWindowSizeInMeter}{Background window size Default value is
'800.0'.}

\item{pfa}{Probability of false alarm Default value is '6.5'.}

\item{estimateBackground}{Rough estimation of background threshold for
quicker processing Default value is 'false'.}
}
\value{
snap_op_adaptive_thresholding object
}
\description{
AdaptiveThresholding: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Detect ships using Constant False Alarm Rate detector."
}
