% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_add_elevation.R
\name{op_add_elevation}
\alias{op_add_elevation}
\title{AddElevation: snap operator function}
\usage{
op_add_elevation(
  operator_id,
  sourceProduct,
  demName = "SRTM 3Sec",
  demResamplingMethod = "BICUBIC_INTERPOLATION",
  externalDEMFile = NULL,
  externalDEMNoDataValue = 0,
  elevationBandName = "elevation"
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{demName}{The digital elevation model. Default value is 'SRTM 3Sec'.}

\item{demResamplingMethod}{Sets parameter 'demResamplingMethod' to \if{html}{\out{<string>}}.
Default value is 'BICUBIC_INTERPOLATION'.}

\item{externalDEMFile}{Sets parameter 'externalDEMFile' to \if{html}{\out{<file>}}.}

\item{externalDEMNoDataValue}{Sets parameter 'externalDEMNoDataValue' to
\if{html}{\out{<double>}}. Default value is '0'.}

\item{elevationBandName}{The elevation band name. Default value is
'elevation'.}
}
\value{
snap_op_add_elevation object
}
\description{
AddElevation: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Creates a DEM band"
}
