% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_add_land_cover.R
\name{op_add_land_cover}
\alias{op_add_land_cover}
\title{AddLandCover: snap operator function}
\usage{
op_add_land_cover(
  operator_id,
  sourceProduct,
  landCoverNames = "AAFC Canada Sand Pct",
  externalFiles = NULL,
  resamplingMethod = "NEAREST_NEIGHBOUR"
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{landCoverNames}{The land cover model. Default value is 'AAFC Canada
Sand Pct'.}

\item{externalFiles}{The external landcover files.}

\item{resamplingMethod}{Sets parameter 'resamplingMethod' to \if{html}{\out{<string>}}.
Default value is 'NEAREST_NEIGHBOUR'.}
}
\value{
snap_op_add_land_cover object
}
\description{
AddLandCover: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Creates a land cover band"
}
