% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_alos_deskewing.R
\name{op_alos_deskewing}
\alias{op_alos_deskewing}
\title{ALOS-Deskewing: snap operator function}
\usage{
op_alos_deskewing(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  demName = "SRTM 3Sec"
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{demName}{The digital elevation model. Default value is 'SRTM 3Sec'.}
}
\value{
snap_op_alos_deskewing object
}
\description{
ALOS-Deskewing: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Deskewing ALOS product"
}
