% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_apply_orbit_file.R
\name{op_apply_orbit_file}
\alias{op_apply_orbit_file}
\title{Apply-Orbit-File: snap operator function}
\usage{
op_apply_orbit_file(
  operator_id,
  sourceProduct,
  orbitType = "Sentinel Precise (Auto Download)",
  polyDegree = 3,
  continueOnFail = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{orbitType}{Sets parameter 'orbitType' to \if{html}{\out{<string>}}. Value must be one
of 'Sentinel Precise (Auto Download)', 'Sentinel Restituted (Auto Download)',
'DORIS Preliminary POR (ENVISAT)', 'DORIS Precise VOR (ENVISAT) (Auto
Download)', 'DELFT Precise (ENVISAT, ERS1&2) (Auto Download)', 'PRARE Precise
(ERS1&2) (Auto Download)', 'Kompsat5 Precise'. Default value is 'Sentinel
Precise (Auto Download)'.}

\item{polyDegree}{Sets parameter 'polyDegree' to \if{html}{\out{<int>}}. Default value is
'3'.}

\item{continueOnFail}{Sets parameter 'continueOnFail' to \if{html}{\out{<boolean>}}. Default
value is 'false'.}
}
\value{
snap_op_apply_orbit_file object
}
\description{
Apply-Orbit-File: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Apply orbit file"
}
