% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_arc_sst.R
\name{op_arc_sst}
\alias{op_arc_sst}
\title{Arc.SST: snap operator function}
\usage{
op_arc_sst(
  operator_id,
  sourceProduct,
  tcwvExpression = 30,
  asdi = TRUE,
  asdiCoefficientsFile = "ASDI_AATSR",
  asdiMaskExpression = NULL,
  dual = TRUE,
  dualCoefficientsFile = "ARC_D2_AATSR",
  dualMaskExpression = NULL,
  nadir = TRUE,
  nadirCoefficientsFile = "ARC_N2_AATSR",
  nadirMaskExpression = NULL,
  invalidSstValue = "-999.0f"
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{tcwvExpression}{TCWV value to use in SST retrieval Default value is
'30.0'.}

\item{asdi}{Enables/disables generation of ATSR Saharan Dust Index Default
value is 'true'.}

\item{asdiCoefficientsFile}{Coefficient file for ASDI Value must be one of
'ASDI_ATSR1', 'ASDI_ATSR2', 'ASDI_AATSR'. Default value is 'ASDI_AATSR'.}

\item{asdiMaskExpression}{ROI-mask used for the ASDI}

\item{dual}{Enables/disables generation of the dual-view SST Default value
is 'true'.}

\item{dualCoefficientsFile}{Coefficient file for the dual-view SST Value
must be one of 'ARC_D2_ATSR1', 'ARC_D2_ATSR2', 'ARC_D2_AATSR', 'ARC_D2_SLSTR',
'ARC_D3_ATSR1', 'ARC_D3_ATSR2', 'ARC_D3_AATSR', 'ARC_D3_SLSTR'. Default value
is 'ARC_D2_AATSR'.}

\item{dualMaskExpression}{ROI-mask used for the dual-view SST}

\item{nadir}{Enables/disables generation of the nadir-view SST Default value
is 'true'.}

\item{nadirCoefficientsFile}{Coefficient file for the nadir-view SST Value
must be one of 'ARC_N2_ATSR1', 'ARC_N2_ATSR2', 'ARC_N2_AATSR', 'ARC_N2_SLSTR',
'ARC_N3_ATSR1', 'ARC_N3_ATSR2', 'ARC_N3_AATSR', 'ARC_N3_SLSTR'. Default value
is 'ARC_N2_AATSR'.}

\item{nadirMaskExpression}{ROI-mask used for the nadir-view SST}

\item{invalidSstValue}{Value used to fill invalid SST pixels Default value
is '-999.0f'.}
}
\value{
snap_op_arc_sst object
}
\description{
Arc.SST: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Computes sea surface temperature (SST) from (A)ATSR and SLSTR products."
}
