% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_arvi_op.R
\name{op_arvi_op}
\alias{op_arvi_op}
\title{ArviOp: snap operator function}
\usage{
op_arvi_op(
  operator_id,
  sourceProduct,
  resampleType = "None",
  redFactor = "1.0F",
  blueFactor = "1.0F",
  nirFactor = "1.0F",
  gammaParameter = "1.0F",
  redSourceBand = NULL,
  blueSourceBand = NULL,
  nirSourceBand = NULL,
  downsampling = NULL,
  upsampling = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{resampleType}{If selected band s differ in size, the resample method
used before computing the index Value must be one of 'None', 'Lowest
resolution', 'Highest resolution'. Default value is 'None'.}

\item{redFactor}{The value of the RED source band is multiplied by this
value. Default value is '1.0F'.}

\item{blueFactor}{The value of the BLUE source band is multiplied by this
value. Default value is '1.0F'.}

\item{nirFactor}{The value of the NIR source band is multiplied by this
value. Default value is '1.0F'.}

\item{gammaParameter}{The gamma parameter is like a weighting function that
depends on the aerosol type Default value is '1.0F'.}

\item{redSourceBand}{The red band for the ARVI computation. If not provided,
the operator will try to find the best fitting band.}

\item{blueSourceBand}{The blue band for the ARVI computation. If not
provided, the operator will try to find the best fitting band.}

\item{nirSourceBand}{The near-infrared band for the ARVI computation. If not
provided, the operator will try to find the best fitting band.}

\item{downsampling}{The method used for aggregation (downsampling to a
coarser resolution). Value must be one of 'First', 'Min', 'Max', 'Mean',
'Median'. Default value is 'First'.}

\item{upsampling}{The method used for interpolation (upsampling to a finer
resolution). Value must be one of 'Nearest', 'Bilinear', 'Bicubic'. Default
value is 'Nearest'.}
}
\value{
snap_op_arvi_op object
}
\description{
ArviOp: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Atmospherically Resistant Vegetation Index belongs to a family of indices
with built-in atmospheric corrections."
}
