% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_azimuth_filter.R
\name{op_azimuth_filter}
\alias{op_azimuth_filter}
\title{AzimuthFilter: snap operator function}
\usage{
op_azimuth_filter(
  operator_id,
  sourceProduct,
  fftLength = 256,
  aziFilterOverlap = 0,
  alphaHamming = 0.75
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{fftLength}{Length of filtering window Value must be one of '64',
'128', '256', '512', '1024', '2048'. Default value is '256'.}

\item{aziFilterOverlap}{Overlap between filtering windows in azimuth
direction [lines] Value must be one of '0', '8', '16', '32', '64', '128',
'256'. Default value is '0'.}

\item{alphaHamming}{Weight for Hamming filter (1 is rectangular window)
Value must be one of '0.5', '0.75', '0.8', '0.9', '1'. Default value is '0.75'.}
}
\value{
snap_op_azimuth_filter object
}
\description{
AzimuthFilter: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Azimuth Filter"
}
