% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_azimuth_shift_estimation_esd.R
\name{op_azimuth_shift_estimation_esd}
\alias{op_azimuth_shift_estimation_esd}
\title{Azimuth-Shift-Estimation-ESD: snap operator function}
\usage{
op_azimuth_shift_estimation_esd(
  operator_id,
  sourceProduct,
  cohThreshold = 0.15,
  numBlocksPerOverlap = 10
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{cohThreshold}{The coherence threshold for outlier removal Valid
interval is (0, 1]. Default value is '0.15'.}

\item{numBlocksPerOverlap}{The number of windows per overlap for ESD Valid
interval is [1, 20]. Default value is '10'.}
}
\value{
snap_op_azimuth_shift_estimation_esd object
}
\description{
Azimuth-Shift-Estimation-ESD: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Estimate azimuth offset for the whole image"
}
