% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_back_geocoding.R
\name{op_back_geocoding}
\alias{op_back_geocoding}
\title{Back-Geocoding: snap operator function}
\usage{
op_back_geocoding(
  operator_id,
  sourceProduct,
  demName = "SRTM 3Sec",
  demResamplingMethod = "BICUBIC_INTERPOLATION",
  externalDEMFile = NULL,
  externalDEMNoDataValue = 0,
  resamplingType = "BISINC_5_POINT_INTERPOLATION",
  maskOutAreaWithoutElevation = TRUE,
  outputRangeAzimuthOffset = FALSE,
  outputDerampDemodPhase = FALSE,
  disableReramp = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{demName}{The digital elevation model. Default value is 'SRTM 3Sec'.}

\item{demResamplingMethod}{Sets parameter 'demResamplingMethod' to \if{html}{\out{<string>}}.
Default value is 'BICUBIC_INTERPOLATION'.}

\item{externalDEMFile}{Sets parameter 'externalDEMFile' to \if{html}{\out{<file>}}.}

\item{externalDEMNoDataValue}{Sets parameter 'externalDEMNoDataValue' to
\if{html}{\out{<double>}}. Default value is '0'.}

\item{resamplingType}{The method to be used when resampling the slave grid
onto the master grid. Default value is 'BISINC_5_POINT_INTERPOLATION'.}

\item{maskOutAreaWithoutElevation}{Sets parameter
'maskOutAreaWithoutElevation' to \if{html}{\out{<boolean>}}. Default value is 'true'.}

\item{outputRangeAzimuthOffset}{Sets parameter 'outputRangeAzimuthOffset' to
\if{html}{\out{<boolean>}}. Default value is 'false'.}

\item{outputDerampDemodPhase}{Sets parameter 'outputDerampDemodPhase' to
\if{html}{\out{<boolean>}}. Default value is 'false'.}

\item{disableReramp}{Sets parameter 'disableReramp' to \if{html}{\out{<boolean>}}. Default
value is 'false'.}
}
\value{
snap_op_back_geocoding object
}
\description{
Back-Geocoding: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Bursts co-registration using orbit and DEM Parameter Options:
-PdemName=\if{html}{\out{<string>}} The digital elevation model. Default value is 'SRTM 3Sec'.
-PdemResamplingMethod=\if{html}{\out{<string>}} Sets parameter 'demResamplingMethod' to
\if{html}{\out{<string>}}. Default value is 'BICUBIC_INTERPOLATION'. -PdisableReramp=\if{html}{\out{<boolean>}}
Sets parameter 'disableReramp' to \if{html}{\out{<boolean>}}. Default value is 'false'.
-PexternalDEMFile=\if{html}{\out{<file>}} Sets parameter 'externalDEMFile' to \if{html}{\out{<file>}}.
-PexternalDEMNoDataValue=\if{html}{\out{<double>}} Sets parameter 'externalDEMNoDataValue' to
\if{html}{\out{<double>}}. Default value is '0'. -PmaskOutAreaWithoutElevation=\if{html}{\out{<boolean>}} Sets
parameter 'maskOutAreaWithoutElevation' to \if{html}{\out{<boolean>}}. Default value is 'true'.
-PoutputDerampDemodPhase=\if{html}{\out{<boolean>}} Sets parameter 'outputDerampDemodPhase' to
\if{html}{\out{<boolean>}}. Default value is 'false'. -PoutputRangeAzimuthOffset=\if{html}{\out{<boolean>}} Sets
parameter 'outputRangeAzimuthOffset' to \if{html}{\out{<boolean>}}. Default value is 'false'.
-PresamplingType=\if{html}{\out{<string>}} The method to be used when resampling the slave grid
onto the master grid. Default value is 'BISINC_5_POINT_INTERPOLATION'."
}
