% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_band_pass_filter.R
\name{op_band_pass_filter}
\alias{op_band_pass_filter}
\title{BandPassFilter: snap operator function}
\usage{
op_band_pass_filter(operator_id, sourceProduct, subband = "low", alpha = 1)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{subband}{Sets parameter 'subband' to \if{html}{\out{<string>}}. Value must be one of
'low', 'high'. Default value is 'low'.}

\item{alpha}{Hamming alpha Value must be one of '0.5', '0.75', '0.8', '0.9',
'1'. Default value is '1'.}
}
\value{
snap_op_band_pass_filter object
}
\description{
BandPassFilter: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Creates a basebanded SLC based on a subband of 1/3 the original bandwidth"
}
