% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_band_select.R
\name{op_band_select}
\alias{op_band_select}
\title{BandSelect: snap operator function}
\usage{
op_band_select(
  operator_id,
  sourceProduct,
  selectedPolarisations = NULL,
  sourceBands = NULL,
  bandNamePattern = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{selectedPolarisations}{The list of polarisations}

\item{sourceBands}{The list of source bands.}

\item{bandNamePattern}{Band name regular expression pattern}
}
\value{
snap_op_band_select object
}
\description{
BandSelect: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Creates a new product with only selected bands"
}
