% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_bands_extractor_op.R
\name{op_bands_extractor_op}
\alias{op_bands_extractor_op}
\title{BandsExtractorOp: snap operator function}
\usage{
op_bands_extractor_op(
  operator_id,
  sourceProduct,
  sourceMaskNames = NULL,
  sourceBandNames = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceMaskNames}{The source masks for the computation.}

\item{sourceBandNames}{The source bands for the computation.}
}
\value{
snap_op_bands_extractor_op object
}
\description{
BandsExtractorOp: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Creates a new product out of the source product containing only the indexes
bands given"
}
