% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_batch_snaphu_unwrap_op.R
\name{op_batch_snaphu_unwrap_op}
\alias{op_batch_snaphu_unwrap_op}
\title{BatchSnaphuUnwrapOp: snap operator function}
\usage{
op_batch_snaphu_unwrap_op(
  operator_id,
  sourceProduct,
  snaphuProcessingLocation = NULL,
  snaphuInstallLocation = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{snaphuProcessingLocation}{Directory that the Snaphu Export operator
wrote to}

\item{snaphuInstallLocation}{Directory to install the SNAPHU binary to}
}
\value{
snap_op_batch_snaphu_unwrap_op object
}
\description{
BatchSnaphuUnwrapOp: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Downloads and executes SNAPHU on a set of two or more interferograms."
}
