% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_biophysical10m_op.R
\name{op_biophysical10m_op}
\alias{op_biophysical10m_op}
\title{Biophysical10mOp: snap operator function}
\usage{
op_biophysical10m_op(
  operator_id,
  sourceProduct,
  sensor = "S2A_10m",
  computeLAI = TRUE,
  computeFapar = TRUE,
  computeFcover = TRUE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sensor}{Sensor Value must be one of 'S2A_10m', 'S2B_10m'. Default
value is 'S2A_10m'.}

\item{computeLAI}{Compute LAI (Leaf Area Index) Default value is 'true'.}

\item{computeFapar}{Compute FAPAR (Fraction of Absorbed Photosynthetically
Active Radiation) Default value is 'true'.}

\item{computeFcover}{Compute FVC (Fraction of Vegetation Cover) Default
value is 'true'.}
}
\value{
snap_op_biophysical10m_op object
}
\description{
Biophysical10mOp: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"The 'Biophysical Processor' operator retrieves LAI from atmospherically
corrected Sentinel-2 products"
}
