% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_biophysical_landsat8op.R
\name{op_biophysical_landsat8op}
\alias{op_biophysical_landsat8op}
\title{BiophysicalLandsat8Op: snap operator function}
\usage{
op_biophysical_landsat8op(
  operator_id,
  sourceProduct,
  computeLAI = TRUE,
  computeFapar = TRUE,
  computeFcover = TRUE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{computeLAI}{Compute LAI (Leaf Area Index) Default value is 'true'.}

\item{computeFapar}{Compute FAPAR (Fraction of Absorbed Photosynthetically
Active Radiation) Default value is 'true'.}

\item{computeFcover}{Compute FVC (Fraction of Vegetation Cover) Default
value is 'true'.}
}
\value{
snap_op_biophysical_landsat8op object
}
\description{
BiophysicalLandsat8Op: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"The 'Biophysical Processor' operator retrieves LAI from atmospherically
corrected Landsat8 products"
}
