% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_biophysical_op.R
\name{op_biophysical_op}
\alias{op_biophysical_op}
\title{BiophysicalOp: snap operator function}
\usage{
op_biophysical_op(
  operator_id,
  sourceProduct,
  sensor = "S2A",
  computeLAI = TRUE,
  computeFapar = TRUE,
  computeFcover = TRUE,
  computeCab = TRUE,
  computeCw = TRUE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sensor}{Sensor Value must be one of 'S2A', 'S2B'. Default value is
'S2A'.}

\item{computeLAI}{Compute LAI (Leaf Area Index) Default value is 'true'.}

\item{computeFapar}{Compute FAPAR (Fraction of Absorbed Photosynthetically
Active Radiation) Default value is 'true'.}

\item{computeFcover}{Compute FVC (Fraction of Vegetation Cover) Default
value is 'true'.}

\item{computeCab}{Compute Cab (Chlorophyll content in the leaf) Default
value is 'true'.}

\item{computeCw}{Compute Cw (Canopy Water Content) Default value is 'true'.}
}
\value{
snap_op_biophysical_op object
}
\description{
BiophysicalOp: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"The 'Biophysical Processor' operator retrieves LAI from atmospherically
corrected Sentinel-2 products"
}
