% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_c2rcc_meris.R
\name{op_c2rcc_meris}
\alias{op_c2rcc_meris}
\title{c2rcc.meris: snap operator function}
\usage{
op_c2rcc_meris(
  operator_id,
  sourceProduct,
  ncepEndProduct = NULL,
  ncepStartProduct = NULL,
  tomsomiEndProduct = NULL,
  tomsomiStartProduct = NULL,
  validPixelExpression = "!l1_flags.INVALID && !l1_flags.LAND_OCEAN",
  salinity = 35,
  temperature = 15,
  ozone = 330,
  press = 1000,
  CHLexp = 1.04,
  thresholdRtosaOOS = 0.003,
  thresholdAcReflecOos = 0.1,
  thresholdCloudTDown865 = 0.955,
  atmosphericAuxDataPath = NULL,
  alternativeNNPath = NULL,
  netSet = "C2RCC-Nets",
  outputAsRrs = FALSE,
  deriveRwFromPathAndTransmittance = FALSE,
  useDefaultSolarFlux = FALSE,
  useEcmwfAuxData = TRUE,
  outputRtoa = TRUE,
  outputRtosaGc = FALSE,
  outputRtosaGcAann = FALSE,
  outputRpath = FALSE,
  outputTdown = FALSE,
  outputTup = FALSE,
  outputAcReflectance = TRUE,
  outputRhown = TRUE,
  outputOos = FALSE,
  outputKd = TRUE,
  outputUncertainties = TRUE,
  CHLfac = NULL,
  TSMexp = NULL,
  TSMfac = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{ncepEndProduct}{The second product providing air pressure values for
pressure interpolation. Use either the TOMSOMI and NCEP products or the
atmosphericAuxdataPath to as source for ozone and air pressure. This is an
optional source.}

\item{ncepStartProduct}{The first product providing air pressure values for
pressure interpolation. Use either the TOMSOMI and NCEP products or the
atmosphericAuxdataPath to as source for ozone and air pressure. This is an
optional source.}

\item{tomsomiEndProduct}{The second product providing ozone values for ozone
interpolation. Use either the TOMSOMI and NCEP products or the
atmosphericAuxdataPath to as source for ozone and air pressure. This is an
optional source.}

\item{tomsomiStartProduct}{The first product providing ozone values for
ozone interpolation. Use either the TOMSOMI and NCEP products or the
atmosphericAuxdataPath to as source for ozone and air pressure. This is an
optional source.}

\item{validPixelExpression}{Defines the pixels which are valid for
processing. Default value is '!l1_flags.INVALID && !l1_flags.LAND_OCEAN'.}

\item{salinity}{The value used as salinity for the scene. Valid interval is
(0.000028, 43). Default value is '35.0'. Parameter unit is 'PSU'.}

\item{temperature}{The value used as temperature for the scene. Valid
interval is (0.000111, 36). Default value is '15.0'. Parameter unit is 'C'.}

\item{ozone}{The value used as ozone if not provided by auxiliary data.
Valid interval is (0, 1000). Default value is '330'. Parameter unit is 'DU'.}

\item{press}{The surface air pressure at sea level if not provided by
auxiliary data. Valid interval is (800, 1040). Default value is '1000'.
Parameter unit is 'hPa'.}

\item{CHLexp}{Chlorophyll exponent ( CHL = iop_apig^CHLexp * CHLfac).
Default value is '1.04'.}

\item{thresholdRtosaOOS}{Threshold for out of scope of nn training dataset
flag for gas corrected top-of-atmosphere reflectances. Default value is
'0.003'.}

\item{thresholdAcReflecOos}{Threshold for out of scope of nn training
dataset flag for atmospherically corrected reflectances. Default value is
'0.1'.}

\item{thresholdCloudTDown865}{Threshold for cloud test based on downwelling
transmittance @865. Default value is '0.955'.}

\item{atmosphericAuxDataPath}{Path to the atmospheric auxiliary data
directory. Use either this or the specified products on the I/O Parameters tab.
If the auxiliary data is not available at this path, the data will
automatically be downloaded.}

\item{alternativeNNPath}{Path to an alternative set of neuronal nets. Use
this to replace the standard set of neuronal nets.}

\item{netSet}{Set of neuronal nets for algorithm. Value must be one of
'C2RCC-Nets', 'C2X-Nets'. Default value is 'C2RCC-Nets'.}

\item{outputAsRrs}{Write remote sensing reflectances instead of water
leaving reflectances. Default value is 'false'.}

\item{deriveRwFromPathAndTransmittance}{Alternative way of calculating water
reflectance. Still experimental. Default value is 'false'.}

\item{useDefaultSolarFlux}{If 'false', use solar flux from source product
Default value is 'false'.}

\item{useEcmwfAuxData}{If selected, the ECMWF auxiliary data (ozon, air
pressure) of the source product is used Default value is 'true'.}

\item{outputRtoa}{Sets parameter 'outputRtoa' to \if{html}{\out{<boolean>}}. Default value is
'true'.}

\item{outputRtosaGc}{Sets parameter 'outputRtosaGc' to \if{html}{\out{<boolean>}}. Default
value is 'false'.}

\item{outputRtosaGcAann}{Sets parameter 'outputRtosaGcAann' to \if{html}{\out{<boolean>}}.
Default value is 'false'.}

\item{outputRpath}{Sets parameter 'outputRpath' to \if{html}{\out{<boolean>}}. Default value
is 'false'.}

\item{outputTdown}{Sets parameter 'outputTdown' to \if{html}{\out{<boolean>}}. Default value
is 'false'.}

\item{outputTup}{Sets parameter 'outputTup' to \if{html}{\out{<boolean>}}. Default value is
'false'.}

\item{outputAcReflectance}{Sets parameter 'outputAcReflectance' to
\if{html}{\out{<boolean>}}. Default value is 'true'.}

\item{outputRhown}{Sets parameter 'outputRhown' to \if{html}{\out{<boolean>}}. Default value
is 'true'.}

\item{outputOos}{Sets parameter 'outputOos' to \if{html}{\out{<boolean>}}. Default value is
'false'.}

\item{outputKd}{Sets parameter 'outputKd' to \if{html}{\out{<boolean>}}. Default value is
'true'.}

\item{outputUncertainties}{Sets parameter 'outputUncertainties' to
\if{html}{\out{<boolean>}}. Default value is 'true'.}

\item{CHLfac}{Chlorophyll factor ( CHL = iop_apig^CHLexp * CHLfac). Default
value is '21.0'.}

\item{TSMexp}{TSM exponent (TSM = TSMfac * iop_btot^TSMexp). Default value
is '3.1'.}

\item{TSMfac}{TSM factor (TSM = TSMfac * iop_btot^TSMexp). Default value is
'1.72'.}
}
\value{
snap_op_c2rcc_meris object
}
\description{
c2rcc.meris: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Performs atmospheric correction and IOP retrieval with uncertainties on
MERIS L1b data products."
}
