% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_c2rcc_modis.R
\name{op_c2rcc_modis}
\alias{op_c2rcc_modis}
\title{c2rcc.modis: snap operator function}
\usage{
op_c2rcc_modis(
  operator_id,
  sourceProduct,
  ncepEndProduct = NULL,
  ncepStartProduct = NULL,
  tomsomiEndProduct = NULL,
  tomsomiStartProduct = NULL,
  validPixelExpression =
    "!(l2_flags.LAND ||  max(rhot_412,max(rhot_443,max(rhot_488,max(rhot_531,max(rhot_547,max(rhot_555,max(rhot_667,max(rhot_678,max(rhot_748,rhot_869)))))))))>0.25)",
  salinity = 35,
  temperature = 15,
  ozone = 330,
  press = 1000,
  atmosphericAuxDataPath = NULL,
  outputRtosa = FALSE,
  outputAsRrs = FALSE,
  outputAngles = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{ncepEndProduct}{The second product providing air pressure values for
pressure interpolation. Use either the TOMSOMI and NCEP products or the
atmosphericAuxdataPath to as source for ozone and air pressure. This is an
optional source.}

\item{ncepStartProduct}{The first product providing air pressure values for
pressure interpolation. Use either the TOMSOMI and NCEP products or the
atmosphericAuxdataPath to as source for ozone and air pressure. This is an
optional source.}

\item{tomsomiEndProduct}{The second product providing ozone values for ozone
interpolation. Use either the TOMSOMI and NCEP products or the
atmosphericAuxdataPath to as source for ozone and air pressure. This is an
optional source.}

\item{tomsomiStartProduct}{The first product providing ozone values for
ozone interpolation. Use either the TOMSOMI and NCEP products or the
atmosphericAuxdataPath to as source for ozone and air pressure. This is an
optional source.}

\item{validPixelExpression}{Defines the pixels which are valid for
processing Default value is '!(l2_flags.LAND ||
max(rhot_412,max(rhot_443,max(rhot_488,max(rhot_531,max(rhot_547,max(rhot_555,max(rhot_667,max(rhot_678,max(rhot_748,rhot_869)))))))))>0.25)'.}

\item{salinity}{The value used as salinity for the scene Valid interval is
(0.000028, 43). Default value is '35.0'. Parameter unit is 'PSU'.}

\item{temperature}{The value used as temperature for the scene Valid
interval is (0.000111, 36). Default value is '15.0'. Parameter unit is 'C'.}

\item{ozone}{The value used as ozone if not provided by auxiliary data Valid
interval is (0, 1000). Default value is '330'. Parameter unit is 'DU'.}

\item{press}{The surface air pressure at sea level if not provided by
auxiliary data Valid interval is (800, 1040). Default value is '1000'.
Parameter unit is 'hPa'.}

\item{atmosphericAuxDataPath}{Path to the atmospheric auxiliary data
directory. Use either this or the specific products. If the auxiliary data
needed for interpolation is not available in this path, the data will
automatically downloaded.}

\item{outputRtosa}{Sets parameter 'outputRtosa' to \if{html}{\out{<boolean>}}. Default value
is 'false'.}

\item{outputAsRrs}{Reflectance values in the target product shall be either
written as remote sensing or water leaving reflectances Default value is
'false'.}

\item{outputAngles}{Sets parameter 'outputAngles' to \if{html}{\out{<boolean>}}. Default
value is 'false'.}
}
\value{
snap_op_c2rcc_modis object
}
\description{
c2rcc.modis: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Performs atmospheric correction and IOP retrieval on MODIS L1C_LAC data
products."
}
