% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_calibration.R
\name{op_calibration}
\alias{op_calibration}
\title{Calibration: snap operator function}
\usage{
op_calibration(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  auxFile = "Latest Auxiliary File",
  externalAuxFile = NULL,
  outputImageInComplex = FALSE,
  outputImageScaleInDb = FALSE,
  createGammaBand = FALSE,
  createBetaBand = FALSE,
  selectedPolarisations = NULL,
  outputSigmaBand = TRUE,
  outputGammaBand = FALSE,
  outputBetaBand = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{auxFile}{The auxiliary file Value must be one of 'Latest Auxiliary
File', 'Product Auxiliary File', 'External Auxiliary File'. Default value is
'Latest Auxiliary File'.}

\item{externalAuxFile}{The antenna elevation pattern gain auxiliary data
file.}

\item{outputImageInComplex}{Output image in complex Default value is
'false'.}

\item{outputImageScaleInDb}{Output image scale Default value is 'false'.}

\item{createGammaBand}{Create gamma0 virtual band Default value is 'false'.}

\item{createBetaBand}{Create beta0 virtual band Default value is 'false'.}

\item{selectedPolarisations}{The list of polarisations}

\item{outputSigmaBand}{Output sigma0 band Default value is 'true'.}

\item{outputGammaBand}{Output gamma0 band Default value is 'false'.}

\item{outputBetaBand}{Output beta0 band Default value is 'false'.}
}
\value{
snap_op_calibration object
}
\description{
Calibration: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Calibration of products"
}
