% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_change_detection.R
\name{op_change_detection}
\alias{op_change_detection}
\title{Change-Detection: snap operator function}
\usage{
op_change_detection(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  maskUpperThreshold = 2,
  maskLowerThreshold = -2,
  includeSourceBands = FALSE,
  outputLogRatio = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{maskUpperThreshold}{Mask upper threshold Default value is '2.0'.}

\item{maskLowerThreshold}{Mask lower threshold Default value is '-2.0'.}

\item{includeSourceBands}{Include source bands Default value is 'false'.}

\item{outputLogRatio}{Output Log Ratio Default value is 'false'.}
}
\value{
snap_op_change_detection object
}
\description{
Change-Detection: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Change Detection."
}
