% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_change_vector_analysis_op.R
\name{op_change_vector_analysis_op}
\alias{op_change_vector_analysis_op}
\title{ChangeVectorAnalysisOp: snap operator function}
\usage{
op_change_vector_analysis_op(
  operator_id,
  sourceProduct1 = NULL,
  sourceProduct2 = NULL,
  sourceBand1 = NULL,
  sourceBand2 = NULL,
  magnitudeThreshold = 0
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct1}{The sources product at the first date. This is a
mandatory source.}

\item{sourceProduct2}{The sources product at the second date. This is a
mandatory source.}

\item{sourceBand1}{Band 1 at the same date}

\item{sourceBand2}{Band 2 at the same date}

\item{magnitudeThreshold}{No change detection magnitude threshold Default
value is '0'.}
}
\value{
snap_op_change_vector_analysis_op object
}
\description{
ChangeVectorAnalysisOp: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"The 'Change Vector Analysis' between two dual bands at two differents
dates."
}
