% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_cloud_prob.R
\name{op_cloud_prob}
\alias{op_cloud_prob}
\title{CloudProb: snap operator function}
\usage{
op_cloud_prob(operator_id, sourceProduct)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}
}
\value{
snap_op_cloud_prob object
}
\description{
CloudProb: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Applies a clear sky conservative cloud detection algorithm. Source Options:
-Ssource=\if{html}{\out{<file>}} The MERIS Level 1b source product. This is a mandatory source."
}
