% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_coherence.R
\name{op_coherence}
\alias{op_coherence}
\title{Coherence: snap operator function}
\usage{
op_coherence(
  operator_id,
  sourceProduct,
  cohWinAz = 10,
  cohWinRg = 10,
  subtractFlatEarthPhase = FALSE,
  srpPolynomialDegree = 5,
  srpNumberPoints = 501,
  orbitDegree = 3,
  squarePixel = TRUE,
  subtractTopographicPhase = FALSE,
  demName = "SRTM 3Sec",
  externalDEMFile = NULL,
  externalDEMNoDataValue = 0,
  externalDEMApplyEGM = TRUE,
  tileExtensionPercent = 100,
  singleMaster = TRUE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{cohWinAz}{Size of coherence estimation window in Azimuth direction
Valid interval is (1, 90]. Default value is '10'.}

\item{cohWinRg}{Size of coherence estimation window in Range direction Valid
interval is (1, 90]. Default value is '10'.}

\item{subtractFlatEarthPhase}{Sets parameter 'subtractFlatEarthPhase' to
\if{html}{\out{<boolean>}}. Default value is 'false'.}

\item{srpPolynomialDegree}{Order of 'Flat earth phase' polynomial Value must
be one of '1', '2', '3', '4', '5', '6', '7', '8'. Default value is '5'.}

\item{srpNumberPoints}{Number of points for the 'flat earth phase'
polynomial estimation Value must be one of '301', '401', '501', '601', '701',
'801', '901', '1001'. Default value is '501'.}

\item{orbitDegree}{Degree of orbit (polynomial) interpolator Value must be
one of '1', '2', '3', '4', '5'. Default value is '3'.}

\item{squarePixel}{Use ground square pixel Default value is 'true'.}

\item{subtractTopographicPhase}{Sets parameter 'subtractTopographicPhase' to
\if{html}{\out{<boolean>}}. Default value is 'false'.}

\item{demName}{The digital elevation model. Default value is 'SRTM 3Sec'.}

\item{externalDEMFile}{Sets parameter 'externalDEMFile' to \if{html}{\out{<file>}}.}

\item{externalDEMNoDataValue}{Sets parameter 'externalDEMNoDataValue' to
\if{html}{\out{<double>}}. Default value is '0'.}

\item{externalDEMApplyEGM}{Sets parameter 'externalDEMApplyEGM' to
\if{html}{\out{<boolean>}}. Default value is 'true'.}

\item{tileExtensionPercent}{Define extension of tile for DEM simulation
(optimization parameter). Default value is '100'.}

\item{singleMaster}{Sets parameter 'singleMaster' to \if{html}{\out{<boolean>}}. Default
value is 'true'.}
}
\value{
snap_op_coherence object
}
\description{
Coherence: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Estimate coherence from stack of coregistered images"
}
