% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_collocate.R
\name{op_collocate}
\alias{op_collocate}
\title{Collocate: snap operator function}
\usage{
op_collocate(
  operator_id,
  sourceProduct,
  reference = NULL,
  secondary = NULL,
  sourceProductPaths = NULL,
  referenceProductName = NULL,
  targetProductType = "COLLOCATED",
  copySecondaryMetadata = FALSE,
  renameReferenceComponents = TRUE,
  renameSecondaryComponents = TRUE,
  referenceComponentPattern = "${ORIGINAL_NAME}_M",
  secondaryComponentPattern = "${ORIGINAL_NAME}_S${SLAVE_NUMBER_ID}",
  resamplingType = "NEAREST_NEIGHBOUR"
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{reference}{The source product which serves as reference. This is an
optional source.}

\item{secondary}{The source product which serves as secondary. This is an
optional source.}

\item{sourceProductPaths}{A comma-separated list of file paths specifying
the source products}

\item{referenceProductName}{The name of the reference product.}

\item{targetProductType}{The product type string for the target product
(informal) Default value is 'COLLOCATED'.}

\item{copySecondaryMetadata}{Copies also the metadata of the secondary
products to the target. Default value is 'false'.}

\item{renameReferenceComponents}{Whether or not components of the reference
product shall be renamed in the target product. Default value is 'true'.}

\item{renameSecondaryComponents}{Whether or not components of the secondary
product(s) shall be renamed in the target product. Default value is 'true'.}

\item{referenceComponentPattern}{The text pattern to be used when renaming
reference components. Default value is '$\{ORIGINAL_NAME\}_M'.}

\item{secondaryComponentPattern}{The text pattern to be used when renaming
secondary components. Default value is '$\{ORIGINAL_NAME\}_S$\{SLAVE_NUMBER_ID\}'.}

\item{resamplingType}{The method to be used when resampling the secondary
grid onto the reference grid. Default value is 'NEAREST_NEIGHBOUR'.}
}
\value{
snap_op_collocate object
}
\description{
Collocate: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Collocates two products based on their geo-codings."
}
