% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_compute_slope_aspect.R
\name{op_compute_slope_aspect}
\alias{op_compute_slope_aspect}
\title{Compute-Slope-Aspect: snap operator function}
\usage{
op_compute_slope_aspect(
  operator_id,
  sourceProduct,
  demName = "SRTM 1Sec HGT",
  demResamplingMethod = "BILINEAR_INTERPOLATION",
  externalDEMFile = NULL,
  externalDEMNoDataValue = 0,
  externalDEMApplyEGM = FALSE,
  demBandName = "elevation"
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{demName}{The digital elevation model. Default value is 'SRTM 1Sec
HGT'.}

\item{demResamplingMethod}{Sets parameter 'demResamplingMethod' to \if{html}{\out{<string>}}.
Default value is 'BILINEAR_INTERPOLATION'.}

\item{externalDEMFile}{Sets parameter 'externalDEMFile' to \if{html}{\out{<file>}}.}

\item{externalDEMNoDataValue}{Sets parameter 'externalDEMNoDataValue' to
\if{html}{\out{<double>}}. Default value is '0'.}

\item{externalDEMApplyEGM}{Sets parameter 'externalDEMApplyEGM' to
\if{html}{\out{<boolean>}}. Default value is 'false'.}

\item{demBandName}{Sets parameter 'demBandName' to \if{html}{\out{<string>}}. Default value
is 'elevation'.}
}
\value{
snap_op_compute_slope_aspect object
}
\description{
Compute-Slope-Aspect: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Compute Slope and Aspect from DEM"
}
