% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_convert_datatype.R
\name{op_convert_datatype}
\alias{op_convert_datatype}
\title{Convert-Datatype: snap operator function}
\usage{
op_convert_datatype(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  targetDataType = "uint8",
  targetScalingStr = "Linear (between 95\% clipped histogram)",
  targetNoDataValue = 0
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{targetDataType}{Sets parameter 'targetDataType' to \if{html}{\out{<string>}}. Value
must be one of 'int8', 'int16', 'int32', 'uint8', 'uint16', 'uint32',
'float32', 'float64'. Default value is 'uint8'.}

\item{targetScalingStr}{Sets parameter 'targetScalingStr' to \if{html}{\out{<string>}}. Value
must be one of 'Truncate', 'Linear (slope and intercept)', 'Linear (between 95\%
clipped histogram)', 'Linear (peak clipped histogram)', 'Logarithmic'. Default
value is 'Linear (between 95\% clipped histogram)'.}

\item{targetNoDataValue}{Sets parameter 'targetNoDataValue' to \if{html}{\out{<double>}}.
Default value is '0'.}
}
\value{
snap_op_convert_datatype object
}
\description{
Convert-Datatype: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Convert product data type"
}
