% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_coregistration_op.R
\name{op_coregistration_op}
\alias{op_coregistration_op}
\title{CoregistrationOp: snap operator function}
\usage{
op_coregistration_op(
  operator_id,
  sourceProduct,
  sourceProduct.1 = NULL,
  masterSourceBand = NULL,
  slaveSourceBand = NULL,
  levels = 6,
  rank = 4,
  iterations = 2,
  radius = "32, 28, 24, 20, 16, 12, 8"
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceProduct.1}{The source product which serves as slave. This is a
mandatory source.}

\item{masterSourceBand}{The master product band}

\item{slaveSourceBand}{The slave product band}

\item{levels}{The number of levels to process the images. Default value is
'6'.}

\item{rank}{Value used to compute the rank. Default value is '4'.}

\item{iterations}{The number of interations for each level and for each
radius. Default value is '2'.}

\item{radius}{The radius integer values splitted by comma. Default value is
'32, 28, 24, 20, 16, 12, 8'.}
}
\value{
snap_op_coregistration_op object
}
\description{
CoregistrationOp: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Coregisters two rasters, not considering their location"
}
