% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_cp_decomposition.R
\name{op_cp_decomposition}
\alias{op_cp_decomposition}
\title{CP-Decomposition: snap operator function}
\usage{
op_cp_decomposition(
  operator_id,
  sourceProduct,
  decomposition = "M-Chi Decomposition",
  windowSizeXStr = 5,
  windowSizeYStr = 5,
  computeAlphaByT3 = TRUE,
  outputRVOG = TRUE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{decomposition}{Sets parameter 'decomposition' to \if{html}{\out{<string>}}. Value must
be one of 'M-Chi Decomposition', 'M-Delta Decomposition', 'H-Alpha
Decomposition', '2 Layer RVOG Model Based Decomposition', 'Model-free
3-component decomposition'. Default value is 'M-Chi Decomposition'.}

\item{windowSizeXStr}{Sets parameter 'windowSizeXStr' to \if{html}{\out{<string>}}. Value
must be one of '3', '5', '7', '9', '11', '13', '15', '17', '19'. Default value
is '5'.}

\item{windowSizeYStr}{Sets parameter 'windowSizeYStr' to \if{html}{\out{<string>}}. Value
must be one of '3', '5', '7', '9', '11', '13', '15', '17', '19'. Default value
is '5'.}

\item{computeAlphaByT3}{Compute alpha by coherency matrix T3 Default value
is 'true'.}

\item{outputRVOG}{Output RVOG parameters mv, ms, alphaS and phi Default
value is 'true'.}
}
\value{
snap_op_cp_decomposition object
}
\description{
CP-Decomposition: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Perform Compact Polarimetric decomposition of a given product"
}
