% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_cp_simulation.R
\name{op_cp_simulation}
\alias{op_cp_simulation}
\title{CP-Simulation: snap operator function}
\usage{
op_cp_simulation(
  operator_id,
  sourceProduct,
  compactMode = "Right Circular Hybrid Mode",
  outputFormat = "Covariance Matrix C2",
  noisePower = -25,
  simulateNoiseFloor = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{compactMode}{The compact mode Value must be one of 'Right Circular
Hybrid Mode', 'Left Circular Hybrid Mode'. Default value is 'Right Circular
Hybrid Mode'.}

\item{outputFormat}{The output simulated compact pol data format Value must
be one of 'Covariance Matrix C2', 'Scatter Vector S2'. Default value is
'Covariance Matrix C2'.}

\item{noisePower}{The noise power Valid interval is (-35, -15). Default
value is '-25'.}

\item{simulateNoiseFloor}{Simulate noise floor Default value is 'false'.}
}
\value{
snap_op_cp_simulation object
}
\description{
CP-Simulation: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Simulation of Compact Pol data from Quad Pol data"
}
