% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_cp_stokes_parameters.R
\name{op_cp_stokes_parameters}
\alias{op_cp_stokes_parameters}
\title{CP-Stokes-Parameters: snap operator function}
\usage{
op_cp_stokes_parameters(
  operator_id,
  sourceProduct,
  windowSizeXStr = 5,
  windowSizeYStr = 5,
  outputStokesVector = FALSE,
  outputDegreeOfPolarization = TRUE,
  outputDegreeOfDepolarization = TRUE,
  outputDegreeOfCircularity = TRUE,
  outputDegreeOfEllipticity = TRUE,
  outputCPR = TRUE,
  outputLPR = TRUE,
  outputRelativePhase = TRUE,
  outputAlphas = TRUE,
  outputConformity = TRUE,
  outputPhasePhi = TRUE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{windowSizeXStr}{Sets parameter 'windowSizeXStr' to \if{html}{\out{<string>}}. Value
must be one of '3', '5', '7', '9', '11', '13', '15', '17', '19'. Default value
is '5'.}

\item{windowSizeYStr}{Sets parameter 'windowSizeYStr' to \if{html}{\out{<string>}}. Value
must be one of '3', '5', '7', '9', '11', '13', '15', '17', '19'. Default value
is '5'.}

\item{outputStokesVector}{Output Stokes vector Default value is 'false'.}

\item{outputDegreeOfPolarization}{Output degree of polarization Default
value is 'true'.}

\item{outputDegreeOfDepolarization}{Output degree of depolarization Default
value is 'true'.}

\item{outputDegreeOfCircularity}{Output degree of circularity Default value
is 'true'.}

\item{outputDegreeOfEllipticity}{Output degree of ellipticity Default value
is 'true'.}

\item{outputCPR}{Output circular polarization ratio Default value is 'true'.}

\item{outputLPR}{Output linear polarization ratio Default value is 'true'.}

\item{outputRelativePhase}{Output relative phase Default value is 'true'.}

\item{outputAlphas}{Output alphas Default value is 'true'.}

\item{outputConformity}{Output conformity coefficient Default value is
'true'.}

\item{outputPhasePhi}{Output phase phi Default value is 'true'.}
}
\value{
snap_op_cp_stokes_parameters object
}
\description{
CP-Stokes-Parameters: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Generates compact polarimetric Stokes child parameters"
}
