% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_create_stack.R
\name{op_create_stack}
\alias{op_create_stack}
\title{CreateStack: snap operator function}
\usage{
op_create_stack(
  operator_id,
  sourceProduct,
  resamplingType = "NONE",
  extent = "Master",
  initialOffsetMethod = "Orbit",
  masterBands = NULL,
  sourceBands = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{resamplingType}{The method to be used when resampling the slave grid
onto the master grid. Default value is 'NONE'.}

\item{extent}{The output image extents. Value must be one of 'Master',
'Minimum', 'Maximum'. Default value is 'Master'.}

\item{initialOffsetMethod}{Method for computing initial offset between
master and slave Value must be one of 'Orbit', 'Product Geolocation'. Default
value is 'Orbit'.}

\item{masterBands}{The list of source bands.}

\item{sourceBands}{The list of source bands.}
}
\value{
snap_op_create_stack object
}
\description{
CreateStack: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Collocates two or more products based on their geo-codings. Parameter
Options: -Pextent=\if{html}{\out{<string>}} The output image extents. Value must be one of
'Master', 'Minimum', 'Maximum'. Default value is 'Master'.
-PinitialOffsetMethod=\if{html}{\out{<string>}} Method for computing initial offset between
master and slave Value must be one of 'Orbit', 'Product Geolocation'. Default
value is 'Orbit'. -PmasterBands=<string,string,string,...> The list of source
bands. -PresamplingType=\if{html}{\out{<string>}} The method to be used when resampling the
slave grid onto the master grid. Default value is 'NONE'.
-PsourceBands=<string,string,string,...> The list of source bands."
}
