% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_cross_channel_snr_correction.R
\name{op_cross_channel_snr_correction}
\alias{op_cross_channel_snr_correction}
\title{Cross-Channel-SNR-Correction: snap operator function}
\usage{
op_cross_channel_snr_correction(operator_id, sourceProduct, windowSize = 5)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{windowSize}{The sliding window size Valid interval is [1, 100].
Default value is '5'.}
}
\value{
snap_op_cross_channel_snr_correction object
}
\description{
Cross-Channel-SNR-Correction: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Compute general polarimetric parameters"
}
