% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_cross_correlation.R
\name{op_cross_correlation}
\alias{op_cross_correlation}
\title{Cross-Correlation: snap operator function}
\usage{
op_cross_correlation(
  operator_id,
  sourceProduct,
  numGCPtoGenerate = 2000,
  coarseRegistrationWindowWidth = 128,
  coarseRegistrationWindowHeight = 128,
  rowInterpFactor = 2,
  columnInterpFactor = 2,
  maxIteration = 10,
  gcpTolerance = 0.5,
  applyFineRegistration = FALSE,
  inSAROptimized = FALSE,
  fineRegistrationWindowWidth = 32,
  fineRegistrationWindowHeight = 32,
  fineRegistrationWindowAccAzimuth = 16,
  fineRegistrationWindowAccRange = 16,
  fineRegistrationOversampling = 16,
  coherenceWindowSize = 3,
  coherenceThreshold = 0.6,
  useSlidingWindow = FALSE,
  computeOffset = FALSE,
  onlyGCPsOnLand = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{numGCPtoGenerate}{The number of GCPs to use in a grid Valid interval
is (10, *). Default value is '2000'.}

\item{coarseRegistrationWindowWidth}{Sets parameter
'coarseRegistrationWindowWidth' to \if{html}{\out{<string>}}. Value must be one of '32', '64',
'128', '256', '512', '1024', '2048'. Default value is '128'.}

\item{coarseRegistrationWindowHeight}{Sets parameter
'coarseRegistrationWindowHeight' to \if{html}{\out{<string>}}. Value must be one of '32', '64',
'128', '256', '512', '1024', '2048'. Default value is '128'.}

\item{rowInterpFactor}{Sets parameter 'rowInterpFactor' to \if{html}{\out{<string>}}. Value
must be one of '2', '4', '8', '16'. Default value is '2'.}

\item{columnInterpFactor}{Sets parameter 'columnInterpFactor' to \if{html}{\out{<string>}}.
Value must be one of '2', '4', '8', '16'. Default value is '2'.}

\item{maxIteration}{The maximum number of iterations Valid interval is (1,
10]. Default value is '10'.}

\item{gcpTolerance}{Tolerance in slave GCP validation check Valid interval
is (0, *). Default value is '0.5'.}

\item{applyFineRegistration}{Sets parameter 'applyFineRegistration' to
\if{html}{\out{<boolean>}}. Default value is 'false'.}

\item{inSAROptimized}{Sets parameter 'inSAROptimized' to \if{html}{\out{<boolean>}}. Default
value is 'false'.}

\item{fineRegistrationWindowWidth}{Sets parameter
'fineRegistrationWindowWidth' to \if{html}{\out{<string>}}. Value must be one of '8', '16',
'32', '64', '128', '256', '512'. Default value is '32'.}

\item{fineRegistrationWindowHeight}{Sets parameter
'fineRegistrationWindowHeight' to \if{html}{\out{<string>}}. Value must be one of '8', '16',
'32', '64', '128', '256', '512'. Default value is '32'.}

\item{fineRegistrationWindowAccAzimuth}{Sets parameter
'fineRegistrationWindowAccAzimuth' to \if{html}{\out{<string>}}. Value must be one of '2', '4',
'8', '16', '32', '64'. Default value is '16'.}

\item{fineRegistrationWindowAccRange}{Sets parameter
'fineRegistrationWindowAccRange' to \if{html}{\out{<string>}}. Value must be one of '2', '4',
'8', '16', '32', '64'. Default value is '16'.}

\item{fineRegistrationOversampling}{Sets parameter
'fineRegistrationOversampling' to \if{html}{\out{<string>}}. Value must be one of '2', '4', '8',
'16', '32', '64'. Default value is '16'.}

\item{coherenceWindowSize}{The coherence window size Valid interval is (1,
16]. Default value is '3'.}

\item{coherenceThreshold}{The coherence threshold Valid interval is (0, *).
Default value is '0.6'.}

\item{useSlidingWindow}{Use sliding window for coherence calculation Default
value is 'false'.}

\item{computeOffset}{Sets parameter 'computeOffset' to \if{html}{\out{<boolean>}}. Default
value is 'false'.}

\item{onlyGCPsOnLand}{Sets parameter 'onlyGCPsOnLand' to \if{html}{\out{<boolean>}}. Default
value is 'false'.}
}
\value{
snap_op_cross_correlation object
}
\description{
Cross-Correlation: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Automatic Selection of Ground Control Points"
}
