% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_cross_resampling.R
\name{op_cross_resampling}
\alias{op_cross_resampling}
\title{CrossResampling: snap operator function}
\usage{
op_cross_resampling(
  operator_id,
  sourceProduct,
  warpPolynomialOrder = 2,
  interpolationMethod = "Cubic convolution (6 points)",
  targetGeometry = "ERS"
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{warpPolynomialOrder}{The order of polynomial function Value must be
one of '1', '2', '3'. Default value is '2'.}

\item{interpolationMethod}{Sets parameter 'interpolationMethod' to \if{html}{\out{<string>}}.
Value must be one of 'Cubic convolution (4 points)', 'Cubic convolution (6
points)', 'Truncated sinc (6 points)', 'Truncated sinc (8 points)', 'Truncated
sinc (16 points)'. Default value is 'Cubic convolution (6 points)'.}

\item{targetGeometry}{Sets parameter 'targetGeometry' to \if{html}{\out{<string>}}. Value
must be one of 'ERS', 'Envisat ASAR'. Default value is 'ERS'.}
}
\value{
snap_op_cross_resampling object
}
\description{
CrossResampling: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Estimate Resampling Polynomial using SAR Image Geometry, and Resample Input
Images"
}
