% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_dark_object_subtraction.R
\name{op_dark_object_subtraction}
\alias{op_dark_object_subtraction}
\title{DarkObjectSubtraction: snap operator function}
\usage{
op_dark_object_subtraction(
  operator_id,
  sourceProduct,
  maskExpression = NULL,
  histogramMinimumPercentile = NULL,
  sourceBandNames = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{maskExpression}{Mask expression defining search area for dark object.}

\item{histogramMinimumPercentile}{Percentile of minimum in image data in
percent (the number means how many percent of the image data are lower than
detected minimum. Value must be one of '0', '1', '5'.}

\item{sourceBandNames}{Bands to be copied to the target. DOS will be applied
on spectral bands only.}
}
\value{
snap_op_dark_object_subtraction object
}
\description{
DarkObjectSubtraction: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Performs dark object subtraction for spectral bands in source product."
}
