% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_deburst_wss.R
\name{op_deburst_wss}
\alias{op_deburst_wss}
\title{DeburstWSS: snap operator function}
\usage{
op_deburst_wss(
  operator_id,
  sourceProduct,
  subSwath = "SS1",
  produceIntensitiesOnly = FALSE,
  average = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{subSwath}{Sets parameter 'subSwath' to \if{html}{\out{<string>}}. Value must be one of
'SS1', 'SS2', 'SS3', 'SS4', 'SS5'. Default value is 'SS1'.}

\item{produceIntensitiesOnly}{Sets parameter 'produceIntensitiesOnly' to
\if{html}{\out{<boolean>}}. Default value is 'false'.}

\item{average}{Sets parameter 'average' to \if{html}{\out{<boolean>}}. Default value is
'false'.}
}
\value{
snap_op_deburst_wss object
}
\description{
DeburstWSS: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Debursts an ASAR WSS product"
}
