% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_demodulate.R
\name{op_demodulate}
\alias{op_demodulate}
\title{Demodulate: snap operator function}
\usage{
op_demodulate(operator_id, sourceProduct)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}
}
\value{
snap_op_demodulate object
}
\description{
Demodulate: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Demodulation and deramping of SLC data Source Options:
-SsourceProduct=\if{html}{\out{<file>}} Sets source 'sourceProduct' to \if{html}{\out{<filepath>}}. This is a
mandatory source."
}
