% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_double_difference_interferogram.R
\name{op_double_difference_interferogram}
\alias{op_double_difference_interferogram}
\title{Double-Difference-Interferogram: snap operator function}
\usage{
op_double_difference_interferogram(
  operator_id,
  sourceProduct,
  outputCoherence = FALSE,
  cohWinSize = 5
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{outputCoherence}{Output coherence for overlapped area Default value is
'false'.}

\item{cohWinSize}{Sets parameter 'cohWinSize' to \if{html}{\out{<string>}}. Value must be one
of '3', '5', '9', '11'. Default value is '5'.}
}
\value{
snap_op_double_difference_interferogram object
}
\description{
Double-Difference-Interferogram: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Compute double difference interferogram"
}
