% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_eap_phase_correction.R
\name{op_eap_phase_correction}
\alias{op_eap_phase_correction}
\title{EAP-Phase-Correction: snap operator function}
\usage{
op_eap_phase_correction(operator_id, sourceProduct)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}
}
\value{
snap_op_eap_phase_correction object
}
\description{
EAP-Phase-Correction: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"EAP Phase Correction Source Options: -Ssource=\if{html}{\out{<file>}} Sets source 'source'
to \if{html}{\out{<filepath>}}. This is a mandatory source."
}
