% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_ellipsoid_correction_gg.R
\name{op_ellipsoid_correction_gg}
\alias{op_ellipsoid_correction_gg}
\title{Ellipsoid-Correction-GG: snap operator function}
\usage{
op_ellipsoid_correction_gg(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  imgResamplingMethod = "BILINEAR_INTERPOLATION",
  mapProjection = "WGS84(DD)"
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{imgResamplingMethod}{Sets parameter 'imgResamplingMethod' to \if{html}{\out{<string>}}.
Value must be one of 'NEAREST_NEIGHBOUR', 'BILINEAR_INTERPOLATION',
'CUBIC_CONVOLUTION'. Default value is 'BILINEAR_INTERPOLATION'.}

\item{mapProjection}{The coordinate reference system in well known text
format Default value is 'WGS84(DD)'.}
}
\value{
snap_op_ellipsoid_correction_gg object
}
\description{
Ellipsoid-Correction-GG: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"GG method for orthorectification"
}
