% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_enhanced_spectral_diversity.R
\name{op_enhanced_spectral_diversity}
\alias{op_enhanced_spectral_diversity}
\title{Enhanced-Spectral-Diversity: snap operator function}
\usage{
op_enhanced_spectral_diversity(
  operator_id,
  sourceProduct,
  fineWinWidthStr = 512,
  fineWinHeightStr = 512,
  fineWinAccAzimuth = 16,
  fineWinAccRange = 16,
  fineWinOversampling = 128,
  xCorrThreshold = 0.1,
  cohThreshold = 0.3,
  numBlocksPerOverlap = 10,
  esdEstimator = "Periodogram",
  weightFunc = "Inv Quadratic",
  temporalBaselineType = "Number of images",
  maxTemporalBaseline = 4,
  integrationMethod = "L1 and L2",
  doNotWriteTargetBands = FALSE,
  useSuppliedRangeShift = FALSE,
  overallRangeShift = 0,
  useSuppliedAzimuthShift = FALSE,
  overallAzimuthShift = 0
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{fineWinWidthStr}{Sets parameter 'fineWinWidthStr' to \if{html}{\out{<string>}}. Value
must be one of '32', '64', '128', '256', '512', '1024', '2048'. Default value
is '512'.}

\item{fineWinHeightStr}{Sets parameter 'fineWinHeightStr' to \if{html}{\out{<string>}}. Value
must be one of '32', '64', '128', '256', '512', '1024', '2048'. Default value
is '512'.}

\item{fineWinAccAzimuth}{Sets parameter 'fineWinAccAzimuth' to \if{html}{\out{<string>}}.
Value must be one of '2', '4', '8', '16', '32', '64'. Default value is '16'.}

\item{fineWinAccRange}{Sets parameter 'fineWinAccRange' to \if{html}{\out{<string>}}. Value
must be one of '2', '4', '8', '16', '32', '64'. Default value is '16'.}

\item{fineWinOversampling}{Sets parameter 'fineWinOversampling' to \if{html}{\out{<string>}}.
Value must be one of '32', '64', '128', '256'. Default value is '128'.}

\item{xCorrThreshold}{The peak cross-correlation threshold Valid interval is
(0, *). Default value is '0.1'.}

\item{cohThreshold}{The coherence threshold for outlier removal Valid
interval is (0, 1]. Default value is '0.3'.}

\item{numBlocksPerOverlap}{The number of windows per overlap for ESD Valid
interval is [1, 20]. Default value is '10'.}

\item{esdEstimator}{ESD estimator used for azimuth shift computation Value
must be one of 'Average', 'Periodogram'. Default value is 'Periodogram'.}

\item{weightFunc}{Weight function of the coherence to use for azimuth shift
estimation Value must be one of 'None', 'Linear', 'Quadratic', 'Inv Quadratic'.
Default value is 'Inv Quadratic'.}

\item{temporalBaselineType}{Baseline type for building the integration
network Value must be one of 'Number of images', 'Number of days'. Default
value is 'Number of images'.}

\item{maxTemporalBaseline}{Maximum temporal baseline (in days or number of
images depending on the Temporal baseline type) between pairs of images to
construct the network. Any number < 1 will generate a network with all of the
possible pairs. Default value is '4'.}

\item{integrationMethod}{Method used for integrating the shifts network.
Value must be one of 'L1', 'L2', 'L1 and L2'. Default value is 'L1 and L2'.}

\item{doNotWriteTargetBands}{Do not write target bands Default value is
'false'.}

\item{useSuppliedRangeShift}{Use user supplied range shift Default value is
'false'.}

\item{overallRangeShift}{The overall range shift Default value is '0.0'.}

\item{useSuppliedAzimuthShift}{Use user supplied azimuth shift Default value
is 'false'.}

\item{overallAzimuthShift}{The overall azimuth shift Default value is '0.0'.}
}
\value{
snap_op_enhanced_spectral_diversity object
}
\description{
Enhanced-Spectral-Diversity: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Estimate constant range and azimuth offsets for a stack of images"
}
