% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_faraday_rotation_correction.R
\name{op_faraday_rotation_correction}
\alias{op_faraday_rotation_correction}
\title{Faraday-Rotation-Correction: snap operator function}
\usage{
op_faraday_rotation_correction(operator_id, sourceProduct, windowSize = 10)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{windowSize}{The sliding window size Valid interval is [1, 100].
Default value is '10'.}
}
\value{
snap_op_faraday_rotation_correction object
}
\description{
Faraday-Rotation-Correction: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Perform Faraday-rotation correction for quad-pol product"
}
