% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_fill_dem_hole.R
\name{op_fill_dem_hole}
\alias{op_fill_dem_hole}
\title{Fill-DEM-Hole: snap operator function}
\usage{
op_fill_dem_hole(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  NoDataValue = 0
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{NoDataValue}{Sets parameter 'NoDataValue' to \if{html}{\out{<double>}}. Default value
is '0.0'.}
}
\value{
snap_op_fill_dem_hole object
}
\description{
Fill-DEM-Hole: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Fill holes in given DEM product file."
}
