% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_flh_mci.R
\name{op_flh_mci}
\alias{op_flh_mci}
\title{FlhMci: snap operator function}
\usage{
op_flh_mci(
  operator_id,
  sourceProduct,
  preset = "NONE",
  lowerBaselineBandName = NULL,
  upperBaselineBandName = NULL,
  signalBandName = NULL,
  lineHeightBandName = NULL,
  slope = TRUE,
  slopeBandName = NULL,
  maskExpression = NULL,
  cloudCorrectionFactor = 1.005,
  invalidFlhMciValue = "NaN"
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{preset}{Sets default values according to the selected preset. The
specific parameters have precedence and override the ones from the preset
Default value is 'NONE'.}

\item{lowerBaselineBandName}{The name for the lower wavelength band defining
the baseline}

\item{upperBaselineBandName}{The name of the upper wavelength band defining
the baseline}

\item{signalBandName}{The name of the signal band, i.e. the band for which
the baseline height is calculated}

\item{lineHeightBandName}{The name of the line height band in the target
product}

\item{slope}{Activates or deactivates calculating the slope parameter
Default value is 'true'.}

\item{slopeBandName}{The name of the slope band in the target product}

\item{maskExpression}{A ROI-mask expression used to identify pixels of
interest}

\item{cloudCorrectionFactor}{The cloud correction factor used during
calculation Default value is '1.005'.}

\item{invalidFlhMciValue}{Value used to fill invalid FLH/MCI pixels Default
value is 'NaN'.}
}
\value{
snap_op_flh_mci object
}
\description{
FlhMci: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Computes fluorescence line height (FLH) or maximum chlorophyll index
(MCI)."
}
