% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_flip.R
\name{op_flip}
\alias{op_flip}
\title{Flip: snap operator function}
\usage{
op_flip(operator_id, sourceProduct, sourceBands = NULL, flipType = "Vertical")
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{flipType}{Sets parameter 'flipType' to \if{html}{\out{<string>}}. Value must be one of
'Horizontal', 'Vertical', 'Horizontal and Vertical'. Default value is
'Vertical'.}
}
\value{
snap_op_flip object
}
\description{
Flip: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"flips a product horizontal/vertical"
}
