% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_forest_cover_change_op.R
\name{op_forest_cover_change_op}
\alias{op_forest_cover_change_op}
\title{ForestCoverChangeOp: snap operator function}
\usage{
op_forest_cover_change_op(
  operator_id,
  previousProduct = NULL,
  recentProduct = NULL,
  forestCoverPercentage = 95,
  landCoverName = "CCILandCover-2015",
  landCoverMapIndices =
    "40, 50, 60, 61, 62, 70, 71, 72, 80, 81, 82, 90, 100, 110, 160, 170",
  mergingCostCriterion = "Spring",
  regionMergingCriterion = "Local Mutual Best Fitting",
  totalIterationsForSecondSegmentation = 10,
  threshold = 5,
  spectralWeight = 0.5,
  shapeWeight = 0.5,
  degreesOfFreedom = 3.3,
  currentProductSourceMaskFile = NULL,
  previousProductSourceMaskFile = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{previousProduct}{The source product to be modified. This is a
mandatory source.}

\item{recentProduct}{The source product to be modified. This is a mandatory
source.}

\item{forestCoverPercentage}{Specifies the percentage of forest cover per
segment Default value is '95.0'.}

\item{landCoverName}{Sets parameter 'landCoverName' to \if{html}{\out{<string>}}. Default
value is 'CCILandCover-2015'.}

\item{landCoverMapIndices}{The indices of forest color from the new added
land cover map Default value is '40, 50, 60, 61, 62, 70, 71, 72, 80, 81, 82,
90, 100, 110, 160, 170'.}

\item{mergingCostCriterion}{The method to compute the region merging. Value
must be one of 'Spring', 'Baatz & Schape', 'Full Lamda Schedule'. Default value
is 'Spring'.}

\item{regionMergingCriterion}{The method to check the region merging. Value
must be one of 'Best Fitting', 'Local Mutual Best Fitting'. Default value is
'Local Mutual Best Fitting'.}

\item{totalIterationsForSecondSegmentation}{The total number of iterations.
Default value is '10'.}

\item{threshold}{The threshold. Default value is '5.0'.}

\item{spectralWeight}{The spectral weight. Default value is '0.5'.}

\item{shapeWeight}{The shape weight. Default value is '0.5'.}

\item{degreesOfFreedom}{Degrees of freedom used for the Chi distribution
trimming process Default value is '3.3'.}

\item{currentProductSourceMaskFile}{A binary raster file to be added as mask
to the output product}

\item{previousProductSourceMaskFile}{A binary raster file to be added as
mask to the output product}
}
\value{
snap_op_forest_cover_change_op object
}
\description{
ForestCoverChangeOp: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Creates forest change masks out of two source products"
}
