% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_fu_classification.R
\name{op_fu_classification}
\alias{op_fu_classification}
\title{FuClassification: snap operator function}
\usage{
op_fu_classification(
  operator_id,
  sourceProduct,
  copyAllSourceBands = FALSE,
  inputIsIrradianceReflectance = FALSE,
  validExpression = NULL,
  reflectanceNamePattern = NULL,
  instrument = "AUTO_DETECT",
  includeDominantLambda = FALSE,
  includeIntermediateResults = TRUE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{copyAllSourceBands}{Weather or not to copy all the bands to the target
product from the source product. Default value is 'false'.}

\item{inputIsIrradianceReflectance}{If enabled, the source reflectances will
be converted to radiance reflectances by dividing it by PI before passing to
the algorithm. Default value is 'false'.}

\item{validExpression}{An expression to filter which pixel are considered.}

\item{reflectanceNamePattern}{The used reflectance band names must match the
given pattern. Useful, if there is more then one spectrum in the product.}

\item{instrument}{The instrument to compute FU for. Default value is
'AUTO_DETECT'.}

\item{includeDominantLambda}{Whether or not the dominant wavelength shall be
derived from the hue angle Default value is 'false'.}

\item{includeIntermediateResults}{Whether or not the intermediate results
shall be written to the target output Default value is 'true'.}
}
\value{
snap_op_fu_classification object
}
\description{
FuClassification: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Colour classification based on the discrete Forel-Ule scale"
}
