% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_fub_water.R
\name{op_fub_water}
\alias{op_fub_water}
\title{FUB.Water: snap operator function}
\usage{
op_fub_water(
  operator_id,
  sourceProduct,
  computeCHL = TRUE,
  computeYS = TRUE,
  computeTSM = TRUE,
  computeAtmCorr = TRUE,
  checkWhetherSuspectIsValid = TRUE,
  expression =
    "not l1_flags.GLINT_RISK and not l1_flags.BRIGHT and not l1_flags.INVALID and not l1_flags.SUSPECT"
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{computeCHL}{Whether chlorophyll-a concentration band shall be computed
Default value is 'true'.}

\item{computeYS}{Whether yellow substances band shall be computed Default
value is 'true'.}

\item{computeTSM}{Whether total suspended matter band shall be computed
Default value is 'true'.}

\item{computeAtmCorr}{Whether atmospheric correction bands shall be computed
Default value is 'true'.}

\item{checkWhetherSuspectIsValid}{Expert parameter. Performs a check whether
the 'l1_flags.SUSPECT' shall be considered in an expression.This parameter is
only considered when the expression contains the term 'and not
l1_flags.SUSPECT' Default value is 'true'.}

\item{expression}{Band maths expression which defines valid pixels. If the
expression is empty,all pixels will be considered. Default value is 'not
l1_flags.GLINT_RISK and not l1_flags.BRIGHT and not l1_flags.INVALID and not
l1_flags.SUSPECT'.}
}
\value{
snap_op_fub_water object
}
\description{
FUB.Water: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"MERIS FUB-CSIRO Coastal Water Processor to retrieve case II water
properties and atmospheric properties"
}
