% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_generic_region_merging_op.R
\name{op_generic_region_merging_op}
\alias{op_generic_region_merging_op}
\title{GenericRegionMergingOp: snap operator function}
\usage{
op_generic_region_merging_op(
  operator_id,
  sourceProduct,
  mergingCostCriterion = "Baatz & Schape",
  regionMergingCriterion = "Local Mutual Best Fitting",
  totalIterationsForSecondSegmentation = 50,
  threshold = 100,
  spectralWeight = 0.5,
  shapeWeight = 0.5,
  sourceBandNames = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{mergingCostCriterion}{The method to compute the region merging. Value
must be one of 'Spring', 'Baatz & Schape', 'Full Lamda Schedule'. Default value
is 'Baatz & Schape'.}

\item{regionMergingCriterion}{The method to check the region merging. Value
must be one of 'Best Fitting', 'Local Mutual Best Fitting'. Default value is
'Local Mutual Best Fitting'.}

\item{totalIterationsForSecondSegmentation}{The total number of iterations.
Default value is '50'.}

\item{threshold}{The threshold. Default value is '100.0'.}

\item{spectralWeight}{The spectral weight. Default value is '0.5'.}

\item{shapeWeight}{The shape weight. Default value is '0.5'.}

\item{sourceBandNames}{The source bands for the computation.}
}
\value{
snap_op_generic_region_merging_op object
}
\description{
GenericRegionMergingOp: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"The 'Generic Region Merging' operator computes the distinct regions from a
product"
}
