% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_goldstein_phase_filtering.R
\name{op_goldstein_phase_filtering}
\alias{op_goldstein_phase_filtering}
\title{GoldsteinPhaseFiltering: snap operator function}
\usage{
op_goldstein_phase_filtering(
  operator_id,
  sourceProduct,
  alpha = 1,
  FFTSizeString = 64,
  windowSizeString = 3,
  useCoherenceMask = FALSE,
  coherenceThreshold = 0.2
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{alpha}{adaptive filter exponent Valid interval is (0, 1]. Default
value is '1.0'.}

\item{FFTSizeString}{Sets parameter 'FFTSizeString' to \if{html}{\out{<string>}}. Value must
be one of '32', '64', '128', '256'. Default value is '64'.}

\item{windowSizeString}{Sets parameter 'windowSizeString' to \if{html}{\out{<string>}}. Value
must be one of '3', '5', '7'. Default value is '3'.}

\item{useCoherenceMask}{Use coherence mask Default value is 'false'.}

\item{coherenceThreshold}{The coherence threshold Valid interval is [0, 1].
Default value is '0.2'.}
}
\value{
snap_op_goldstein_phase_filtering object
}
\description{
GoldsteinPhaseFiltering: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Phase Filtering"
}
