% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_grd_post.R
\name{op_grd_post}
\alias{op_grd_post}
\title{GRD-Post: snap operator function}
\usage{
op_grd_post(operator_id, sourceProduct)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}
}
\value{
snap_op_grd_post object
}
\description{
GRD-Post: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Applies GRD post-processing Source Options: -Ssource=\if{html}{\out{<file>}} Sets source
'source' to \if{html}{\out{<filepath>}}. This is a mandatory source."
}
