% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_horizontal_vertical_motion.R
\name{op_horizontal_vertical_motion}
\alias{op_horizontal_vertical_motion}
\title{HorizontalVerticalMotion: snap operator function}
\usage{
op_horizontal_vertical_motion(
  operator_id,
  sourceProduct,
  refPixelX = 0,
  refPixelY = 0
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{refPixelX}{X position for reference pixel Default value is '0'.}

\item{refPixelY}{Y position for reference pixel Default value is '0'.}
}
\value{
snap_op_horizontal_vertical_motion object
}
\description{
HorizontalVerticalMotion: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Computation of Horizontal/Vertical Motion Components"
}
