% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_iem_hybrid_inversion.R
\name{op_iem_hybrid_inversion}
\alias{op_iem_hybrid_inversion}
\title{IEM-Hybrid-Inversion: snap operator function}
\usage{
op_iem_hybrid_inversion(
  operator_id,
  sourceProduct,
  N = 5,
  M = 5,
  doRemainingOutliersFilter = TRUE,
  lutFile = NULL,
  outputRMS = FALSE,
  outputCL = FALSE,
  thresholdRDC = 0.5
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{N}{closest sigma match from LUT search Default value is '5'.}

\item{M}{Length (pixels) of side of square neighbourhood (M) Default value
is '5'.}

\item{doRemainingOutliersFilter}{Replace remaining outlier with neighbours's
average Default value is 'true'.}

\item{lutFile}{Sets parameter 'lutFile' to \if{html}{\out{<file>}}.}

\item{outputRMS}{Optional rms in output Default value is 'false'.}

\item{outputCL}{Optional cl in output Default value is 'false'.}

\item{thresholdRDC}{RDC deviation threshold Default value is '0.5'.}
}
\value{
snap_op_iem_hybrid_inversion object
}
\description{
IEM-Hybrid-Inversion: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Performs IEM inversion using Hybrid approach"
}
