% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_image_filter.R
\name{op_image_filter}
\alias{op_image_filter}
\title{Image-Filter: snap operator function}
\usage{
op_image_filter(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  selectedFilterName = NULL,
  userDefinedKernelFile = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{selectedFilterName}{Sets parameter 'selectedFilterName' to \if{html}{\out{<string>}}.}

\item{userDefinedKernelFile}{The kernel file}
}
\value{
snap_op_image_filter object
}
\description{
Image-Filter: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Common Image Processing Filters"
}
