% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_import_vector.R
\name{op_import_vector}
\alias{op_import_vector}
\title{Import-Vector: snap operator function}
\usage{
op_import_vector(
  operator_id,
  sourceProduct,
  vectorFile = NULL,
  separateShapes = TRUE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{vectorFile}{Sets parameter 'vectorFile' to \if{html}{\out{<file>}}.}

\item{separateShapes}{Sets parameter 'separateShapes' to \if{html}{\out{<boolean>}}. Default
value is 'true'.}
}
\value{
snap_op_import_vector object
}
\description{
Import-Vector: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Imports a shape file into a product"
}
