% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_integer_interferogram.R
\name{op_integer_interferogram}
\alias{op_integer_interferogram}
\title{IntegerInterferogram: snap operator function}
\usage{
op_integer_interferogram(operator_id, sourceProduct)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}
}
\value{
snap_op_integer_interferogram object
}
\description{
IntegerInterferogram: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Create integer interferogram"
}
