% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_interferogram.R
\name{op_interferogram}
\alias{op_interferogram}
\title{Interferogram: snap operator function}
\usage{
op_interferogram(
  operator_id,
  sourceProduct,
  subtractFlatEarthPhase = TRUE,
  srpPolynomialDegree = 5,
  srpNumberPoints = 501,
  orbitDegree = 3,
  includeCoherence = TRUE,
  cohWinAz = 10,
  cohWinRg = 10,
  squarePixel = TRUE,
  subtractTopographicPhase = FALSE,
  demName = "SRTM 3Sec",
  externalDEMFile = NULL,
  externalDEMNoDataValue = 0,
  externalDEMApplyEGM = TRUE,
  tileExtensionPercent = 100,
  outputElevation = FALSE,
  outputLatLon = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{subtractFlatEarthPhase}{Sets parameter 'subtractFlatEarthPhase' to
\if{html}{\out{<boolean>}}. Default value is 'true'.}

\item{srpPolynomialDegree}{Order of 'Flat earth phase' polynomial Value must
be one of '1', '2', '3', '4', '5', '6', '7', '8'. Default value is '5'.}

\item{srpNumberPoints}{Number of points for the 'flat earth phase'
polynomial estimation Value must be one of '301', '401', '501', '601', '701',
'801', '901', '1001'. Default value is '501'.}

\item{orbitDegree}{Degree of orbit (polynomial) interpolator Value must be
one of '1', '2', '3', '4', '5'. Default value is '3'.}

\item{includeCoherence}{Sets parameter 'includeCoherence' to \if{html}{\out{<boolean>}}.
Default value is 'true'.}

\item{cohWinAz}{Size of coherence estimation window in Azimuth direction
Default value is '10'.}

\item{cohWinRg}{Size of coherence estimation window in Range direction
Default value is '10'.}

\item{squarePixel}{Use ground square pixel Default value is 'true'.}

\item{subtractTopographicPhase}{Sets parameter 'subtractTopographicPhase' to
\if{html}{\out{<boolean>}}. Default value is 'false'.}

\item{demName}{The digital elevation model. Default value is 'SRTM 3Sec'.}

\item{externalDEMFile}{Sets parameter 'externalDEMFile' to \if{html}{\out{<file>}}.}

\item{externalDEMNoDataValue}{Sets parameter 'externalDEMNoDataValue' to
\if{html}{\out{<double>}}. Default value is '0'.}

\item{externalDEMApplyEGM}{Sets parameter 'externalDEMApplyEGM' to
\if{html}{\out{<boolean>}}. Default value is 'true'.}

\item{tileExtensionPercent}{Define extension of tile for DEM simulation
(optimization parameter). Default value is '100'.}

\item{outputElevation}{Sets parameter 'outputElevation' to \if{html}{\out{<boolean>}}.
Default value is 'false'.}

\item{outputLatLon}{Sets parameter 'outputLatLon' to \if{html}{\out{<boolean>}}. Default
value is 'false'.}
}
\value{
snap_op_interferogram object
}
\description{
Interferogram: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Compute interferograms from stack of coregistered S-1 images"
}
