% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_ionospheric_correction.R
\name{op_ionospheric_correction}
\alias{op_ionospheric_correction}
\title{IonosphericCorrection: snap operator function}
\usage{
op_ionospheric_correction(
  operator_id,
  sourceProduct,
  sigma = 81,
  coherenceThreshold = 0.6,
  minCoherence = 0.2
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sigma}{Standard deviation for Gaussian filter Default value is '81'.}

\item{coherenceThreshold}{Coherence threshold Valid interval is [0, 1].
Default value is '0.6'.}

\item{minCoherence}{Minimum coherence for output mask Valid interval is [0,
1]. Default value is '0.2'.}
}
\value{
snap_op_ionospheric_correction object
}
\description{
IonosphericCorrection: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Estimation of Ionospheric Phase Screens Parameter Options:
-PcoherenceThreshold=\if{html}{\out{<double>}} Coherence threshold Valid interval is [0, 1].
Default value is '0.6'. -PminCoherence=\if{html}{\out{<double>}} Minimum coherence for output
mask Valid interval is [0, 1]. Default value is '0.2'. -Psigma=\if{html}{\out{<int>}} Standard
deviation for Gaussian filter Default value is '81'."
}
