% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_ipvi_op.R
\name{op_ipvi_op}
\alias{op_ipvi_op}
\title{IpviOp: snap operator function}
\usage{
op_ipvi_op(
  operator_id,
  sourceProduct,
  resampleType = "None",
  redFactor = "1.0F",
  nirFactor = "1.0F",
  redSourceBand = NULL,
  nirSourceBand = NULL,
  downsampling = NULL,
  upsampling = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{resampleType}{If selected band s differ in size, the resample method
used before computing the index Value must be one of 'None', 'Lowest
resolution', 'Highest resolution'. Default value is 'None'.}

\item{redFactor}{The value of the red source band is multiplied by this
value. Default value is '1.0F'.}

\item{nirFactor}{The value of the NIR source band is multiplied by this
value. Default value is '1.0F'.}

\item{redSourceBand}{The red band for the IPVI computation. If not provided,
the operator will try to find the best fitting band.}

\item{nirSourceBand}{The near-infrared band for the IPVI computation. If not
provided, the operator will try to find the best fitting band.}

\item{downsampling}{The method used for aggregation (downsampling to a
coarser resolution). Value must be one of 'First', 'Min', 'Max', 'Mean',
'Median'. Default value is 'First'.}

\item{upsampling}{The method used for interpolation (upsampling to a finer
resolution). Value must be one of 'Nearest', 'Bilinear', 'Bicubic'. Default
value is 'Nearest'.}
}
\value{
snap_op_ipvi_op object
}
\description{
IpviOp: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Infrared Percentage Vegetation Index retrieves the Isovegetation lines
converge at origin"
}
