% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_kmeans_cluster_analysis.R
\name{op_kmeans_cluster_analysis}
\alias{op_kmeans_cluster_analysis}
\title{KMeansClusterAnalysis: snap operator function}
\usage{
op_kmeans_cluster_analysis(
  operator_id,
  sourceProduct,
  clusterCount = 14,
  iterationCount = 30,
  randomSeed = 31415,
  roiMaskName = NULL,
  sourceBandNames = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{clusterCount}{Number of clusters Valid interval is (0,100]. Default
value is '14'.}

\item{iterationCount}{Number of iterations Valid interval is (0,10000].
Default value is '30'.}

\item{randomSeed}{Seed for the random generator, used for initialising the
algorithm. Default value is '31415'.}

\item{roiMaskName}{The name of the ROI-Mask that should be used.}

\item{sourceBandNames}{The names of the bands being used for the cluster
analysis.}
}
\value{
snap_op_kmeans_cluster_analysis object
}
\description{
KMeansClusterAnalysis: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Performs a K-Means cluster analysis."
}
