% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_land_cover_mask.R
\name{op_land_cover_mask}
\alias{op_land_cover_mask}
\title{Land-Cover-Mask: snap operator function}
\usage{
op_land_cover_mask(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  landCoverBand = NULL,
  validLandCoverClasses = NULL,
  validPixelExpression = NULL,
  includeOtherBands = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{landCoverBand}{Land cover band}

\item{validLandCoverClasses}{Land cover classes to include}

\item{validPixelExpression}{Valid pixel expression}

\item{includeOtherBands}{Add other bands unmasked Default value is 'false'.}
}
\value{
snap_op_land_cover_mask object
}
\description{
Land-Cover-Mask: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Perform a masking based on land cover classes"
}
