% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_land_sea_mask.R
\name{op_land_sea_mask}
\alias{op_land_sea_mask}
\title{Land-Sea-Mask: snap operator function}
\usage{
op_land_sea_mask(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  landMask = TRUE,
  useSRTM = TRUE,
  geometry = NULL,
  invertGeometry = FALSE,
  shorelineExtension = 0
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{landMask}{Sets parameter 'landMask' to \if{html}{\out{<boolean>}}. Default value is
'true'.}

\item{useSRTM}{Sets parameter 'useSRTM' to \if{html}{\out{<boolean>}}. Default value is
'true'.}

\item{geometry}{Sets parameter 'geometry' to \if{html}{\out{<string>}}.}

\item{invertGeometry}{Sets parameter 'invertGeometry' to \if{html}{\out{<boolean>}}. Default
value is 'false'.}

\item{shorelineExtension}{Sets parameter 'shorelineExtension' to \if{html}{\out{<integer>}}.
Default value is '0'.}
}
\value{
snap_op_land_sea_mask object
}
\description{
Land-Sea-Mask: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Creates a bitmask defining land vs ocean."
}
