% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_land_water_mask.R
\name{op_land_water_mask}
\alias{op_land_water_mask}
\title{LandWaterMask: snap operator function}
\usage{
op_land_water_mask(
  operator_id,
  sourceProduct,
  resolution = 50,
  subSamplingFactorX = 1,
  subSamplingFactorY = 1
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{resolution}{Specifies on which resolution the water mask shall be
based. Value must be one of '50', '150', '1000'. Default value is '50'.
Parameter unit is 'm/pixel'.}

\item{subSamplingFactorX}{Specifies the factor between the resolution of the
source product and the watermask in x direction. A value of '1' means no
subsampling at all. Default value is '1'. This is a mandatory parameter.}

\item{subSamplingFactorY}{Specifies the factor between the resolution of the
source product and the watermask iny direction. A value of '1' means no
subsampling at all. Default value is '1'. This is a mandatory parameter.}
}
\value{
snap_op_land_water_mask object
}
\description{
LandWaterMask: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Operator creating a target product with a single band containing a
land/water-mask."
}
