% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_linear_to_fromd_b.R
\name{op_linear_to_fromd_b}
\alias{op_linear_to_fromd_b}
\title{LinearToFromdB: snap operator function}
\usage{
op_linear_to_fromd_b(operator_id, sourceProduct, sourceBands = NULL)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}
}
\value{
snap_op_linear_to_fromd_b object
}
\description{
LinearToFromdB: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Converts bands to/from dB"
}
