% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_mcari_op.R
\name{op_mcari_op}
\alias{op_mcari_op}
\title{McariOp: snap operator function}
\usage{
op_mcari_op(
  operator_id,
  sourceProduct,
  resampleType = "None",
  red1Factor = "1.0F",
  red2Factor = "1.0F",
  greenFactor = "1.0F",
  red1SourceBand = NULL,
  red2SourceBand = NULL,
  greenSourceBand = NULL,
  downsampling = NULL,
  upsampling = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{resampleType}{If selected band s differ in size, the resample method
used before computing the index Value must be one of 'None', 'Lowest
resolution', 'Highest resolution'. Default value is 'None'.}

\item{red1Factor}{The value of the red source band (B4) is multiplied by
this value. Default value is '1.0F'.}

\item{red2Factor}{The value of the red source band (B5) is multiplied by
this value. Default value is '1.0F'.}

\item{greenFactor}{The value of the green source band is multiplied by this
value. Default value is '1.0F'.}

\item{red1SourceBand}{The first red band for the MCARI computation. Choose
B4 for Sentinel-2. If not provided, the operator will try to find the best
fitting band.}

\item{red2SourceBand}{The second red band for the MCARI computation. Choose
B5 for Sentinel-2. If not provided, the operator will try to find the best
fitting band.}

\item{greenSourceBand}{The green band for the MCARI computation. If not
provided, the operator will try to find the best fitting band.}

\item{downsampling}{The method used for aggregation (downsampling to a
coarser resolution). Value must be one of 'First', 'Min', 'Max', 'Mean',
'Median'. Default value is 'First'.}

\item{upsampling}{The method used for interpolation (upsampling to a finer
resolution). Value must be one of 'Nearest', 'Bilinear', 'Bicubic'. Default
value is 'Nearest'.}
}
\value{
snap_op_mcari_op object
}
\description{
McariOp: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Modified Chlorophyll Absorption Ratio Index, developed to be responsive to
chlorophyll variation"
}
