% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_mci_s2.R
\name{op_mci_s2}
\alias{op_mci_s2}
\title{Mci.s2: snap operator function}
\usage{
op_mci_s2(
  operator_id,
  sourceProduct,
  lowerBaselineBandName = NULL,
  upperBaselineBandName = NULL,
  signalBandName = NULL,
  lineHeightBandName = NULL,
  slope = TRUE,
  slopeBandName = NULL,
  maskExpression = NULL,
  cloudCorrectionFactor = 1.005,
  invalidMciValue = "NaN"
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{lowerBaselineBandName}{The name for the lower wavelength band defining
the baseline}

\item{upperBaselineBandName}{The name of the upper wavelength band defining
the baseline}

\item{signalBandName}{The name of the signal band, i.e. the band for which
the baseline height is calculated}

\item{lineHeightBandName}{The name of the MCI band in the target product}

\item{slope}{Activates or deactivates calculating the slope parameter
Default value is 'true'.}

\item{slopeBandName}{The name of the slope band in the target product}

\item{maskExpression}{A ROI-mask expression used to identify pixels of
interest}

\item{cloudCorrectionFactor}{The cloud correction factor used during
calculation Default value is '1.005'.}

\item{invalidMciValue}{Value used to fill invalid MCI pixels Default value
is 'NaN'.}
}
\value{
snap_op_mci_s2 object
}
\description{
Mci.s2: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Computes maximum chlorophyll index (MCI) for Sentinel-2 MSI."
}
