% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_meris_adapt_4to3.R
\name{op_meris_adapt_4to3}
\alias{op_meris_adapt_4to3}
\title{Meris.Adapt.4To3: snap operator function}
\usage{
op_meris_adapt_4to3(operator_id, sourceProduct)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}
}
\value{
snap_op_meris_adapt_4to3 object
}
\description{
Meris.Adapt.4To3: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Provides the adaptation of MERIS L1b products from 4th to 3rd reprocessing.
Source Options: -SsourceProduct=\if{html}{\out{<file>}} The MERIS L1b 4th reprocessing source
product. This is a mandatory source."
}
