% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_meris_correct_radiometry.R
\name{op_meris_correct_radiometry}
\alias{op_meris_correct_radiometry}
\title{Meris.CorrectRadiometry: snap operator function}
\usage{
op_meris_correct_radiometry(
  operator_id,
  sourceProduct,
  doCalibration = TRUE,
  sourceRacFile = NULL,
  targetRacFile = NULL,
  doSmile = TRUE,
  doEqualization = TRUE,
  reproVersion = "AUTO_DETECT",
  doRadToRefl = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{doCalibration}{Whether to perform the calibration. Default value is
'true'.}

\item{sourceRacFile}{The radiometric correction auxiliary file for the
source product. The default
'MER_RAC_AXVIEC20050708_135553_20021224_121445_20041213_220000'}

\item{targetRacFile}{The radiometric correction auxiliary file for the
target product. The default
'MER_RAC_AXVACR20091016_154511_20021224_121445_20041213_220000'}

\item{doSmile}{Whether to perform Smile-effect correction. Default value is
'true'.}

\item{doEqualization}{Perform removal of detector-to-detector systematic
radiometric differences in MERIS L1b data products. Default value is 'true'.}

\item{reproVersion}{The version of the reprocessing the product comes from.
Is only used if equalisation is enabled. Value must be one of 'AUTO_DETECT',
'REPROCESSING_2', 'REPROCESSING_3'. Default value is 'AUTO_DETECT'.}

\item{doRadToRefl}{Whether to perform radiance-to-reflectance conversion.
When selecting ENVISAT as target format, the radiance to reflectance conversion
can not be performed. Default value is 'false'.}
}
\value{
snap_op_meris_correct_radiometry object
}
\description{
Meris.CorrectRadiometry: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Performs radiometric corrections on MERIS L1b data products."
}
