% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_meris_n1patcher.R
\name{op_meris_n1patcher}
\alias{op_meris_n1patcher}
\title{Meris.N1Patcher: snap operator function}
\usage{
op_meris_n1patcher(
  operator_id,
  input = NULL,
  n1 = NULL,
  copyAllTiePoints = FALSE,
  patchedFile = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{input}{The source product provides the data to be written into the
patched file. This is a mandatory source.}

\item{n1}{The N1 file which is used as a template. This is a mandatory
source.}

\item{copyAllTiePoints}{If set to 'false' only the lat and lon tie-points
will be copied to the target product Default value is 'false'.}

\item{patchedFile}{The file to which the patched L1b product is written.
This is a mandatory parameter. Value must not be empty.}
}
\value{
snap_op_meris_n1patcher object
}
\description{
Meris.N1Patcher: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Copies an existing N1 file and replaces the data for the radiance bands"
}
