% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_minimum_distance_classifier.R
\name{op_minimum_distance_classifier}
\alias{op_minimum_distance_classifier}
\title{Minimum-Distance-Classifier: snap operator function}
\usage{
op_minimum_distance_classifier(
  operator_id,
  sourceProduct,
  numTrainSamples = 5000,
  savedClassifierName = NULL,
  doLoadClassifier = FALSE,
  doClassValQuantization = TRUE,
  minClassValue = 0,
  classValStepSize = 5,
  classLevels = 101,
  trainingBands = NULL,
  trainOnRaster = TRUE,
  trainingVectors = NULL,
  featureBands = NULL,
  labelSource = NULL,
  evaluateClassifier = NULL,
  evaluateFeaturePowerSet = FALSE,
  minPowerSetSize = 2,
  maxPowerSetSize = 7
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{numTrainSamples}{The number of training samples Valid interval is
(1,*]. Default value is '5000'.}

\item{savedClassifierName}{The saved classifier name}

\item{doLoadClassifier}{Choose to save or load classifier Default value is
'false'.}

\item{doClassValQuantization}{Quantization for raster traiing Default value
is 'true'.}

\item{minClassValue}{Quantization min class value for raster traiing Default
value is '0.0'.}

\item{classValStepSize}{Quantization step size for raster traiing Default
value is '5.0'.}

\item{classLevels}{Quantization class levels for raster traiing Default
value is '101'.}

\item{trainingBands}{Raster bands to train on}

\item{trainOnRaster}{Train on raster or vector data Default value is 'true'.}

\item{trainingVectors}{Vectors to train on}

\item{featureBands}{Names of bands to be used as features}

\item{labelSource}{'VectorNodeName' or specific Attribute name}

\item{evaluateClassifier}{Evaluate classifier and features}

\item{evaluateFeaturePowerSet}{Evaluate the power set of features Default
value is 'false'.}

\item{minPowerSetSize}{Minimum size of the power set of features Default
value is '2'.}

\item{maxPowerSetSize}{Maximum size of the power set of features Default
value is '7'.}
}
\value{
snap_op_minimum_distance_classifier object
}
\description{
Minimum-Distance-Classifier: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Minimum Distance classifier Parameter Options: -PclassLevels=\if{html}{\out{<int>}}
Quantization class levels for raster traiing Default value is '101'.
-PclassValStepSize=\if{html}{\out{<double>}} Quantization step size for raster traiing Default
value is '5.0'. -PdoClassValQuantization=\if{html}{\out{<boolean>}} Quantization for raster
traiing Default value is 'true'. -PdoLoadClassifier=\if{html}{\out{<boolean>}} Choose to save or
load classifier Default value is 'false'. -PevaluateClassifier=\if{html}{\out{<boolean>}}
Evaluate classifier and features -PevaluateFeaturePowerSet=\if{html}{\out{<boolean>}} Evaluate
the power set of features Default value is 'false'.
-PfeatureBands=<string,string,string,...> Names of bands to be used as features
-PlabelSource=\if{html}{\out{<string>}} 'VectorNodeName' or specific Attribute name
-PmaxPowerSetSize=\if{html}{\out{<integer>}} Maximum size of the power set of features Default
value is '7'. -PminClassValue=\if{html}{\out{<double>}} Quantization min class value for raster
traiing Default value is '0.0'. -PminPowerSetSize=\if{html}{\out{<integer>}} Minimum size of the
power set of features Default value is '2'. -PnumTrainSamples=\if{html}{\out{<int>}} The number
of training samples Valid interval is (1,*]. Default value is '5000'.
-PsavedClassifierName=\if{html}{\out{<string>}} The saved classifier name
-PtrainingBands=<string,string,string,...> Raster bands to train on
-PtrainingVectors=<string,string,string,...> Vectors to train on
-PtrainOnRaster=\if{html}{\out{<boolean>}} Train on raster or vector data Default value is
'true'."
}
