% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_mndwi_op.R
\name{op_mndwi_op}
\alias{op_mndwi_op}
\title{MndwiOp: snap operator function}
\usage{
op_mndwi_op(
  operator_id,
  sourceProduct,
  resampleType = "None",
  greenFactor = "1.0F",
  mirFactor = "1.0F",
  greenSourceBand = NULL,
  mirSourceBand = NULL,
  downsampling = NULL,
  upsampling = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{resampleType}{If selected band s differ in size, the resample method
used before computing the index Value must be one of 'None', 'Lowest
resolution', 'Highest resolution'. Default value is 'None'.}

\item{greenFactor}{The value of the green source band is multiplied by this
value. Default value is '1.0F'.}

\item{mirFactor}{The value of the MIR source band is multiplied by this
value. Default value is '1.0F'.}

\item{greenSourceBand}{The green band for the MNDWI computation. If not
provided, the operator will try to find the best fitting band.}

\item{mirSourceBand}{The mid-infrared band for the MNDWI computation. If not
provided, the operator will try to find the best fitting band.}

\item{downsampling}{The method used for aggregation (downsampling to a
coarser resolution). Value must be one of 'First', 'Min', 'Max', 'Mean',
'Median'. Default value is 'First'.}

\item{upsampling}{The method used for interpolation (upsampling to a finer
resolution). Value must be one of 'Nearest', 'Bilinear', 'Bicubic'. Default
value is 'Nearest'.}
}
\value{
snap_op_mndwi_op object
}
\description{
MndwiOp: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Modified Normalized Difference Water Index, allowing for the measurement of
surface water extent"
}
