% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_mph_chl.R
\name{op_mph_chl}
\alias{op_mph_chl}
\title{MphChl: snap operator function}
\usage{
op_mph_chl(
  operator_id,
  sourceProduct,
  validPixelExpression = NULL,
  cyanoMaxValue = 1000,
  chlThreshForFloatFlag = 500,
  exportMph = FALSE,
  applyLowPassFilter = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{validPixelExpression}{Expression defining pixels considered for
processing. If not set, all valid pixels over water are processed.}

\item{cyanoMaxValue}{Maximum chlorophyll, arithmetically higher values are
capped. Default value is '1000.0'.}

\item{chlThreshForFloatFlag}{Chlorophyll threshold, above which all
cyanobacteria dominated waters are 'float'. Default value is '500.0'.}

\item{exportMph}{Switch to true to write 'mph' band. Default value is
'false'.}

\item{applyLowPassFilter}{Switch to true to apply a 3x3 low-pass filter on
the result. Default value is 'false'.}
}
\value{
snap_op_mph_chl object
}
\description{
MphChl: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"This operator computes maximum peak height of chlorophyll (MPH/CHL)."
}
