% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_multi_master_in_sar.R
\name{op_multi_master_in_sar}
\alias{op_multi_master_in_sar}
\title{MultiMasterInSAR: snap operator function}
\usage{
op_multi_master_in_sar(
  operator_id,
  sourceProduct,
  orbitDegree = 4,
  pairs = NULL,
  includeWavenumber = TRUE,
  includeIncidenceAngle = TRUE,
  includeLatLon = TRUE,
  cohWindowAz = 10,
  cohWindowRg = 10
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{orbitDegree}{Degree of orbit (polynomial) interpolator Value must be
one of '1', '2', '3', '4', '5'. Default value is '4'.}

\item{pairs}{List of interferometric pairs}

\item{includeWavenumber}{Sets parameter 'includeWavenumber' to \if{html}{\out{<boolean>}}.
Default value is 'true'.}

\item{includeIncidenceAngle}{Sets parameter 'includeIncidenceAngle' to
\if{html}{\out{<boolean>}}. Default value is 'true'.}

\item{includeLatLon}{Sets parameter 'includeLatLon' to \if{html}{\out{<boolean>}}. Default
value is 'true'.}

\item{cohWindowAz}{Size of coherence estimation window in azimuth Default
value is '10'.}

\item{cohWindowRg}{Size of coherence estimation window in range Default
value is '10'.}
}
\value{
snap_op_multi_master_in_sar object
}
\description{
MultiMasterInSAR: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Multi-master InSAR processing"
}
