% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_multi_master_stack_generator.R
\name{op_multi_master_stack_generator}
\alias{op_multi_master_stack_generator}
\title{MultiMasterStackGenerator: snap operator function}
\usage{
op_multi_master_stack_generator(
  operator_id,
  sourceProduct,
  outputFolder = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{outputFolder}{Output folder}
}
\value{
snap_op_multi_master_stack_generator object
}
\description{
MultiMasterStackGenerator: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Generates a set of master-slave pairs from a coregistered stack for use in
SBAS processing"
}
