% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_multi_temporal_speckle_filter.R
\name{op_multi_temporal_speckle_filter}
\alias{op_multi_temporal_speckle_filter}
\title{Multi-Temporal-Speckle-Filter: snap operator function}
\usage{
op_multi_temporal_speckle_filter(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  filter = "Lee Sigma",
  filterSizeX = 3,
  filterSizeY = 3,
  dampingFactor = 2,
  estimateENL = FALSE,
  enl = 1,
  numLooksStr = 1,
  windowSize = "7x7",
  targetWindowSizeStr = "3x3",
  sigmaStr = 0.9,
  anSize = 50
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{filter}{Sets parameter 'filter' to \if{html}{\out{<string>}}. Value must be one of
'None', 'Boxcar', 'Median', 'Frost', 'Gamma Map', 'Lee', 'Refined Lee', 'Lee
Sigma', 'IDAN'. Default value is 'Lee Sigma'.}

\item{filterSizeX}{The kernel x dimension Valid interval is (1, 100].
Default value is '3'.}

\item{filterSizeY}{The kernel y dimension Valid interval is (1, 100].
Default value is '3'.}

\item{dampingFactor}{The damping factor (Frost filter only) Valid interval
is (0, 100]. Default value is '2'.}

\item{estimateENL}{Sets parameter 'estimateENL' to \if{html}{\out{<boolean>}}. Default value
is 'false'.}

\item{enl}{The number of looks Valid interval is (0, *). Default value is
'1.0'.}

\item{numLooksStr}{Sets parameter 'numLooksStr' to \if{html}{\out{<string>}}. Value must be
one of '1', '2', '3', '4'. Default value is '1'.}

\item{windowSize}{Sets parameter 'windowSize' to \if{html}{\out{<string>}}. Value must be one
of '5x5', '7x7', '9x9', '11x11', '13x13', '15x15', '17x17'. Default value is
'7x7'.}

\item{targetWindowSizeStr}{Sets parameter 'targetWindowSizeStr' to \if{html}{\out{<string>}}.
Value must be one of '3x3', '5x5'. Default value is '3x3'.}

\item{sigmaStr}{Sets parameter 'sigmaStr' to \if{html}{\out{<string>}}. Value must be one of
'0.5', '0.6', '0.7', '0.8', '0.9'. Default value is '0.9'.}

\item{anSize}{The Adaptive Neighbourhood size Valid interval is (1, 200].
Default value is '50'.}
}
\value{
snap_op_multi_temporal_speckle_filter object
}
\description{
Multi-Temporal-Speckle-Filter: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Speckle Reduction using Multitemporal Filtering"
}
