% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_multitemporal_compositing.R
\name{op_multitemporal_compositing}
\alias{op_multitemporal_compositing}
\title{Multitemporal-Compositing: snap operator function}
\usage{
op_multitemporal_compositing(operator_id, sourceProduct)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}
}
\value{
snap_op_multitemporal_compositing object
}
\description{
Multitemporal-Compositing: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Compute composite image from multi-temporal RTCs Source Options:
-SsourceProduct=\if{html}{\out{<file>}} Sets source 'sourceProduct' to \if{html}{\out{<filepath>}}. This is a
mandatory source."
}
