% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_ndi45op.R
\name{op_ndi45op}
\alias{op_ndi45op}
\title{Ndi45Op: snap operator function}
\usage{
op_ndi45op(
  operator_id,
  sourceProduct,
  resampleType = "None",
  redB4Factor = "1.0F",
  redB5Factor = "1.0F",
  redSourceBand4 = NULL,
  redSourceBand5 = NULL,
  downsampling = NULL,
  upsampling = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{resampleType}{If selected band s differ in size, the resample method
used before computing the index Value must be one of 'None', 'Lowest
resolution', 'Highest resolution'. Default value is 'None'.}

\item{redB4Factor}{The value of the red source band (B4) is multiplied by
this value. Default value is '1.0F'.}

\item{redB5Factor}{The value of the red source band (B5) is multiplied by
this value. Default value is '1.0F'.}

\item{redSourceBand4}{The red band (B4) for the NDI45 computation. If not
provided, the operator will try to find the best fitting band.}

\item{redSourceBand5}{The red band (B5) for the NDI45 computation. If not
provided, the operator will try to find the best fitting band.}

\item{downsampling}{The method used for aggregation (downsampling to a
coarser resolution). Value must be one of 'First', 'Min', 'Max', 'Mean',
'Median'. Default value is 'First'.}

\item{upsampling}{The method used for interpolation (upsampling to a finer
resolution). Value must be one of 'Nearest', 'Bilinear', 'Bicubic'. Default
value is 'Nearest'.}
}
\value{
snap_op_ndi45op object
}
\description{
Ndi45Op: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Normalized Difference Index using bands 4 and 5"
}
