% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_ndwi_op.R
\name{op_ndwi_op}
\alias{op_ndwi_op}
\title{NdwiOp: snap operator function}
\usage{
op_ndwi_op(
  operator_id,
  sourceProduct,
  resampleType = "None",
  mirFactor = "1.0F",
  nirFactor = "1.0F",
  mirSourceBand = NULL,
  nirSourceBand = NULL,
  downsampling = NULL,
  upsampling = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{resampleType}{If selected band s differ in size, the resample method
used before computing the index Value must be one of 'None', 'Lowest
resolution', 'Highest resolution'. Default value is 'None'.}

\item{mirFactor}{The value of the MIR source band is multiplied by this
value. Default value is '1.0F'.}

\item{nirFactor}{The value of the NIR source band is multiplied by this
value. Default value is '1.0F'.}

\item{mirSourceBand}{The mid-infrared band for the NDWI computation. If not
provided, the operator will try to find the best fitting band.}

\item{nirSourceBand}{The near-infrared band for the NDWI computation. If not
provided, the operator will try to find the best fitting band.}

\item{downsampling}{The method used for aggregation (downsampling to a
coarser resolution). Value must be one of 'First', 'Min', 'Max', 'Mean',
'Median'. Default value is 'First'.}

\item{upsampling}{The method used for interpolation (upsampling to a finer
resolution). Value must be one of 'Nearest', 'Bilinear', 'Bicubic'. Default
value is 'Nearest'.}
}
\value{
snap_op_ndwi_op object
}
\description{
NdwiOp: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"The Normalized Difference Water Index was developed for the extraction of
water features"
}
