% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_object_discrimination.R
\name{op_object_discrimination}
\alias{op_object_discrimination}
\title{Object-Discrimination: snap operator function}
\usage{
op_object_discrimination(
  operator_id,
  sourceProduct,
  minTargetSizeInMeter = 50,
  maxTargetSizeInMeter = 600
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{minTargetSizeInMeter}{Minimum target size Default value is '50.0'.}

\item{maxTargetSizeInMeter}{Maximum target size Default value is '600.0'.}
}
\value{
snap_op_object_discrimination object
}
\description{
Object-Discrimination: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Remove false alarms from the detected objects."
}
