% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_offset_tracking.R
\name{op_offset_tracking}
\alias{op_offset_tracking}
\title{Offset-Tracking: snap operator function}
\usage{
op_offset_tracking(
  operator_id,
  sourceProduct,
  gridAzimuthSpacing = 40,
  gridRangeSpacing = 40,
  registrationWindowWidth = 128,
  registrationWindowHeight = 128,
  xCorrThreshold = 0.1,
  registrationOversampling = 16,
  averageBoxSize = 5,
  maxVelocity = 5,
  radius = 4,
  resamplingType = "BICUBIC_INTERPOLATION",
  spatialAverage = TRUE,
  fillHoles = TRUE,
  roiVector = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{gridAzimuthSpacing}{The output grid azimuth spacing in pixels Valid
interval is (1, *). Default value is '40'.}

\item{gridRangeSpacing}{The output grid range spacing in pixels Valid
interval is (1, *). Default value is '40'.}

\item{registrationWindowWidth}{Sets parameter 'registrationWindowWidth' to
\if{html}{\out{<string>}}. Value must be one of '32', '64', '128', '256', '512', '1024', '2048'.
Default value is '128'.}

\item{registrationWindowHeight}{Sets parameter 'registrationWindowHeight' to
\if{html}{\out{<string>}}. Value must be one of '32', '64', '128', '256', '512', '1024', '2048'.
Default value is '128'.}

\item{xCorrThreshold}{The cross-correlation threshold Valid interval is (0,
*). Default value is '0.1'.}

\item{registrationOversampling}{Sets parameter 'registrationOversampling' to
\if{html}{\out{<string>}}. Value must be one of '2', '4', '8', '16', '32', '64', '128', '256',
'512'. Default value is '16'.}

\item{averageBoxSize}{Sets parameter 'averageBoxSize' to \if{html}{\out{<string>}}. Value
must be one of '3', '5', '9', '11'. Default value is '5'.}

\item{maxVelocity}{The threshold for eliminating invalid GCPs Valid interval
is (0, *). Default value is '5.0'.}

\item{radius}{Radius for Hole-Filling Valid interval is (0, *). Default
value is '4'.}

\item{resamplingType}{Methods for velocity interpolation. Value must be one
of 'NEAREST_NEIGHBOUR', 'BILINEAR_INTERPOLATION', 'BICUBIC_INTERPOLATION',
'BISINC_5_POINT_INTERPOLATION', 'CUBIC_CONVOLUTION'. Default value is
'BICUBIC_INTERPOLATION'.}

\item{spatialAverage}{Sets parameter 'spatialAverage' to \if{html}{\out{<boolean>}}. Default
value is 'true'.}

\item{fillHoles}{Sets parameter 'fillHoles' to \if{html}{\out{<boolean>}}. Default value is
'true'.}

\item{roiVector}{Sets parameter 'roiVector' to \if{html}{\out{<string>}}.}
}
\value{
snap_op_offset_tracking object
}
\description{
Offset-Tracking: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Create velocity vectors from offset tracking"
}
