% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_oil_spill_clustering.R
\name{op_oil_spill_clustering}
\alias{op_oil_spill_clustering}
\title{Oil-Spill-Clustering: snap operator function}
\usage{
op_oil_spill_clustering(operator_id, sourceProduct, minClusterSizeInKm2 = 0.1)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{minClusterSizeInKm2}{Minimum cluster size Default value is '0.1'.}
}
\value{
snap_op_oil_spill_clustering object
}
\description{
Oil-Spill-Clustering: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Remove small clusters from detected area."
}
