% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_oil_spill_detection.R
\name{op_oil_spill_detection}
\alias{op_oil_spill_detection}
\title{Oil-Spill-Detection: snap operator function}
\usage{
op_oil_spill_detection(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  backgroundWindowDim = 0.5,
  k = 2
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{backgroundWindowDim}{Background window dimension (km) Default value is
'0.5'.}

\item{k}{Threshold shift from background mean Default value is '2.0'.}
}
\value{
snap_op_oil_spill_detection object
}
\description{
Oil-Spill-Detection: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Detect oil spill."
}
