% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_olci_anomaly_flagging.R
\name{op_olci_anomaly_flagging}
\alias{op_olci_anomaly_flagging}
\title{OlciAnomalyFlagging: snap operator function}
\usage{
op_olci_anomaly_flagging(
  operator_id,
  l1bProduct = NULL,
  writeSlopeInformation = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{l1bProduct}{OLCI L1b or fully compatible product. This is a mandatory
source.}

\item{writeSlopeInformation}{If set to true, the operator adds two bands
containing the maximal spectral slope and the band index where the peak is
observed. Default value is 'false'.}
}
\value{
snap_op_olci_anomaly_flagging object
}
\description{
OlciAnomalyFlagging: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Adds a flagging band indicating saturated pixels and altitude data
overflows"
}
