% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_olci_o2a_harmonisation.R
\name{op_olci_o2a_harmonisation}
\alias{op_olci_o2a_harmonisation}
\title{OlciO2aHarmonisation: snap operator function}
\usage{
op_olci_o2a_harmonisation(
  operator_id,
  l1bProduct = NULL,
  alternativeAltitudeBandName = NULL,
  processOnlyBand13 = TRUE,
  writeHarmonisedRadiances = TRUE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{l1bProduct}{OLCI L1b or fully compatible product. May contain an
optional altitude band, i.e. introduced from an external DEM. This is a
mandatory source.}

\item{alternativeAltitudeBandName}{Name of alternative altitude band in
source product (i.e. introduced from an external DEM). Altitude is expected in
meters.}

\item{processOnlyBand13}{If set to true, only band 13 needed for cloud
detection will be processed, otherwise bands 13-15. Default value is 'true'.}

\item{writeHarmonisedRadiances}{If set to true, harmonised radiances of
processed band(s) will be written to target product. Default value is 'true'.}
}
\value{
snap_op_olci_o2a_harmonisation object
}
\description{
OlciO2aHarmonisation: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Performs O2A band harmonisation on OLCI L1b product. Implements update v4
of R.Preusker, June 2020."
}
