% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_olci_sensor_harmonisation.R
\name{op_olci_sensor_harmonisation}
\alias{op_olci_sensor_harmonisation}
\title{OlciSensorHarmonisation: snap operator function}
\usage{
op_olci_sensor_harmonisation(
  operator_id,
  l1bProduct = NULL,
  performSensorCrossCalibration = FALSE,
  copyInputBands = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{l1bProduct}{OLCI L1b or fully compatible product. This is a mandatory
source.}

\item{performSensorCrossCalibration}{If set to true, in addition to the
camera homogenisation, sensor cross-calibration (i.e. S3A->S3B or S3B->S3A) is
performed using linear regression Default value is 'false'.}

\item{copyInputBands}{If set to true, all bands of the input product (except
for the radiances) are copied to the target product. If set to false, only the
tie-point rasters are copied Default value is 'false'.}
}
\value{
snap_op_olci_sensor_harmonisation object
}
\description{
OlciSensorHarmonisation: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Performs sensor harmonisation on OLCI L1b product. Implements algorithm
described in 'OLCI A/B Tandem Phase Analysis'"
}
