% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_orientation_angle_correction.R
\name{op_orientation_angle_correction}
\alias{op_orientation_angle_correction}
\title{Orientation-Angle-Correction: snap operator function}
\usage{
op_orientation_angle_correction(
  operator_id,
  sourceProduct,
  outputOrientationAngle = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{outputOrientationAngle}{Output Orientation Angle Default value is
'false'.}
}
\value{
snap_op_orientation_angle_correction object
}
\description{
Orientation-Angle-Correction: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Perform polarization orientation angle correction for given coherency
matrix"
}
