% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_oversample.R
\name{op_oversample}
\alias{op_oversample}
\title{Oversample: snap operator function}
\usage{
op_oversample(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  outputImageBy = "Ratio",
  targetImageHeight = 1000,
  targetImageWidth = 1000,
  widthRatio = 2,
  heightRatio = 2,
  rangeSpacing = 12.5,
  azimuthSpacing = 12.5,
  usePRFTileSize = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{outputImageBy}{Sets parameter 'outputImageBy' to \if{html}{\out{<string>}}. Value must
be one of 'Image Size', 'Ratio', 'Pixel Spacing'. Default value is 'Ratio'.}

\item{targetImageHeight}{The row dimension of the output image Default value
is '1000'.}

\item{targetImageWidth}{The col dimension of the output image Default value
is '1000'.}

\item{widthRatio}{The width ratio of the output/input images Default value
is '2.0'.}

\item{heightRatio}{The height ratio of the output/input images Default value
is '2.0'.}

\item{rangeSpacing}{The range pixel spacing Default value is '12.5'.}

\item{azimuthSpacing}{The azimuth pixel spacing Default value is '12.5'.}

\item{usePRFTileSize}{use PRF as azimuth tile size and range line as range
tile size Default value is 'false'.}
}
\value{
snap_op_oversample object
}
\description{
Oversample: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Oversample the datset"
}
