% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_owtclassification.R
\name{op_owtclassification}
\alias{op_owtclassification}
\title{OWTClassification: snap operator function}
\usage{
op_owtclassification(
  operator_id,
  sourceProduct,
  owtType = "COASTAL",
  reflectancesPrefix = "reflec",
  inputReflectanceIs = "RADIANCE_REFLECTANCES",
  writeInputReflectances = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{owtType}{Sets parameter 'owtType' to <oWT_TYPE>. Default value is
'COASTAL'.}

\item{reflectancesPrefix}{Sets parameter 'reflectancesPrefix' to \if{html}{\out{<string>}}.
Default value is 'reflec'.}

\item{inputReflectanceIs}{Sets parameter 'inputReflectanceIs' to
\if{html}{\out{<reflectanceEnum>}}. Default value is 'RADIANCE_REFLECTANCES'.}

\item{writeInputReflectances}{Sets parameter 'writeInputReflectances' to
\if{html}{\out{<boolean>}}. Default value is 'false'.}
}
\value{
snap_op_owtclassification object
}
\description{
OWTClassification: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Performs an optical water type classification based on atmospherically
corrected reflectances."
}
