% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_pca.R
\name{op_pca}
\alias{op_pca}
\title{PCA: snap operator function}
\usage{
op_pca(
  operator_id,
  sourceProduct,
  componentCount = -1,
  roiMaskName = NULL,
  removeNonRoiPixels = FALSE,
  sourceBandNames = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{componentCount}{The maximum number of principal components to compute.
Default value is '-1'.}

\item{roiMaskName}{The name of the ROI mask that should be used.}

\item{removeNonRoiPixels}{Removes all non-ROI pixels in the target product.
Default value is 'false'.}

\item{sourceBandNames}{The names of the bands being used for the cluster
analysis.}
}
\value{
snap_op_pca object
}
\description{
PCA: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Performs a Principal Component Analysis."
}
