% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_pdu_stitching.R
\name{op_pdu_stitching}
\alias{op_pdu_stitching}
\title{PduStitching: snap operator function}
\usage{
op_pdu_stitching(
  operator_id,
  sourceProduct,
  sourceProductPaths = NULL,
  targetDir = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceProductPaths}{A comma-separated list of file paths specifying
the product dissemination units. Each path may contain the wildcards '**'
(matches recursively any directory), '*' (matches any character sequence in
path names) and '?' (matches any single character). If not given, the parameter
'sourceProducts' must be provided.}

\item{targetDir}{The directory to which the stitched product shall be
written. Within this directory, a folder of the SLSTR L1B naming format will be
created. If no target directory is given, the product will be written to the
user directory.}
}
\value{
snap_op_pdu_stitching object
}
\description{
PduStitching: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Stitches multiple SLSTR L1B product dissemination units (PDUs) of the same
orbit to a single product. Parameter Options:
-PsourceProductPaths=<string,string,string,...> A comma-separated list of file
paths specifying the product dissemination units. Each path may contain the
wildcards '**' (matches recursively any directory), '*' (matches any character
sequence in path names) and '?' (matches any single character). If not given,
the parameter 'sourceProducts' must be provided. -PtargetDir=\if{html}{\out{<file>}} The
directory to which the stitched product shall be written. Within this
directory, a folder of the SLSTR L1B naming format will be created. If no
target directory is given, the product will be written to the user directory."
}
