% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_phase_to_displacement.R
\name{op_phase_to_displacement}
\alias{op_phase_to_displacement}
\title{PhaseToDisplacement: snap operator function}
\usage{
op_phase_to_displacement(operator_id, sourceProduct)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}
}
\value{
snap_op_phase_to_displacement object
}
\description{
PhaseToDisplacement: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Phase To Displacement Conversion along LOS Source Options: -Ssource=\if{html}{\out{<file>}}
Sets source 'source' to \if{html}{\out{<filepath>}}. This is a mandatory source."
}
