% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_phase_to_height.R
\name{op_phase_to_height}
\alias{op_phase_to_height}
\title{PhaseToHeight: snap operator function}
\usage{
op_phase_to_height(
  operator_id,
  sourceProduct,
  nPoints = 200,
  nHeights = 3,
  degree1D = 2,
  degree2D = 5,
  orbitDegree = 3
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{nPoints}{Number of points for evaluation of flat earth phase at
different altitudes Value must be one of '100', '200', '300', '400', '500'.
Default value is '200'.}

\item{nHeights}{Number of height samples in range [0,5000) Value must be one
of '2', '3', '4', '5'. Default value is '3'.}

\item{degree1D}{Degree of the 1D polynomial to fit reference phase through.
Value must be one of '1', '2', '3', '4', '5'. Default value is '2'.}

\item{degree2D}{Degree of the 2D polynomial to fit reference phase through.
Value must be one of '1', '2', '3', '4', '5', '6', '7', '8'. Default value is
'5'.}

\item{orbitDegree}{Degree of orbit (polynomial) interpolator Value must be
one of '2', '3', '4', '5'. Default value is '3'.}
}
\value{
snap_op_phase_to_height object
}
\description{
PhaseToHeight: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Phase to Height conversion"
}
