% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_polarimetric_classification.R
\name{op_polarimetric_classification}
\alias{op_polarimetric_classification}
\title{Polarimetric-Classification: snap operator function}
\usage{
op_polarimetric_classification(
  operator_id,
  sourceProduct,
  classification = "H Alpha Wishart",
  windowSize = 5,
  maxIterations = 3,
  numInitialClasses = 90,
  numFinalClasses = 15,
  mixedCategoryThreshold = 0.5,
  decomposition = "Sinclair Decomposition"
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{classification}{Sets parameter 'classification' to \if{html}{\out{<string>}}. Value
must be one of 'Cloude-Pottier', 'Cloude-Pottier Dual Pol', 'H Alpha Wishart',
'H Alpha Wishart Dual Pol', 'Freeman-Durden Wishart', 'General Wishart'.
Default value is 'H Alpha Wishart'.}

\item{windowSize}{The sliding window size Valid interval is (1, 100].
Default value is '5'.}

\item{maxIterations}{The maximum number of iterations Valid interval is [1,
100]. Default value is '3'.}

\item{numInitialClasses}{The initial number of classes Valid interval is [9,
1000]. Default value is '90'.}

\item{numFinalClasses}{The desired number of classes Valid interval is [9,
100]. Default value is '15'.}

\item{mixedCategoryThreshold}{The threshold for classifying pixels to mixed
category Valid interval is (0, *). Default value is '0.5'.}

\item{decomposition}{Sets parameter 'decomposition' to \if{html}{\out{<string>}}. Value must
be one of 'Sinclair Decomposition', 'Pauli Decomposition', 'Freeman-Durden
Decomposition', 'Generalized Freeman-Durden Decomposition', 'Yamaguchi
Decomposition', 'van Zyl Decomposition', 'H-A-Alpha Quad Pol Decomposition',
'Cloude Decomposition', 'Touzi Decomposition'. Default value is 'Sinclair
Decomposition'.}
}
\value{
snap_op_polarimetric_classification object
}
\description{
Polarimetric-Classification: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Perform Polarimetric classification of a given product"
}
