% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_polarimetric_decomposition.R
\name{op_polarimetric_decomposition}
\alias{op_polarimetric_decomposition}
\title{Polarimetric-Decomposition: snap operator function}
\usage{
op_polarimetric_decomposition(
  operator_id,
  sourceProduct,
  decomposition = "Sinclair Decomposition",
  windowSize = 5,
  outputHAAlpha = FALSE,
  outputBetaDeltaGammaLambda = FALSE,
  outputAlpha123 = FALSE,
  outputLambda123 = FALSE,
  outputTouziParamSet0 = FALSE,
  outputTouziParamSet1 = FALSE,
  outputTouziParamSet2 = FALSE,
  outputTouziParamSet3 = FALSE,
  outputHuynenParamSet0 = TRUE,
  outputHuynenParamSet1 = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{decomposition}{Sets parameter 'decomposition' to \if{html}{\out{<string>}}. Value must
be one of 'Sinclair Decomposition', 'Pauli Decomposition', 'Freeman-Durden
Decomposition', 'Generalized Freeman-Durden Decomposition', 'Yamaguchi
Decomposition', 'van Zyl Decomposition', 'H-A-Alpha Quad Pol Decomposition',
'H-Alpha Dual Pol Decomposition', 'Cloude Decomposition', 'Touzi
Decomposition', 'Huynen Decomposition', 'Yang Decomposition', 'Krogager
Decomposition', 'Cameron Decomposition', 'Model-free 3-component
Decomposition', 'Model-free 4-component Decomposition'. Default value is
'Sinclair Decomposition'.}

\item{windowSize}{The sliding window size Valid interval is [1, 100].
Default value is '5'.}

\item{outputHAAlpha}{Output entropy, anisotropy, alpha Default value is
'false'.}

\item{outputBetaDeltaGammaLambda}{Output beta, delta, gamma, lambda Default
value is 'false'.}

\item{outputAlpha123}{Output alpha 1, 2, 3 Default value is 'false'.}

\item{outputLambda123}{Output lambda 1, 2, 3 Default value is 'false'.}

\item{outputTouziParamSet0}{Output psi, tau, alpha, phi Default value is
'false'.}

\item{outputTouziParamSet1}{Output psi1, tau1, alpha1, phi1 Default value is
'false'.}

\item{outputTouziParamSet2}{Output psi2, tau2, alpha2, phi2 Default value is
'false'.}

\item{outputTouziParamSet3}{Output psi3, tau3, alpha3, phi3 Default value is
'false'.}

\item{outputHuynenParamSet0}{Output 2A0_b, B0_plus_B, B0_minus_B Default
value is 'true'.}

\item{outputHuynenParamSet1}{Output A0, B0, B, C, D, E, F, G, H Default
value is 'false'.}
}
\value{
snap_op_polarimetric_decomposition object
}
\description{
Polarimetric-Decomposition: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Perform Polarimetric decomposition of a given product"
}
