% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_polarimetric_matrices.R
\name{op_polarimetric_matrices}
\alias{op_polarimetric_matrices}
\title{Polarimetric-Matrices: snap operator function}
\usage{
op_polarimetric_matrices(operator_id, sourceProduct, matrix = "T3")
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{matrix}{The covariance or coherency matrix Value must be one of 'C2',
'C3', 'C4', 'T3', 'T4'. Default value is 'T3'.}
}
\value{
snap_op_polarimetric_matrices object
}
\description{
Polarimetric-Matrices: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Generates covariance or coherency matrix for given product"
}
