% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_polarimetric_parameters.R
\name{op_polarimetric_parameters}
\alias{op_polarimetric_parameters}
\title{Polarimetric-Parameters: snap operator function}
\usage{
op_polarimetric_parameters(
  operator_id,
  sourceProduct,
  useMeanMatrix = TRUE,
  windowSizeXStr = 5,
  windowSizeYStr = 5,
  outputSpan = TRUE,
  outputPedestalHeight = FALSE,
  outputRVI = FALSE,
  outputRFDI = FALSE,
  outputCSI = FALSE,
  outputVSI = FALSE,
  outputBMI = FALSE,
  outputITI = FALSE,
  outputHHVVRatio = FALSE,
  outputHHHVRatio = FALSE,
  outputVVVHRatio = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{useMeanMatrix}{Use mean coherency or covariance matrix Default value
is 'true'.}

\item{windowSizeXStr}{Sets parameter 'windowSizeXStr' to \if{html}{\out{<string>}}. Value
must be one of '3', '5', '7', '9', '11', '13', '15', '17', '19'. Default value
is '5'.}

\item{windowSizeYStr}{Sets parameter 'windowSizeYStr' to \if{html}{\out{<string>}}. Value
must be one of '3', '5', '7', '9', '11', '13', '15', '17', '19'. Default value
is '5'.}

\item{outputSpan}{Output Span Default value is 'true'.}

\item{outputPedestalHeight}{Output pedestal height Default value is 'false'.}

\item{outputRVI}{Output RVI Default value is 'false'.}

\item{outputRFDI}{Output RFDI Default value is 'false'.}

\item{outputCSI}{Output CSI Default value is 'false'.}

\item{outputVSI}{Output VSI Default value is 'false'.}

\item{outputBMI}{Output BMI Default value is 'false'.}

\item{outputITI}{Output ITI Default value is 'false'.}

\item{outputHHVVRatio}{Output Co-Pol HH/VV Default value is 'false'.}

\item{outputHHHVRatio}{Output Cross-Pol HH/HV Default value is 'false'.}

\item{outputVVVHRatio}{Output Cross-Pol VV/VH Default value is 'false'.}
}
\value{
snap_op_polarimetric_parameters object
}
\description{
Polarimetric-Parameters: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Compute general polarimetric parameters"
}
