% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_polarimetric_speckle_filter.R
\name{op_polarimetric_speckle_filter}
\alias{op_polarimetric_speckle_filter}
\title{Polarimetric-Speckle-Filter: snap operator function}
\usage{
op_polarimetric_speckle_filter(
  operator_id,
  sourceProduct,
  filter = "Refined Lee Filter",
  filterSize = 5,
  numLooksStr = 1,
  windowSize = "7x7",
  targetWindowSizeStr = "3x3",
  anSize = 50,
  sigmaStr = 0.9,
  searchWindowSizeStr = 15,
  patchSizeStr = 5,
  scaleSizeStr = 1
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{filter}{Sets parameter 'filter' to \if{html}{\out{<string>}}. Value must be one of 'Box
Car Filter', 'IDAN Filter', 'Refined Lee Filter', 'Improved Lee Sigma Filter'.
Default value is 'Refined Lee Filter'.}

\item{filterSize}{The filter size Valid interval is (1, 100]. Default value
is '5'.}

\item{numLooksStr}{Sets parameter 'numLooksStr' to \if{html}{\out{<string>}}. Value must be
one of '1', '2', '3', '4'. Default value is '1'.}

\item{windowSize}{Sets parameter 'windowSize' to \if{html}{\out{<string>}}. Value must be one
of '5x5', '7x7', '9x9', '11x11', '13x13', '15x15', '17x17'. Default value is
'7x7'.}

\item{targetWindowSizeStr}{Sets parameter 'targetWindowSizeStr' to \if{html}{\out{<string>}}.
Value must be one of '3x3', '5x5'. Default value is '3x3'.}

\item{anSize}{The Adaptive Neighbourhood size Valid interval is (1, 200].
Default value is '50'.}

\item{sigmaStr}{Sets parameter 'sigmaStr' to \if{html}{\out{<string>}}. Value must be one of
'0.5', '0.6', '0.7', '0.8', '0.9'. Default value is '0.9'.}

\item{searchWindowSizeStr}{The search window size Value must be one of '3',
'5', '7', '9', '11', '13', '15', '17', '19', '21', '23', '25'. Default value is
'15'.}

\item{patchSizeStr}{The patch size Value must be one of '3', '5', '7', '9',
'11'. Default value is '5'.}

\item{scaleSizeStr}{The scale size Value must be one of '0', '1', '2'.
Default value is '1'.}
}
\value{
snap_op_polarimetric_speckle_filter object
}
\description{
Polarimetric-Speckle-Filter: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Polarimetric Speckle Reduction"
}
