% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_ppe_filtering.R
\name{op_ppe_filtering}
\alias{op_ppe_filtering}
\title{PpeFiltering: snap operator function}
\usage{
op_ppe_filtering(
  operator_id,
  sourceProduct,
  cutOff = 0.7,
  numberOfMAD = 10,
  validExpression = "not quality_flags_land"
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{cutOff}{Minimum threshold to differentiate with the neighboring
pixels. Default value is '0.7'.}

\item{numberOfMAD}{Multiplier of MAD (Median Absolute Deviation) used for
the threshold. Default value is '10'.}

\item{validExpression}{An expression to filter which pixel are considered.
Default value is 'not quality_flags_land'.}
}
\value{
snap_op_ppe_filtering object
}
\description{
PpeFiltering: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Performs Prompt Particle Event (PPE) filtering on OLCI L1B"
}
