% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_principle_components.R
\name{op_principle_components}
\alias{op_principle_components}
\title{Principle-Components: snap operator function}
\usage{
op_principle_components(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  selectEigenvaluesBy = "Eigenvalue Threshold",
  eigenvalueThreshold = 100,
  numPCA = 1,
  showEigenvalues = 1,
  subtractMeanImage = 1
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{selectEigenvaluesBy}{Sets parameter 'selectEigenvaluesBy' to \if{html}{\out{<string>}}.
Value must be one of 'Eigenvalue Threshold', 'Number of Eigenvalues'. Default
value is 'Eigenvalue Threshold'.}

\item{eigenvalueThreshold}{The threshold for selecting eigenvalues Valid
interval is (0, 100]. Default value is '100'.}

\item{numPCA}{The number of PCA images output Valid interval is (0, 100].
Default value is '1'.}

\item{showEigenvalues}{Show the eigenvalues Default value is '1'.}

\item{subtractMeanImage}{Subtract mean image Default value is '1'.}
}
\value{
snap_op_principle_components object
}
\description{
Principle-Components: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Principle Component Analysis"
}
