% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_product_set_reader.R
\name{op_product_set_reader}
\alias{op_product_set_reader}
\title{ProductSet-Reader: snap operator function}
\usage{
op_product_set_reader(operator_id, fileList = NULL)
}
\arguments{
\item{operator_id}{character operator id}

\item{fileList}{Sets parameter 'fileList' to <string,string,string,...>.}
}
\value{
snap_op_product_set_reader object
}
\description{
ProductSet-Reader: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Adds a list of sources Parameter Options:
-PfileList=<string,string,string,...> Sets parameter 'fileList' to
<string,string,string,...>."
}
