% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_rad2refl.R
\name{op_rad2refl}
\alias{op_rad2refl}
\title{Rad2Refl: snap operator function}
\usage{
op_rad2refl(
  operator_id,
  sourceProduct,
  sensor = "OLCI",
  conversionMode = "RAD_TO_REFL",
  copyTiePointGrids = FALSE,
  copyFlagBandsAndMasks = FALSE,
  copyNonSpectralBands = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sensor}{The sensor Value must be one of 'MERIS', 'OLCI', 'SLSTR_500m'.
Default value is 'OLCI'.}

\item{conversionMode}{Conversion mode: from rad to refl, or backwards Value
must be one of 'RAD_TO_REFL', 'REFL_TO_RAD'. Default value is 'RAD_TO_REFL'.}

\item{copyTiePointGrids}{If set, all tie point grids from source product are
written to target product Default value is 'false'.}

\item{copyFlagBandsAndMasks}{If set, all flag bands and masks from source
product are written to target product Default value is 'false'.}

\item{copyNonSpectralBands}{If set, all other non-spectral bands from source
product are written to target product Default value is 'false'.}
}
\value{
snap_op_rad2refl object
}
\description{
Rad2Refl: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Provides conversion from radiances to reflectances or backwards."
}
