% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_range_filter.R
\name{op_range_filter}
\alias{op_range_filter}
\title{RangeFilter: snap operator function}
\usage{
op_range_filter(
  operator_id,
  sourceProduct,
  fftLength = 8,
  alphaHamming = 0.75,
  nlMean = 15,
  snrThresh = 5,
  ovsmpFactor = 1,
  doWeightCorrel = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{fftLength}{Length of filtering window Value must be one of '8', '16',
'32', '64', '128', '256', '512', '1024'. Default value is '8'.}

\item{alphaHamming}{Weight for Hamming filter (1 is rectangular window)
Value must be one of '0.5', '0.75', '0.8', '0.9', '1'. Default value is '0.75'.}

\item{nlMean}{Input value for (walking) mean averaging to reduce noise.
Value must be one of '5', '10', '15', '20', '25'. Default value is '15'.}

\item{snrThresh}{Threshold on SNR for peak estimation Value must be one of
'3', '4', '5', '6', '7'. Default value is '5'.}

\item{ovsmpFactor}{Oversampling factor (in range only). Value must be one of
'1', '2', '4'. Default value is '1'.}

\item{doWeightCorrel}{Use weight values to bias higher frequencies Value
must be one of 'true', 'false'. Default value is 'off'.}
}
\value{
snap_op_range_filter object
}
\description{
RangeFilter: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Range Filter"
}
