% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_read.R
\name{op_read}
\alias{op_read}
\title{Read: snap operator function}
\usage{
op_read(
  operator_id,
  file = NULL,
  formatName = NULL,
  pixelRegion = NULL,
  geometryRegion = NULL,
  copyMetadata = TRUE,
  sourceBands = NULL,
  sourceMasks = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{file}{The file from which the data product is read. This is a
mandatory parameter. Value must not be empty.}

\item{formatName}{An (optional) format name. Value must not be empty.}

\item{pixelRegion}{The subset region in pixel coordinates. Use the following
format: \{x>,\{y>,\{width>,\{height> If not given, the entire scene is used. The
'geoRegion' parameter has precedence over this parameter.}

\item{geometryRegion}{The subset region in geographical coordinates using
WKT-format, e.g. POLYGON((\{lon1\} \{lat1\}, \{lon2\} \{lat2\}, ..., \{lon1\} \{lat1\}))
(make sure to quote the option due to spaces in \{geometry\}). If not given, the
entire scene is used.}

\item{copyMetadata}{Whether to copy the metadata of the source product.
Default value is 'true'.}

\item{sourceBands}{The list of source bands.}

\item{sourceMasks}{The list of source masks.}
}
\value{
snap_op_read object
}
\description{
Read: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Reads a data product from a given file location. Parameter Options:
-PcopyMetadata=\if{html}{\out{<boolean>}} Whether to copy the metadata of the source product.
Default value is 'true'. -Pfile=\if{html}{\out{<file>}} The file from which the data product is
read. This is a mandatory parameter. Value must not be empty.
-PformatName=\if{html}{\out{<string>}} An (optional) format name. Value must not be empty.
-PgeometryRegion=\if{html}{\out{<geometry>}} The subset region in geographical coordinates using
WKT-format, e.g. POLYGON((\{lon1\} \{lat1\}, \{lon2\} \{lat2\}, ..., \{lon1\} \{lat1\}))
(make sure to quote the option due to spaces in \{geometry\}). If not given, the
entire scene is used. -PpixelRegion=\if{html}{\out{<rectangle>}} The subset region in pixel
coordinates. Use the following format: \{x>,\{y>,\{width>,\{height> If not given,
the entire scene is used. The 'geoRegion' parameter has precedence over this
parameter. -PsourceBands=<string,string,string,...> The list of source bands.
-PsourceMasks=<string,string,string,...> The list of source masks.
-PuseAdvancedOptions=\if{html}{\out{<boolean>}} Whether to use advanced options for reading of
the source product. Default value is 'false'."
}
