% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_reflectance_to_radiance_op.R
\name{op_reflectance_to_radiance_op}
\alias{op_reflectance_to_radiance_op}
\title{ReflectanceToRadianceOp: snap operator function}
\usage{
op_reflectance_to_radiance_op(
  operator_id,
  sourceProduct,
  solarIrradiance = NULL,
  u = NULL,
  incidenceAngle = NULL,
  sourceBands = NULL,
  copyMasks = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{solarIrradiance}{The solar irradiance.}

\item{u}{U}

\item{incidenceAngle}{The incidence angle in degrees.}

\item{sourceBands}{The source bands for the computation.}

\item{copyMasks}{Copy masks from the source product Default value is
'false'.}
}
\value{
snap_op_reflectance_to_radiance_op object
}
\description{
ReflectanceToRadianceOp: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"The 'Reflectance To Radiance Processor' operator retrieves the radiance
from reflectance using Sentinel-2 products"
}
