% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_remodulate.R
\name{op_remodulate}
\alias{op_remodulate}
\title{Remodulate: snap operator function}
\usage{
op_remodulate(operator_id, sourceProduct)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}
}
\value{
snap_op_remodulate object
}
\description{
Remodulate: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Remodulation and reramping of SLC data Source Options:
-SsourceProduct=\if{html}{\out{<file>}} Sets source 'sourceProduct' to \if{html}{\out{<filepath>}}. This is a
mandatory source."
}
