% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_remove_antenna_pattern.R
\name{op_remove_antenna_pattern}
\alias{op_remove_antenna_pattern}
\title{RemoveAntennaPattern: snap operator function}
\usage{
op_remove_antenna_pattern(operator_id, sourceProduct, sourceBands = NULL)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}
}
\value{
snap_op_remove_antenna_pattern object
}
\description{
RemoveAntennaPattern: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Remove Antenna Pattern"
}
