% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_remove_grd_border_noise.R
\name{op_remove_grd_border_noise}
\alias{op_remove_grd_border_noise}
\title{Remove-GRD-Border-Noise: snap operator function}
\usage{
op_remove_grd_border_noise(
  operator_id,
  sourceProduct,
  selectedPolarisations = NULL,
  borderLimit = 500,
  trimThreshold = 0.5
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{selectedPolarisations}{The list of polarisations}

\item{borderLimit}{The border margin limit Default value is '500'.}

\item{trimThreshold}{The trim threshold Default value is '0.5'.}
}
\value{
snap_op_remove_grd_border_noise object
}
\description{
Remove-GRD-Border-Noise: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Mask no-value pixels for GRD product"
}
