% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_replace_metadata.R
\name{op_replace_metadata}
\alias{op_replace_metadata}
\title{ReplaceMetadata: snap operator function}
\usage{
op_replace_metadata(
  operator_id,
  sourceProduct,
  note = "Replace the metadata of the first product with that of the second"
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{note}{Sets parameter 'note' to \if{html}{\out{<string>}}. Default value is 'Replace the
metadata of the first product with that of the second'.}
}
\value{
snap_op_replace_metadata object
}
\description{
ReplaceMetadata: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Replace the metadata of the first product with that of the second Parameter
Options: -Pnote=\if{html}{\out{<string>}} Sets parameter 'note' to \if{html}{\out{<string>}}. Default value is
'Replace the metadata of the first product with that of the second'."
}
