% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_reproject.R
\name{op_reproject}
\alias{op_reproject}
\title{Reproject: snap operator function}
\usage{
op_reproject(
  operator_id,
  sourceProduct,
  collocateWith = NULL,
  wktFile = NULL,
  crs = NULL,
  referencePixelX = NULL,
  referencePixelY = NULL,
  easting = NULL,
  northing = NULL,
  orientation = 0,
  pixelSizeX = NULL,
  pixelSizeY = NULL,
  width = NULL,
  height = NULL,
  tileSizeX = NULL,
  tileSizeY = NULL,
  orthorectify = FALSE,
  elevationModelName = NULL,
  noDataValue = NULL,
  includeTiePointGrids = TRUE,
  addDeltaBands = FALSE,
  resampling = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{collocateWith}{The source product will be collocated with this
product. This is an optional source.}

\item{wktFile}{A file which contains the target Coordinate Reference System
in WKT format.}

\item{crs}{A text specifying the target Coordinate Reference System, either
in WKT or as an authority code. For appropriate EPSG authority codes see
(www.epsg-registry.org). AUTO authority can be used with code 42001 (UTM), and
42002 (Transverse Mercator) where the scene center is used as reference.
Examples: EPSG:4326, AUTO:42001}

\item{referencePixelX}{The X-position of the reference pixel.}

\item{referencePixelY}{The Y-position of the reference pixel.}

\item{easting}{The easting of the reference pixel.}

\item{northing}{The northing of the reference pixel.}

\item{orientation}{The orientation of the output product (in degree). Valid
interval is [-360,360]. Default value is '0'.}

\item{pixelSizeX}{The pixel size in X direction given in CRS units.}

\item{pixelSizeY}{The pixel size in Y direction given in CRS units.}

\item{width}{The width of the target product.}

\item{height}{The height of the target product.}

\item{tileSizeX}{The tile size in X direction.}

\item{tileSizeY}{The tile size in Y direction.}

\item{orthorectify}{Whether the source product should be orthorectified.
(Not applicable to all products) Default value is 'false'.}

\item{elevationModelName}{The name of the elevation model for the
orthorectification. If not given tie-point data is used.}

\item{noDataValue}{The value used to indicate no-data.}

\item{includeTiePointGrids}{Whether tie-point grids should be included in
the output product. Default value is 'true'.}

\item{addDeltaBands}{Whether to add delta longitude and latitude bands.
Default value is 'false'.}

\item{resampling}{The method used for resampling of floating-point raster
data. Value must be one of 'Nearest', 'Bilinear', 'Bicubic'. Default value is
'Nearest'.}
}
\value{
snap_op_reproject object
}
\description{
Reproject: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Reprojection of a source product to a target Coordinate Reference System."
}
