% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_resample.R
\name{op_resample}
\alias{op_resample}
\title{Resample: snap operator function}
\usage{
op_resample(
  operator_id,
  sourceProduct,
  targetWidth = NULL,
  targetHeight = NULL,
  targetResolution = NULL,
  resamplingPreset = NULL,
  bandResamplings = NULL,
  resampleOnPyramidLevels = TRUE,
  downsampling = NULL,
  flagDownsampling = NULL,
  referenceBand = NULL,
  upsampling = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{targetWidth}{The width that all bands of the target product shall
have. If this is set, targetHeight must be set, too. Either this and
targetHeight or referenceBand or targetResolution must be set.}

\item{targetHeight}{The height that all bands of the target product shall
have. If this is set, targetWidth must be set, too. Either this and targetWidth
or referenceBand or targetResolution must be set.}

\item{targetResolution}{The resolution that all bands of the target product
shall have. The same value will be applied to scale image widths and heights.
Either this or referenceBand or targetwidth and targetHeight must be set.}

\item{resamplingPreset}{The resampling preset. This will over rules the
settings for upsampling, downsampling and flagDownsampling.}

\item{bandResamplings}{The band resamplings. This will over rules the
settings for resamplingPreset.}

\item{resampleOnPyramidLevels}{This setting will increase performance when
viewing the image, but accurate resamplings are only retrieved when zooming in
on a pixel. Default value is 'true'.}

\item{downsampling}{The method used for aggregation (downsampling to a
coarser resolution). Value must be one of 'First', 'Min', 'Max', 'Mean',
'Median'. Default value is 'First'.}

\item{flagDownsampling}{The method used for aggregation (downsampling to a
coarser resolution) of flags. Value must be one of 'First', 'FlagAnd',
'FlagOr', 'FlagMedianAnd', 'FlagMedianOr'. Default value is 'First'.}

\item{referenceBand}{The name of the reference band. All other bands will be
re-sampled to match its size and resolution. Either this or targetResolutionor
targetWidth and targetHeight must be set.}

\item{upsampling}{The method used for interpolation (upsampling to a finer
resolution). Value must be one of 'Nearest', 'Bilinear', 'Bicubic'. Default
value is 'Nearest'.}
}
\value{
snap_op_resample object
}
\description{
Resample: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Resampling of a multi-size source product to a single-size target product."
}
