% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_s2rep_op.R
\name{op_s2rep_op}
\alias{op_s2rep_op}
\title{S2repOp: snap operator function}
\usage{
op_s2rep_op(
  operator_id,
  sourceProduct,
  resampleType = "None",
  redB4Factor = "1.0F",
  redB5Factor = "1.0F",
  redB6Factor = "1.0F",
  nirFactor = "1.0F",
  redSourceBand4 = NULL,
  redSourceBand5 = NULL,
  redSourceBand6 = NULL,
  nirSourceBand = NULL,
  downsampling = NULL,
  upsampling = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{resampleType}{If selected band s differ in size, the resample method
used before computing the index Value must be one of 'None', 'Lowest
resolution', 'Highest resolution'. Default value is 'None'.}

\item{redB4Factor}{The value of the red source band (B4) is multiplied by
this value. Default value is '1.0F'.}

\item{redB5Factor}{The value of the red source band (B5) is multiplied by
this value. Default value is '1.0F'.}

\item{redB6Factor}{The value of the red source band (B6) is multiplied by
this value. Default value is '1.0F'.}

\item{nirFactor}{The value of the NIR source band is multiplied by this
value. Default value is '1.0F'.}

\item{redSourceBand4}{The red band (B4) for the S2REP computation. If not
provided, the operator will try to find the best fitting band.}

\item{redSourceBand5}{The red band (B5) for the S2REP computation. If not
provided, the operator will try to find the best fitting band.}

\item{redSourceBand6}{The red band (B6) for the S2REP computation. If not
provided, the operator will try to find the best fitting band.}

\item{nirSourceBand}{The near-infrared band for the S2REP computation. If
not provided, the operator will try to find the best fitting band.}

\item{downsampling}{The method used for aggregation (downsampling to a
coarser resolution). Value must be one of 'First', 'Min', 'Max', 'Mean',
'Median'. Default value is 'First'.}

\item{upsampling}{The method used for interpolation (upsampling to a finer
resolution). Value must be one of 'Nearest', 'Bilinear', 'Bicubic'. Default
value is 'Nearest'.}
}
\value{
snap_op_s2rep_op object
}
\description{
S2repOp: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Sentinel-2 red-edge position index"
}
