% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_sar_simulation.R
\name{op_sar_simulation}
\alias{op_sar_simulation}
\title{SAR-Simulation: snap operator function}
\usage{
op_sar_simulation(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  demName = "SRTM 3Sec",
  demResamplingMethod = "BICUBIC_INTERPOLATION",
  externalDEMFile = NULL,
  externalDEMNoDataValue = 0,
  externalDEMApplyEGM = TRUE,
  saveLayoverShadowMask = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{demName}{The digital elevation model. Default value is 'SRTM 3Sec'.}

\item{demResamplingMethod}{Sets parameter 'demResamplingMethod' to \if{html}{\out{<string>}}.
Default value is 'BICUBIC_INTERPOLATION'.}

\item{externalDEMFile}{Sets parameter 'externalDEMFile' to \if{html}{\out{<file>}}.}

\item{externalDEMNoDataValue}{Sets parameter 'externalDEMNoDataValue' to
\if{html}{\out{<double>}}. Default value is '0'.}

\item{externalDEMApplyEGM}{Sets parameter 'externalDEMApplyEGM' to
\if{html}{\out{<boolean>}}. Default value is 'true'.}

\item{saveLayoverShadowMask}{Sets parameter 'saveLayoverShadowMask' to
\if{html}{\out{<boolean>}}. Default value is 'false'.}
}
\value{
snap_op_sar_simulation object
}
\description{
SAR-Simulation: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Rigorous SAR Simulation"
}
