% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_sarsim_terrain_correction.R
\name{op_sarsim_terrain_correction}
\alias{op_sarsim_terrain_correction}
\title{SARSim-Terrain-Correction: snap operator function}
\usage{
op_sarsim_terrain_correction(
  operator_id,
  sourceProduct,
  rmsThreshold = 1,
  warpPolynomialOrder = 1,
  imgResamplingMethod = "BILINEAR_INTERPOLATION",
  pixelSpacingInMeter = 0,
  pixelSpacingInDegree = 0,
  mapProjection = NULL,
  alignToStandardGrid = FALSE,
  standardGridOriginX = 0,
  standardGridOriginY = 0,
  saveDEM = FALSE,
  saveLatLon = FALSE,
  saveLocalIncidenceAngle = FALSE,
  saveProjectedLocalIncidenceAngle = FALSE,
  saveSelectedSourceBand = TRUE,
  outputComplex = FALSE,
  applyRadiometricNormalization = FALSE,
  saveSigmaNought = FALSE,
  saveGammaNought = FALSE,
  saveBetaNought = FALSE,
  incidenceAngleForSigma0 = "Use projected local incidence angle from DEM",
  incidenceAngleForGamma0 = "Use projected local incidence angle from DEM",
  auxFile = "Latest Auxiliary File",
  externalAuxFile = NULL,
  openShiftsFile = FALSE,
  openResidualsFile = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{rmsThreshold}{The RMS threshold for eliminating invalid GCPs Valid
interval is (0, *). Default value is '1.0'.}

\item{warpPolynomialOrder}{The order of WARP polynomial function Value must
be one of '1', '2', '3'. Default value is '1'.}

\item{imgResamplingMethod}{Sets parameter 'imgResamplingMethod' to \if{html}{\out{<string>}}.
Value must be one of 'NEAREST_NEIGHBOUR', 'BILINEAR_INTERPOLATION',
'CUBIC_CONVOLUTION'. Default value is 'BILINEAR_INTERPOLATION'.}

\item{pixelSpacingInMeter}{The pixel spacing in meters Default value is '0'.}

\item{pixelSpacingInDegree}{The pixel spacing in degrees Default value is
'0'.}

\item{mapProjection}{The coordinate reference system in well known text
format}

\item{alignToStandardGrid}{Force the image grid to be aligned with a
specific point Default value is 'false'.}

\item{standardGridOriginX}{x-coordinate of the standard grid's origin point
Default value is '0'.}

\item{standardGridOriginY}{y-coordinate of the standard grid's origin point
Default value is '0'.}

\item{saveDEM}{Sets parameter 'saveDEM' to \if{html}{\out{<boolean>}}. Default value is
'false'.}

\item{saveLatLon}{Sets parameter 'saveLatLon' to \if{html}{\out{<boolean>}}. Default value is
'false'.}

\item{saveLocalIncidenceAngle}{Sets parameter 'saveLocalIncidenceAngle' to
\if{html}{\out{<boolean>}}. Default value is 'false'.}

\item{saveProjectedLocalIncidenceAngle}{Sets parameter
'saveProjectedLocalIncidenceAngle' to \if{html}{\out{<boolean>}}. Default value is 'false'.}

\item{saveSelectedSourceBand}{Sets parameter 'saveSelectedSourceBand' to
\if{html}{\out{<boolean>}}. Default value is 'true'.}

\item{outputComplex}{Sets parameter 'outputComplex' to \if{html}{\out{<boolean>}}. Default
value is 'false'.}

\item{applyRadiometricNormalization}{Sets parameter
'applyRadiometricNormalization' to \if{html}{\out{<boolean>}}. Default value is 'false'.}

\item{saveSigmaNought}{Sets parameter 'saveSigmaNought' to \if{html}{\out{<boolean>}}.
Default value is 'false'.}

\item{saveGammaNought}{Sets parameter 'saveGammaNought' to \if{html}{\out{<boolean>}}.
Default value is 'false'.}

\item{saveBetaNought}{Sets parameter 'saveBetaNought' to \if{html}{\out{<boolean>}}. Default
value is 'false'.}

\item{incidenceAngleForSigma0}{Sets parameter 'incidenceAngleForSigma0' to
\if{html}{\out{<string>}}. Value must be one of 'Use incidence angle from Ellipsoid', 'Use
projected local incidence angle from DEM', 'Use local incidence angle from
DEM'. Default value is 'Use projected local incidence angle from DEM'.}

\item{incidenceAngleForGamma0}{Sets parameter 'incidenceAngleForGamma0' to
\if{html}{\out{<string>}}. Value must be one of 'Use incidence angle from Ellipsoid', 'Use
projected local incidence angle from DEM', 'Use local incidence angle from
DEM'. Default value is 'Use projected local incidence angle from DEM'.}

\item{auxFile}{The auxiliary file Value must be one of 'Latest Auxiliary
File', 'Product Auxiliary File', 'External Auxiliary File'. Default value is
'Latest Auxiliary File'.}

\item{externalAuxFile}{The antenne elevation pattern gain auxiliary data
file.}

\item{openShiftsFile}{Show range and azimuth shifts file in a text viewer
Default value is 'false'.}

\item{openResidualsFile}{Show the Residuals file in a text viewer Default
value is 'false'.}
}
\value{
snap_op_sarsim_terrain_correction object
}
\description{
SARSim-Terrain-Correction: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Orthorectification with SAR simulation"
}
