% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_set_no_data_value.R
\name{op_set_no_data_value}
\alias{op_set_no_data_value}
\title{SetNoDataValue: snap operator function}
\usage{
op_set_no_data_value(
  operator_id,
  sourceProduct,
  noDataValueUsed = TRUE,
  noDataValue = 0
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{noDataValueUsed}{Sets parameter 'noDataValueUsed' to \if{html}{\out{<boolean>}}.
Default value is 'true'.}

\item{noDataValue}{Sets parameter 'noDataValue' to \if{html}{\out{<double>}}. Default value
is '0.0'.}
}
\value{
snap_op_set_no_data_value object
}
\description{
SetNoDataValue: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Set NoDataValueUsed flag and NoDataValue for all bands"
}
