% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_slice_assembly.R
\name{op_slice_assembly}
\alias{op_slice_assembly}
\title{SliceAssembly: snap operator function}
\usage{
op_slice_assembly(operator_id, sourceProduct, selectedPolarisations = NULL)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{selectedPolarisations}{The list of polarisations}
}
\value{
snap_op_slice_assembly object
}
\description{
SliceAssembly: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Merges Sentinel-1 slice products Parameter Options:
-PselectedPolarisations=<string,string,string,...> The list of polarisations"
}
