% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_sm_dielectric_modeling.R
\name{op_sm_dielectric_modeling}
\alias{op_sm_dielectric_modeling}
\title{SM-Dielectric-Modeling: snap operator function}
\usage{
op_sm_dielectric_modeling(
  operator_id,
  sourceProduct,
  modelToUse = "Hallikainen",
  minSM = 0,
  maxSM = 0.55,
  outputRDC = TRUE,
  outputLandCover = TRUE,
  effectiveSoilTemperature = 18
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{modelToUse}{Choice of dielectric models for SM inversion Value must be
one of 'Hallikainen', 'Mironov'. Default value is 'Hallikainen'.}

\item{minSM}{Minimum soil moisture value Valid interval is [0.0, 1.0).
Default value is '0.0'. Parameter unit is 'm^3/m^3'.}

\item{maxSM}{Maximum soil moisture value Valid interval is (0.0, 1.0].
Default value is '0.55'. Parameter unit is 'm^3/m^3'.}

\item{outputRDC}{Optional RDC in output Default value is 'true'.}

\item{outputLandCover}{Optional LandCover in output Default value is 'true'.}

\item{effectiveSoilTemperature}{Effective soil temperature Valid interval is
[-50.0, 50.0]. Default value is '18.0'. Parameter unit is 'Celsius'.}
}
\value{
snap_op_sm_dielectric_modeling object
}
\description{
SM-Dielectric-Modeling: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Performs SM inversion using dielectric model"
}
