% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_smac_op.R
\name{op_smac_op}
\alias{op_smac_op}
\title{SmacOp: snap operator function}
\usage{
op_smac_op(
  operator_id,
  sourceProduct,
  tauAero550 = 0.2,
  uH2o = 3,
  uO3 = 0.15,
  surfPress = 1013,
  useMerisADS = FALSE,
  aerosolType = "CONTINENTAL",
  invalidPixel = 0,
  maskExpression = NULL,
  maskExpressionForward = NULL,
  bandNames = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{tauAero550}{Aerosol optical depth Default value is '0.2'.}

\item{uH2o}{Relative humidity Default value is '3.0'. Parameter unit is
'g/cm²'.}

\item{uO3}{Ozone content Default value is '0.15'. Parameter unit is 'cm *
atm'.}

\item{surfPress}{Surface pressure Default value is '1013.0'. Parameter unit
is 'hPa'.}

\item{useMerisADS}{Use ECMWF data in the MERIS ADS Default value is 'false'.}

\item{aerosolType}{Aerosol type Default value is 'CONTINENTAL'. This is a
mandatory parameter.}

\item{invalidPixel}{Default reflectance for invalid pixel Default value is
'0.0'.}

\item{maskExpression}{Mask expression for the whole view (MERIS) or the
nadir view (AATSR)}

\item{maskExpressionForward}{Mask expression for the forward view (AATSR
only)}

\item{bandNames}{Bands to process This is a mandatory parameter.}
}
\value{
snap_op_smac_op object
}
\description{
SmacOp: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Applies the Simplified Method for Atmospheric Corrections of Envisat
MERIS/(A)ATSR measurements."
}
