% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_snaphu_import.R
\name{op_snaphu_import}
\alias{op_snaphu_import}
\title{SnaphuImport: snap operator function}
\usage{
op_snaphu_import(operator_id, sourceProduct, doNotKeepWrapped = FALSE)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{doNotKeepWrapped}{Sets parameter 'doNotKeepWrapped' to \if{html}{\out{<boolean>}}.
Default value is 'false'.}
}
\value{
snap_op_snaphu_import object
}
\description{
SnaphuImport: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Ingest SNAPHU results into InSAR product. Parameter Options:
-PdoNotKeepWrapped=\if{html}{\out{<boolean>}} Sets parameter 'doNotKeepWrapped' to \if{html}{\out{<boolean>}}.
Default value is 'false'."
}
