% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_speckle_divergence.R
\name{op_speckle_divergence}
\alias{op_speckle_divergence}
\title{Speckle-Divergence: snap operator function}
\usage{
op_speckle_divergence(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  windowSizeStr = "15x15"
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{windowSizeStr}{Sets parameter 'windowSizeStr' to \if{html}{\out{<string>}}. Value must
be one of '3x3', '5x5', '7x7', '9x9', '11x11', '13x13', '15x15', '17x17'.
Default value is '15x15'.}
}
\value{
snap_op_speckle_divergence object
}
\description{
Speckle-Divergence: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Detect urban area."
}
