% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_srgr.R
\name{op_srgr}
\alias{op_srgr}
\title{SRGR: snap operator function}
\usage{
op_srgr(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  warpPolynomialOrder = 4,
  interpolationMethod = "Linear interpolation"
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{warpPolynomialOrder}{The order of WARP polynomial function Value must
be one of '1', '2', '3', '4'. Default value is '4'.}

\item{interpolationMethod}{Sets parameter 'interpolationMethod' to \if{html}{\out{<string>}}.
Value must be one of 'Nearest-neighbor interpolation', 'Linear interpolation',
'Cubic interpolation', 'Cubic2 interpolation', 'Sinc interpolation'. Default
value is 'Linear interpolation'.}
}
\value{
snap_op_srgr object
}
\description{
SRGR: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Converts Slant Range to Ground Range"
}
