% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_stack_averaging.R
\name{op_stack_averaging}
\alias{op_stack_averaging}
\title{Stack-Averaging: snap operator function}
\usage{
op_stack_averaging(operator_id, sourceProduct, statistic = "Mean Average")
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{statistic}{Sets parameter 'statistic' to \if{html}{\out{<string>}}. Value must be one
of 'Mean Average', 'Minimum', 'Maximum', 'Standard Deviation', 'Coefficient of
Variation'. Default value is 'Mean Average'.}
}
\value{
snap_op_stack_averaging object
}
\description{
Stack-Averaging: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Averaging multi-temporal images"
}
