% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_stack_split.R
\name{op_stack_split}
\alias{op_stack_split}
\title{Stack-Split: snap operator function}
\usage{
op_stack_split(
  operator_id,
  sourceProduct,
  targetFolder = "target",
  formatName = "BEAM-DIMAP"
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{targetFolder}{The output folder to which the data product is written.
Default value is 'target'.}

\item{formatName}{The name of the output file format. Default value is
'BEAM-DIMAP'.}
}
\value{
snap_op_stack_split object
}
\description{
Stack-Split: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Writes all bands to files."
}
