% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_stamps_export.R
\name{op_stamps_export}
\alias{op_stamps_export}
\title{StampsExport: snap operator function}
\usage{
op_stamps_export(
  operator_id,
  sourceProduct,
  targetFolder = NULL,
  psiFormat = TRUE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{targetFolder}{The output folder to which the data product is written.}

\item{psiFormat}{Format for PSI or SBAS Default value is 'true'.}
}
\value{
snap_op_stamps_export object
}
\description{
StampsExport: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Export data for StaMPS processing Parameter Options: -PpsiFormat=\if{html}{\out{<boolean>}}
Format for PSI or SBAS Default value is 'true'. -PtargetFolder=\if{html}{\out{<file>}} The
output folder to which the data product is written."
}
