% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_sub_graph.R
\name{op_sub_graph}
\alias{op_sub_graph}
\title{SubGraph: snap operator function}
\usage{
op_sub_graph(operator_id, graphFile = NULL)
}
\arguments{
\item{operator_id}{character operator id}

\item{graphFile}{Sets parameter 'graphFile' to \if{html}{\out{<file>}}.}
}
\value{
snap_op_sub_graph object
}
\description{
SubGraph: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Encapsulates a graph within a graph. Parameter Options: -PgraphFile=\if{html}{\out{<file>}}
Sets parameter 'graphFile' to \if{html}{\out{<file>}}."
}
