% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_subset.R
\name{op_subset}
\alias{op_subset}
\title{Subset: snap operator function}
\usage{
op_subset(
  operator_id,
  sourceProduct,
  region = NULL,
  referenceBand = NULL,
  geoRegion = NULL,
  subSamplingX = 1,
  subSamplingY = 1,
  fullSwath = FALSE,
  copyMetadata = FALSE,
  sourceBands = NULL,
  tiePointGrids = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{region}{The subset region in pixel coordinates. Use the following
format: \{x\},\{y\},\{width\},\{height\} If not given, the entire scene is used. The
'geoRegion' parameter has precedence over this parameter.}

\item{referenceBand}{The band used to indicate the pixel coordinates.}

\item{geoRegion}{The subset region in geographical coordinates using
WKT-format, e.g. POLYGON((\{lon1\} \{lat1\}, \{lon2\} \{lat2\}, ..., \{lon1\} \{lat1\}))
(make sure to quote the option due to spaces in \{geometry\}). If not given, the
entire scene is used.}

\item{subSamplingX}{The pixel sub-sampling step in X (horizontal image
direction) Default value is '1'.}

\item{subSamplingY}{The pixel sub-sampling step in Y (vertical image
direction) Default value is '1'.}

\item{fullSwath}{Forces the operator to extend the subset region to the full
swath. Default value is 'false'.}

\item{copyMetadata}{Whether to copy the metadata of the source product.
Default value is 'false'.}

\item{sourceBands}{The list of source bands.}

\item{tiePointGrids}{The list of tie-point grid names.}
}
\value{
snap_op_subset object
}
\description{
Subset: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Create a spatial and/or spectral subset of a data product."
}
