% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_supervised_wishart_classification.R
\name{op_supervised_wishart_classification}
\alias{op_supervised_wishart_classification}
\title{Supervised-Wishart-Classification: snap operator function}
\usage{
op_supervised_wishart_classification(
  operator_id,
  sourceProduct,
  trainingDataSet = NULL,
  windowSize = 5
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{trainingDataSet}{The training data set file}

\item{windowSize}{The sliding window size Valid interval is (1, 100].
Default value is '5'.}
}
\value{
snap_op_supervised_wishart_classification object
}
\description{
Supervised-Wishart-Classification: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Perform supervised Wishart classification"
}
