% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_temporal_percentile.R
\name{op_temporal_percentile}
\alias{op_temporal_percentile}
\title{TemporalPercentile: snap operator function}
\usage{
op_temporal_percentile(
  operator_id,
  sourceProduct,
  sourceProductPaths = NULL,
  startDate = NULL,
  endDate = NULL,
  keepIntermediateTimeSeriesProduct = TRUE,
  timeSeriesOutputDir = NULL,
  crs = "EPSG:4326",
  westBound = -15,
  northBound = 75,
  eastBound = 30,
  southBound = 35,
  pixelSizeX = 0.05,
  pixelSizeY = 0.05,
  sourceBandName = NULL,
  bandMathsExpression = NULL,
  percentileBandNamePrefix = NULL,
  validPixelExpression = "true",
  percentiles = 90,
  gapFillingMethod = "gapFillingLinearInterpolation",
  startValueFallback = 0,
  endValueFallback = 0,
  resampling = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceProductPaths}{A comma-separated list of file paths specifying
the source products. Source products to be considered for percentile
computation. Each path may contain the wildcards '\strong{' (matches recursively any
directory), '*' (matches any character sequence in path names) and '?' (matches
any single character). If, for example, all NetCDF files under /eodata/ shall
be considered, use '/eodata/}/*.nc'.}

\item{startDate}{The start date. If not given, it is taken from the 'oldest'
source product. Products that have a start date earlier than the start date
given by this parameter are not considered. Format for valid values is
'yyyy-MM-dd HH:mm:ss'.}

\item{endDate}{The end date. If not given, it is taken from the 'newest'
source product. Products that have an end date later than the end date given by
this parameter are not considered. Format for valid values is 'yyyy-MM-dd
HH:mm:ss'.}

\item{keepIntermediateTimeSeriesProduct}{Determines whether the time series
product which is created during computation should be written to disk. Default
value is 'true'.}

\item{timeSeriesOutputDir}{The output directory for the intermediate time
series product. If not given, the time series product will be written to the
working directory.}

\item{crs}{A text specifying the target Coordinate Reference System, either
in WKT or as an authority code. For appropriate EPSG authority codes see
(www.epsg-registry.org). AUTO authority can be used with code 42001 (UTM), and
42002 (Transverse Mercator) where the scene center is used as reference.
Examples: EPSG:4326, AUTO:42001 Default value is 'EPSG:4326'.}

\item{westBound}{The most-western longitude. All values west of this
longitude will not be considered. Valid interval is [-180,180]. Default value
is '-15.0'.}

\item{northBound}{The most-northern latitude. All values north of this
latitude will not be considered. Valid interval is [-90,90]. Default value is
'75.0'.}

\item{eastBound}{The most-eastern longitude. All values east of this
longitude will not be considered. Valid interval is [-180,180]. Default value
is '30.0'.}

\item{southBound}{The most-southern latitude. All values south of this
latitude will not be considered. Valid interval is [-90,90]. Default value is
'35.0'.}

\item{pixelSizeX}{Size of a pixel in X-direction in map units. Default value
is '0.05'.}

\item{pixelSizeY}{Size of a pixel in Y-direction in map units. Default value
is '0.05'.}

\item{sourceBandName}{The name of the band in the source products. Either
this or 'bandMathsExpression' must be provided.}

\item{bandMathsExpression}{A band maths expression serving as input band.
Either this or 'sourceBandName' must be provided.}

\item{percentileBandNamePrefix}{If given, this is the percentile band name
prefix. If empty, the resulting percentile band’s name prefix will be either
the 'sourceBandName' or created from the 'bandMathsExpression'.}

\item{validPixelExpression}{The valid pixel expression serving as criterion
for whether to consider pixels for computation. Default value is 'true'.}

\item{percentiles}{The percentiles. Default value is '90'.}

\item{gapFillingMethod}{The gap filling method for percentile calculation.
Value must be one of 'noGapFilling', 'gapFillingLinearInterpolation',
'gapFillingSplineInterpolation', 'gapFillingQuadraticInterpolation'. Default
value is 'gapFillingLinearInterpolation'.}

\item{startValueFallback}{The fallback value for the start of a pixel time
series. It will be considered if there is no valid value at the pixel of the
oldest collocated mean band. This would be the case, if, e.g., there is a
cloudy day at the time period start. Default value is '0.0'.}

\item{endValueFallback}{The fallback value for the end of a pixel time
series. It will be considered ifthere is no valid value at the pixel of the
newest collocated mean band. This would be the case, if, e.g., there is a
cloudy day at the time period end. Default value is '0.0'.}

\item{resampling}{The method used for resampling of floating-point raster
data, if source products must be reprojected to the target CRS. Value must be
one of 'Nearest', 'Bilinear', 'Bicubic'. Default value is 'Nearest'.}
}
\value{
snap_op_temporal_percentile object
}
\description{
TemporalPercentile: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Computes percentiles over a given time period. Parameter Options:
-PbandMathsExpression=\if{html}{\out{<string>}} A band maths expression serving as input band.
Either this or 'sourceBandName' must be provided. -Pcrs=\if{html}{\out{<string>}} A text
specifying the target Coordinate Reference System, either in WKT or as an
authority code. For appropriate EPSG authority codes see
(www.epsg-registry.org). AUTO authority can be used with code 42001 (UTM), and
42002 (Transverse Mercator) where the scene center is used as reference.
Examples: EPSG:4326, AUTO:42001 Default value is 'EPSG:4326'.
-PeastBound=\if{html}{\out{<double>}} The most-eastern longitude. All values east of this
longitude will not be considered. Valid interval is [-180,180]. Default value
is '30.0'. -PendDate=\if{html}{\out{<uTC>}} The end date. If not given, it is taken from the
'newest' source product. Products that have an end date later than the end date
given by this parameter are not considered. Format for valid values is
'yyyy-MM-dd HH:mm:ss'. -PendValueFallback=\if{html}{\out{<double>}} The fallback value for the
end of a pixel time series. It will be considered ifthere is no valid value at
the pixel of the newest collocated mean band. This would be the case, if, e.g.,
there is a cloudy day at the time period end. Default value is '0.0'.
-PgapFillingMethod=\if{html}{\out{<string>}} The gap filling method for percentile calculation.
Value must be one of 'noGapFilling', 'gapFillingLinearInterpolation',
'gapFillingSplineInterpolation', 'gapFillingQuadraticInterpolation'. Default
value is 'gapFillingLinearInterpolation'.
-PkeepIntermediateTimeSeriesProduct=\if{html}{\out{<boolean>}} Determines whether the time
series product which is created during computation should be written to disk.
Default value is 'true'. -PnorthBound=\if{html}{\out{<double>}} The most-northern latitude. All
values north of this latitude will not be considered. Valid interval is
[-90,90]. Default value is '75.0'. -PpercentileBandNamePrefix=\if{html}{\out{<string>}} If
given, this is the percentile band name prefix. If empty, the resulting
percentile band’s name prefix will be either the 'sourceBandName' or created
from the 'bandMathsExpression'. -Ppercentiles=<int,int,int,...> The
percentiles. Default value is '90'. -PpixelSizeX=\if{html}{\out{<double>}} Size of a pixel in
X-direction in map units. Default value is '0.05'. -PpixelSizeY=\if{html}{\out{<double>}} Size
of a pixel in Y-direction in map units. Default value is '0.05'.
-Presampling=\if{html}{\out{<string>}} The method used for resampling of floating-point raster
data, if source products must be reprojected to the target CRS. Value must be
one of 'Nearest', 'Bilinear', 'Bicubic'. Default value is 'Nearest'.
-PsourceBandName=\if{html}{\out{<string>}} The name of the band in the source products. Either
this or 'bandMathsExpression' must be provided.
-PsourceProductPaths=<string,string,string,...> A comma-separated list of file
paths specifying the source products. Source products to be considered for
percentile computation. Each path may contain the wildcards '\strong{' (matches
recursively any directory), '*' (matches any character sequence in path names)
and '?' (matches any single character). If, for example, all NetCDF files under
/eodata/ shall be considered, use '/eodata/}/*.nc'. -PsouthBound=\if{html}{\out{<double>}} The
most-southern latitude. All values south of this latitude will not be
considered. Valid interval is [-90,90]. Default value is '35.0'.
-PstartDate=\if{html}{\out{<uTC>}} The start date. If not given, it is taken from the 'oldest'
source product. Products that have a start date earlier than the start date
given by this parameter are not considered. Format for valid values is
'yyyy-MM-dd HH:mm:ss'. -PstartValueFallback=\if{html}{\out{<double>}} The fallback value for the
start of a pixel time series. It will be considered if there is no valid value
at the pixel of the oldest collocated mean band. This would be the case, if,
e.g., there is a cloudy day at the time period start. Default value is '0.0'.
-PtimeSeriesOutputDir=\if{html}{\out{<file>}} The output directory for the intermediate time
series product. If not given, the time series product will be written to the
working directory. -PvalidPixelExpression=\if{html}{\out{<string>}} The valid pixel expression
serving as criterion for whether to consider pixels for computation. Default
value is 'true'. -PwestBound=\if{html}{\out{<double>}} The most-western longitude. All values
west of this longitude will not be considered. Valid interval is [-180,180].
Default value is '-15.0'."
}
