% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_terrain_flattening.R
\name{op_terrain_flattening}
\alias{op_terrain_flattening}
\title{Terrain-Flattening: snap operator function}
\usage{
op_terrain_flattening(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  demName = "SRTM 1Sec HGT",
  demResamplingMethod = "BILINEAR_INTERPOLATION",
  externalDEMFile = NULL,
  externalDEMNoDataValue = 0,
  externalDEMApplyEGM = FALSE,
  outputSimulatedImage = FALSE,
  outputSigma0 = FALSE,
  nodataValueAtSea = TRUE,
  additionalOverlap = 0.1,
  oversamplingMultiple = 1
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{demName}{The digital elevation model. Default value is 'SRTM 1Sec
HGT'.}

\item{demResamplingMethod}{Sets parameter 'demResamplingMethod' to \if{html}{\out{<string>}}.
Default value is 'BILINEAR_INTERPOLATION'.}

\item{externalDEMFile}{Sets parameter 'externalDEMFile' to \if{html}{\out{<file>}}.}

\item{externalDEMNoDataValue}{Sets parameter 'externalDEMNoDataValue' to
\if{html}{\out{<double>}}. Default value is '0'.}

\item{externalDEMApplyEGM}{Sets parameter 'externalDEMApplyEGM' to
\if{html}{\out{<boolean>}}. Default value is 'false'.}

\item{outputSimulatedImage}{Sets parameter 'outputSimulatedImage' to
\if{html}{\out{<boolean>}}. Default value is 'false'.}

\item{outputSigma0}{Sets parameter 'outputSigma0' to \if{html}{\out{<boolean>}}. Default
value is 'false'.}

\item{nodataValueAtSea}{Mask the sea with no data value (faster) Default
value is 'true'.}

\item{additionalOverlap}{The additional overlap percentage Valid interval is
[0, 1]. Default value is '0.1'.}

\item{oversamplingMultiple}{The oversampling factor Valid interval is [1,
4]. Default value is '1.0'.}
}
\value{
snap_op_terrain_flattening object
}
\description{
Terrain-Flattening: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Terrain Flattening"
}
