% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_terrain_mask.R
\name{op_terrain_mask}
\alias{op_terrain_mask}
\title{Terrain-Mask: snap operator function}
\usage{
op_terrain_mask(
  operator_id,
  sourceProduct,
  demName = "SRTM 3Sec",
  demResamplingMethod = "NEAREST_NEIGHBOUR",
  externalDEMFile = NULL,
  externalDEMNoDataValue = 0,
  windowSizeStr = "15x15",
  thresholdInMeter = 40
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{demName}{The digital elevation model. Value must be one of 'ACE',
'ASTER 1sec GDEM', 'GETASSE30', 'SRTM 1Sec HGT', 'SRTM 3Sec'. Default value is
'SRTM 3Sec'.}

\item{demResamplingMethod}{Sets parameter 'demResamplingMethod' to \if{html}{\out{<string>}}.
Value must be one of 'NEAREST_NEIGHBOUR', 'BILINEAR_INTERPOLATION',
'CUBIC_CONVOLUTION', 'BICUBIC_INTERPOLATION', 'BISINC_5_POINT_INTERPOLATION'.
Default value is 'NEAREST_NEIGHBOUR'.}

\item{externalDEMFile}{Sets parameter 'externalDEMFile' to \if{html}{\out{<file>}}.}

\item{externalDEMNoDataValue}{Sets parameter 'externalDEMNoDataValue' to
\if{html}{\out{<double>}}. Default value is '0'.}

\item{windowSizeStr}{Sets parameter 'windowSizeStr' to \if{html}{\out{<string>}}. Value must
be one of '5x5', '7x7', '9x9', '11x11', '13x13', '15x15', '17x17'. Default
value is '15x15'.}

\item{thresholdInMeter}{Threshold for detection Valid interval is (0, *).
Default value is '40.0'.}
}
\value{
snap_op_terrain_mask object
}
\description{
Terrain-Mask: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Terrain Mask Generation"
}
