% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_thermal_noise_removal.R
\name{op_thermal_noise_removal}
\alias{op_thermal_noise_removal}
\title{ThermalNoiseRemoval: snap operator function}
\usage{
op_thermal_noise_removal(
  operator_id,
  sourceProduct,
  selectedPolarisations = NULL,
  removeThermalNoise = TRUE,
  outputNoise = FALSE,
  reIntroduceThermalNoise = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{selectedPolarisations}{The list of polarisations}

\item{removeThermalNoise}{Remove thermal noise Default value is 'true'.}

\item{outputNoise}{Output noise Default value is 'false'.}

\item{reIntroduceThermalNoise}{Re-introduce thermal noise Default value is
'false'.}
}
\value{
snap_op_thermal_noise_removal object
}
\description{
ThermalNoiseRemoval: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Removes thermal noise from products"
}
