% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_three_pass_din_sar.R
\name{op_three_pass_din_sar}
\alias{op_three_pass_din_sar}
\title{Three-passDInSAR: snap operator function}
\usage{
op_three_pass_din_sar(operator_id, sourceProduct, orbitDegree = 3)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{orbitDegree}{Degree of orbit interpolation polynomial Valid interval
is (1, 10]. Default value is '3'.}
}
\value{
snap_op_three_pass_din_sar object
}
\description{
Three-passDInSAR: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Differential Interferometry Parameter Options: -PorbitDegree=\if{html}{\out{<int>}} Degree
of orbit interpolation polynomial Valid interval is (1, 10]. Default value is
'3'."
}
