% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_tile_cache.R
\name{op_tile_cache}
\alias{op_tile_cache}
\title{TileCache: snap operator function}
\usage{
op_tile_cache(operator_id, sourceProduct, cacheSize = 1000)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{cacheSize}{The cache size in MB. Set it to 0 to use default tile
cache. Default value is '1000'. Parameter unit is 'MB'.}
}
\value{
snap_op_tile_cache object
}
\description{
TileCache: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Experimental Operator which provides a dedicated cache for its source
product. A guide on how this operator is used is provided at
https://senbox.atlassian.net/wiki/x/VQCTLw."
}
