% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_tile_writer.R
\name{op_tile_writer}
\alias{op_tile_writer}
\title{TileWriter: snap operator function}
\usage{
op_tile_writer(
  operator_id,
  sourceProduct,
  file = NULL,
  formatName = "BEAM-DIMAP",
  divisionBy = "Tiles",
  numberOfTiles = 4,
  pixelSizeX = 200,
  pixelSizeY = 200,
  overlap = 0
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{file}{The output file to which the data product is written.}

\item{formatName}{The name of the output file format. Default value is
'BEAM-DIMAP'.}

\item{divisionBy}{How to divide the tiles Value must be one of 'Tiles',
'Pixels'. Default value is 'Tiles'.}

\item{numberOfTiles}{The number of output tiles Value must be one of '2',
'4', '9', '16', '36', '64', '100', '256'. Default value is '4'.}

\item{pixelSizeX}{Tile pixel width Default value is '200'.}

\item{pixelSizeY}{Tile pixel height Default value is '200'.}

\item{overlap}{Tile overlap Default value is '0'.}
}
\value{
snap_op_tile_writer object
}
\description{
TileWriter: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Writes a data product to a tiles."
}
