% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_tool_adapter_op.R
\name{op_tool_adapter_op}
\alias{op_tool_adapter_op}
\title{ToolAdapterOp: snap operator function}
\usage{
op_tool_adapter_op(operator_id)
}
\arguments{
\item{operator_id}{character operator id}
}
\value{
snap_op_tool_adapter_op object
}
\description{
ToolAdapterOp: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Tool Adapter Operator"
}
