% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_topo_phase_removal.R
\name{op_topo_phase_removal}
\alias{op_topo_phase_removal}
\title{TopoPhaseRemoval: snap operator function}
\usage{
op_topo_phase_removal(
  operator_id,
  sourceProduct,
  orbitDegree = 3,
  demName = "SRTM 3Sec",
  externalDEMFile = NULL,
  externalDEMNoDataValue = 0,
  tileExtensionPercent = 100,
  outputTopoPhaseBand = FALSE,
  outputElevationBand = FALSE,
  outputLatLonBands = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{orbitDegree}{Degree of orbit interpolation polynomial Valid interval
is (1, 10]. Default value is '3'.}

\item{demName}{The digital elevation model. Default value is 'SRTM 3Sec'.}

\item{externalDEMFile}{Sets parameter 'externalDEMFile' to \if{html}{\out{<file>}}.}

\item{externalDEMNoDataValue}{Sets parameter 'externalDEMNoDataValue' to
\if{html}{\out{<double>}}. Default value is '0'.}

\item{tileExtensionPercent}{Define extension of tile for DEM simulation
(optimization parameter). Default value is '100'.}

\item{outputTopoPhaseBand}{Output topographic phase band. Default value is
'false'.}

\item{outputElevationBand}{Output elevation band. Default value is 'false'.}

\item{outputLatLonBands}{Output lat/lon bands. Default value is 'false'.}
}
\value{
snap_op_topo_phase_removal object
}
\description{
TopoPhaseRemoval: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Compute and subtract TOPO phase"
}
