% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_topsar_deburst.R
\name{op_topsar_deburst}
\alias{op_topsar_deburst}
\title{TOPSAR-Deburst: snap operator function}
\usage{
op_topsar_deburst(operator_id, sourceProduct, selectedPolarisations = NULL)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{selectedPolarisations}{The list of polarisations}
}
\value{
snap_op_topsar_deburst object
}
\description{
TOPSAR-Deburst: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Debursts a Sentinel-1 TOPSAR product"
}
