% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_topsar_deramp_demod.R
\name{op_topsar_deramp_demod}
\alias{op_topsar_deramp_demod}
\title{TOPSAR-DerampDemod: snap operator function}
\usage{
op_topsar_deramp_demod(
  operator_id,
  sourceProduct,
  outputDerampDemodPhase = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{outputDerampDemodPhase}{Sets parameter 'outputDerampDemodPhase' to
\if{html}{\out{<boolean>}}. Default value is 'false'.}
}
\value{
snap_op_topsar_deramp_demod object
}
\description{
TOPSAR-DerampDemod: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Bursts co-registration using orbit and DEM"
}
