% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_topsar_split.R
\name{op_topsar_split}
\alias{op_topsar_split}
\title{TOPSAR-Split: snap operator function}
\usage{
op_topsar_split(
  operator_id,
  sourceProduct,
  subswath = NULL,
  selectedPolarisations = NULL,
  firstBurstIndex = 1,
  lastBurstIndex = 9999,
  wktAoi = NULL
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{subswath}{The list of source bands.}

\item{selectedPolarisations}{The list of polarisations}

\item{firstBurstIndex}{The first burst index Valid interval is [1, *).
Default value is '1'.}

\item{lastBurstIndex}{The last burst index Valid interval is [1, *). Default
value is '9999'.}

\item{wktAoi}{WKT polygon to be used for selecting bursts}
}
\value{
snap_op_topsar_split object
}
\description{
TOPSAR-Split: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Creates a new product with only the selected subswath"
}
