% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_undersample.R
\name{op_undersample}
\alias{op_undersample}
\title{Undersample: snap operator function}
\usage{
op_undersample(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  method = "LowPass Filtering",
  filterSize = "3x3",
  subSamplingX = 2,
  subSamplingY = 2,
  outputImageBy = "Ratio",
  targetImageHeight = 1000,
  targetImageWidth = 1000,
  widthRatio = 0.5,
  heightRatio = 0.5,
  rangeSpacing = 12.5,
  azimuthSpacing = 12.5
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{method}{Sets parameter 'method' to \if{html}{\out{<string>}}. Value must be one of
'Sub-Sampling', 'LowPass Filtering'. Default value is 'LowPass Filtering'.}

\item{filterSize}{Sets parameter 'filterSize' to \if{html}{\out{<string>}}. Value must be one
of '3x3', '5x5', '7x7'. Default value is '3x3'.}

\item{subSamplingX}{Sets parameter 'subSamplingX' to \if{html}{\out{<int>}}. Default value is
'2'.}

\item{subSamplingY}{Sets parameter 'subSamplingY' to \if{html}{\out{<int>}}. Default value is
'2'.}

\item{outputImageBy}{Sets parameter 'outputImageBy' to \if{html}{\out{<string>}}. Value must
be one of 'Image Size', 'Ratio', 'Pixel Spacing'. Default value is 'Ratio'.}

\item{targetImageHeight}{The row dimension of the output image Default value
is '1000'.}

\item{targetImageWidth}{The col dimension of the output image Default value
is '1000'.}

\item{widthRatio}{The width ratio of the output/input images Default value
is '0.5'.}

\item{heightRatio}{The height ratio of the output/input images Default value
is '0.5'.}

\item{rangeSpacing}{The range pixel spacing Default value is '12.5'.}

\item{azimuthSpacing}{The azimuth pixel spacing Default value is '12.5'.}
}
\value{
snap_op_undersample object
}
\description{
Undersample: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Undersample the datset"
}
