% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_update_geo_reference.R
\name{op_update_geo_reference}
\alias{op_update_geo_reference}
\title{Update-Geo-Reference: snap operator function}
\usage{
op_update_geo_reference(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  demName = "SRTM 3Sec",
  demResamplingMethod = "BICUBIC_INTERPOLATION",
  externalDEMFile = NULL,
  externalDEMNoDataValue = 0,
  reGridMethod = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{demName}{The digital elevation model. Value must be one of 'ACE',
'ASTER 1sec GDEM', 'GETASSE30', 'SRTM 1Sec HGT', 'SRTM 3Sec'. Default value is
'SRTM 3Sec'.}

\item{demResamplingMethod}{Sets parameter 'demResamplingMethod' to \if{html}{\out{<string>}}.
Default value is 'BICUBIC_INTERPOLATION'.}

\item{externalDEMFile}{Sets parameter 'externalDEMFile' to \if{html}{\out{<file>}}.}

\item{externalDEMNoDataValue}{Sets parameter 'externalDEMNoDataValue' to
\if{html}{\out{<double>}}. Default value is '0'.}

\item{reGridMethod}{Sets parameter 'reGridMethod' to \if{html}{\out{<boolean>}}. Default
value is 'false'.}
}
\value{
snap_op_update_geo_reference object
}
\description{
Update-Geo-Reference: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Update Geo Reference"
}
