% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_warp.R
\name{op_warp}
\alias{op_warp}
\title{Warp: snap operator function}
\usage{
op_warp(
  operator_id,
  sourceProduct,
  rmsThreshold = 0.05,
  warpPolynomialOrder = 2,
  interpolationMethod = "Cubic convolution (6 points)",
  demRefinement = FALSE,
  demName = "SRTM 3Sec",
  excludeMaster = FALSE,
  openResidualsFile = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{rmsThreshold}{Confidence level for outlier detection procedure, lower
value accepts more outliers Value must be one of '0.001', '0.05', '0.1', '0.5',
'1.0'. Default value is '0.05'.}

\item{warpPolynomialOrder}{The order of WARP polynomial function Value must
be one of '1', '2', '3'. Default value is '2'.}

\item{interpolationMethod}{Sets parameter 'interpolationMethod' to \if{html}{\out{<string>}}.
Value must be one of 'Nearest-neighbor interpolation', 'Bilinear
interpolation', 'Bicubic interpolation', 'Bicubic2 interpolation', 'Linear
interpolation', 'Cubic convolution (4 points)', 'Cubic convolution (6 points)',
'Truncated sinc (6 points)', 'Truncated sinc (8 points)', 'Truncated sinc (16
points)'. Default value is 'Cubic convolution (6 points)'.}

\item{demRefinement}{Refine estimated offsets using a-priori DEM Default
value is 'false'.}

\item{demName}{The digital elevation model. Default value is 'SRTM 3Sec'.}

\item{excludeMaster}{Sets parameter 'excludeMaster' to \if{html}{\out{<boolean>}}. Default
value is 'false'.}

\item{openResidualsFile}{Show the Residuals file in a text viewer Default
value is 'false'.}
}
\value{
snap_op_warp object
}
\description{
Warp: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Create Warp Function And Get Co-registrated Images"
}
