% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_wind_field_estimation.R
\name{op_wind_field_estimation}
\alias{op_wind_field_estimation}
\title{Wind-Field-Estimation: snap operator function}
\usage{
op_wind_field_estimation(
  operator_id,
  sourceProduct,
  sourceBands = NULL,
  windowSizeInKm = 20
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{sourceBands}{The list of source bands.}

\item{windowSizeInKm}{Window size Default value is '20.0'.}
}
\value{
snap_op_wind_field_estimation object
}
\description{
Wind-Field-Estimation: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Estimate wind speed and direction"
}
