% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_write.R
\name{op_write}
\alias{op_write}
\title{Write: snap operator function}
\usage{
op_write(
  operator_id,
  sourceProduct,
  file = NULL,
  formatName = "BEAM-DIMAP",
  deleteOutputOnFailure = TRUE,
  writeEntireTileRows = FALSE,
  clearCacheAfterRowWrite = FALSE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{sourceProduct}{The source product(s) as input to the operator}

\item{file}{The output file to which the data product is written.}

\item{formatName}{The name of the output file format. Default value is
'BEAM-DIMAP'.}

\item{deleteOutputOnFailure}{If true, all output files are deleted after a
failed write operation. Default value is 'true'.}

\item{writeEntireTileRows}{If true, the write operation waits until an
entire tile row is computed. Default value is 'false'.}

\item{clearCacheAfterRowWrite}{If true, the internal tile cache is cleared
after a tile row has been written. Ignored if writeEntireTileRows=false.
Default value is 'false'.}
}
\value{
snap_op_write object
}
\description{
Write: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Writes a data product to a file."
}
