% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-operator-help.R
\name{print.snap_operator_help}
\alias{print.snap_operator_help}
\title{print method for snap operator helper object}
\arguments{
\item{x}{snap_operator_help object}

\item{xml}{logical print the xml graph}
}
\description{
print method for snap operator helper object
}
