% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-classes.R
\name{print.snap_operator_helper}
\alias{print.snap_operator_helper}
\title{print method for snap operator helper object}
\arguments{
\item{x}{snap_operator_helper object}

\item{xml}{logical print the xml graph}
}
\description{
print method for snap operator helper object
}
