% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-snap-op-methods.R, R/snap-graph.R
\name{run_graph}
\alias{run_graph}
\title{print method for snap_graph object}
\usage{
run_graph(x, ...)
}
\arguments{
\item{x}{snap_graph object}

\item{...}{not used}

\item{gpt_path}{character path to gpt executable}
}
\description{
print method for snap_graph object

run snap graph
}
