% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_snapshot_data.R
\name{save_csv}
\alias{save_csv}
\title{Save a data.frame to a CSV file}
\usage{
save_csv(x)
}
\arguments{
\item{x}{A data.frame to save}
}
\value{
\link{character} Path to the temporary CSV file
}
\description{
Save a data.frame to a CSV file
}
\keyword{internal}
