% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_snapshot_object.R
\name{save_deparse}
\alias{save_deparse}
\title{Save an R object to a text file using deparse}
\usage{
save_deparse(x)
}
\arguments{
\item{x}{An R object to save}
}
\value{
\link{character} Path to the temporary text file
}
\description{
Save an R object to a text file using deparse
}
\keyword{internal}
