% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_snapshot_object.R
\name{save_diffobj}
\alias{save_diffobj}
\title{Save an R object using diffobj-friendly format}
\usage{
save_diffobj(x)
}
\arguments{
\item{x}{An R object to save}
}
\value{
\link{character} Path to the temporary text file
}
\description{
Saves an R object in a text format that is optimized for visualization
with diffobj in testthat's snapshot_review(). This uses a combination
of str() output which provides a hierarchical view of the object structure.
}
\examples{
\dontrun{
obj <- list(a = 1:5, b = letters[1:3], c = list(x = 1, y = 2))
path <- save_diffobj(obj)
cat(readLines(path), sep = "\n")
}
}
\keyword{internal}
