% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-snap-op-methods.R
\name{save_graph}
\alias{save_graph}
\title{write snap graph to file}
\usage{
save_graph(x, ...)
}
\arguments{
\item{x}{snap_graph object}

\item{...}{not used}
}
\description{
write snap graph to file
}
