% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_snapshot_object.R
\name{save_json}
\alias{save_json}
\title{Save an R object to a JSON file}
\usage{
save_json(x)
}
\arguments{
\item{x}{An R object to save}
}
\value{
\link{character} Path to the temporary JSON file
}
\description{
Save an R object to a JSON file
}
\keyword{internal}
