% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_snapshot_object.R
\name{save_rds}
\alias{save_rds}
\title{Save an R object to an RDS file}
\usage{
save_rds(x)
}
\arguments{
\item{x}{An R object to save}
}
\value{
\link{character} Path to the temporary RDS file
}
\description{
Save an R object to an RDS file
}
\keyword{internal}
