% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNAP.R
\name{SNAP}
\alias{SNAP}
\title{SNAP Wrapper Function}
\usage{
SNAP(RAIN, GI, FIRE, LIGCELL, SAND, Gdays = 240)
}
\arguments{
\item{RAIN}{Mean annual precipitation (mm/yr)}

\item{GI}{Grazing intensity (proportion of aboveground biomass removed per year, 0 - 1).}

\item{FIRE}{Proportion of aboveground biomass removed by fire per year (0 - 1).}

\item{LIGCELL}{Lignin and cellulose fraction of aboveground plant biomass (0 - 1).}

\item{SAND}{Sand fraction in the top 40 cm of soil (0 - 1).}

\item{Gdays}{Number of days in growing season (days). Default is 240.}
}
\description{
A wrapper function around all SNAP model equations to calculate SOCeq based on below input parameters.
}
