% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap-graph.R
\name{snap_graph}
\alias{snap_graph}
\title{create a snap_graph object
A snap graph object is used to compose snap gpt graphs}
\usage{
snap_graph(...)
}
\arguments{
\item{...}{snap_operator objects}
}
\description{
create a snap_graph object
A snap graph object is used to compose snap gpt graphs
}
