% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap_operator.R
\name{snap_operator}
\alias{snap_operator}
\alias{snap_read_operator}
\title{An S7 class for containing snap operators}
\usage{
snap_operator(
  operator = class_missing,
  operator_id = class_missing,
  operator_sources = class_missing,
  created_with = class_missing,
  xml_graph = class_missing
)

snap_read_operator(
  operator = class_missing,
  operator_id = class_missing,
  created_with = class_missing,
  xml_graph = class_missing
)
}
\arguments{
\item{operator}{character name of the snap operator}

\item{operator_id}{character id of the snap operator}

\item{created_with}{function name of the function that created the object}

\item{xml_graph}{character xml graph of the snap operator}
}
\description{
An S7 class for containing snap operators

An S7 class for containing snap_read_operators
}
\keyword{internal}
