% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snap_operator_helper.R
\name{snap_operator_help}
\alias{snap_operator_help}
\alias{show_xml}
\title{function for creating a new snap operator helper object}
\usage{
snap_operator_help(
  operator,
  gpt_path = "/home/hugh/esa-snap/bin/gpt",
  check_operator = TRUE,
  node = TRUE
)

show_xml(x, ...)
}
\arguments{
\item{operator}{character name of the operator}

\item{gpt_path}{character path to the gpt executable}

\item{check_operator}{logical check if the operator is valid}

\item{node}{logical return the node or the full xml graph}

\item{x}{snap_operator_help object}
}
\value{
snap_operator_help S7 object with the following slots:
operator, description, parameters, and xml_graph
}
\description{
function for creating a new snap operator helper object

show xml graph method for snap operator helper object
}
