% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-builders.R
\name{write_op_test}
\alias{write_op_test}
\title{function to write tests for functions created in \code{build_xml_engine}}
\usage{
write_op_test(op_name, gpt_op, null_src, node)
}
\description{
function to write tests for functions created in \code{build_xml_engine}
}
\keyword{internal}
