% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_SOCeq.R
\name{calc_SOCeq}
\alias{calc_SOCeq}
\title{Equilibrium Soil Organic Carbon (SOCeq)}
\usage{
calc_SOCeq(PDSOCt, DDSOCt, SAND, RAIN)
}
\arguments{
\item{PDSOCt}{Output of calc_PDSOCt()}

\item{DDSOCt}{Output of calc_DDSOCt()}

\item{SAND}{Sand \% in top 30 cm soil}

\item{RAIN}{MAP for year t (mm/year)}
}
\description{
To calculate the change in SOC for year t (deltaSOCt), we need the combination of PDSOCt and DDSOCt, but also the maximum rate of microbial respiration for year t (MRESPt). A key input to MRESPt is WETDAYS, which is based on RAIN, the Sand \% and SOCt.
}
\details{
SOCt is unknown to us though, so it's not helpful to have to rely on it as an input... We're interested in the SOCt at equilibrium (SOCeq from Eq 12).

By setting deltaSOCt = 0, we can solve for the SOCt term in MRESPt to yield SOCeq.

Not generally called directly but incorporated in wrapper function.
}
\seealso{
\link{SNAP} which wraps this function.

\link{calc_PDSOCt}

\link{calc_DDSOCt}
}
