% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op_read.R
\name{op_read}
\alias{op_read}
\title{Read: snap operator function}
\usage{
op_read(
  operator_id,
  operator_sources,
  file = NULL,
  formatName = NULL,
  bandNames = NULL,
  maskNames = NULL,
  pixelRegion = NULL,
  geometryRegion = NULL,
  useAdvancedOptions = FALSE,
  copyMetadata = TRUE
)
}
\arguments{
\item{operator_id}{character operator id}

\item{operator_sources}{character vector of operator sources that match
upstream \code{operator_id} values}

\item{file}{The file from which the data product is read. This is a
mandatory parameter. Value must not be empty.}

\item{formatName}{An (optional) format name. Value must not be empty.}

\item{bandNames}{The list of source bands.}

\item{maskNames}{The list of source masks.}

\item{pixelRegion}{The subset region in pixel coordinates. Use the following
format: x>,y>,width>,zheight> If not given, the entire scene is used. The
'geoRegion' parameter has precedence over this parameter.}

\item{geometryRegion}{The subset region in geographical coordinates using
WKT-format, e.g. POLYGON(({lon1} {lat1}, {lon2} {lat2}, ..., {lon1} {lat1}))
(make sure to quote the option due to spaces in {geometry}). If not given, the
entire scene is used.}

\item{useAdvancedOptions}{Whether to use advanced options for reading of the
source product. Default value is 'false'.}

\item{copyMetadata}{Whether to copy the metadata of the source product.
Default value is 'true'.}
}
\value{
xml2 xml graph
}
\description{
Read: snap operator function
}
\details{
Descrscription from '\verb{gpt \{operator\} -h}':

"Reads a data product from a given file location."
}
