% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-snap-op-methods.R, R/snap-operator-help.R,
%   R/snap-operator.R
\name{show_xml}
\alias{show_xml}
\title{show xml graph method for snap_operator_helper object}
\usage{
show_xml(x, ...)
}
\arguments{
\item{x}{snap_read_operator object}

\item{...}{not used}
}
\description{
show xml graph method for snap_operator_helper object

show xml graph method for snap_operator_help object

show xml graph method for snap_operator_help object

show xml graph method for snap_read_operator object
}
