% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNAP.R
\name{SNAP}
\alias{SNAP}
\title{SNAP Model Implementation}
\usage{
SNAP(SAND, RAIN, LIGCELL, GI, Fire)
}
\arguments{
\item{SAND}{Sand \% in top 30 cm soil}

\item{RAIN}{MAP for year t (mm/year)}

\item{LIGCELL}{Lignin and cellulose content of livestock feed for year t (proportion)}

\item{GI}{Grazing intensity for year t (proportion)}

\item{Fire}{Average number of fires per year (#/year)}
}
\description{
The five input parameters to the SNAP model are used to estimate the equilibrium soil organic carbon stocks in g/m^2.
}
