% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapr-package.R
\docType{package}
\name{snapr-package}
\alias{snapr}
\alias{snapr-package}
\title{snapr: Convenient Snapshot Testing Functions for R Packages}
\description{
Provides convenient snapshot testing functions for R packages, including expect_snapshot_data() for data.frames and expect_snapshot_object() for any R object.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/d-morrison/snapr}
  \item \url{https://d-morrison.github.io/snapr/}
  \item Report bugs at \url{https://github.com/d-morrison/snapr/issues}
}

}
\author{
\strong{Maintainer}: Douglas Ezra Morrison \email{demorrison@ucdavis.edu}

}
