% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapr-package.R
\docType{package}
\name{snapr-package}
\alias{snapr}
\alias{snapr-package}
\title{snapr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/d-morrison/snapr}
  \item Report bugs at \url{https://github.com/d-morrison/snapr/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
