% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNAPr-package.R
\docType{package}
\name{SNAPr-package}
\alias{SNAPr}
\alias{SNAPr-package}
\title{SNAPr: A model for simulating the impacts of grazing on SOC stocks and stock changes}
\description{
R implementation of the SNAP model first published by Ritchie 2014 to simulate the impacts of grazing, climate, soil, and vegetation parameters on SOC stocks and stock changes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jpjazzcat/SNAPr}
}

}
\author{
\strong{Maintainer}: Jacob Penner \email{jacob.penner@tnc.org} (\href{https://orcid.org/0000-0001-9578-4458}{ORCID})

}
\keyword{internal}
